Fit a Mean Parametrized Conway-Maxwell-Poisson Generalized Linear Model

The function glm.cmp is used to fit a mean parametrized Conway-Maxwell-Poisson generalized linear model with a log-link by using Fisher Scoring iteration.

glm.cmp(formula, data, offset = NULL,
   lambdalb = 1e-10, lambdaub = 1299, maxlambdaiter = 1e3, tol = 1e-6)

Arguments

formula
an object of class 'formula': a symblic desciption of the model to be fitted.

data
an optional data frame containing the variables in the model

offset
this can be used to specify an *a priori* known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector of length equal to the number of cases.

lambdalb, lambdaub
numeric: the lower and upper end points for the interval to be searched for lambda(s).

maxlambdaiter
numeric: the maximum number of iterations allowed to solve for lambda(s).

tol
numeric: the convergence threshold. A lambda is said to satisfy the mean constraint if the absolute difference between the calculated mean and a fitted values is less than tol.

Value

A fitted model object of class cmp similar to one obtained from glm or glm.nb.

The function summary (i.e., summary.cmp) can be used to obtain and print a summary of the results.

The function plot (i.e., plot.cmp) can be used to produce a range of diagnostic plots.

The generic assessor functions coef (i.e., coef.cmp), logLik (i.e., logLik.cmp) fitted (i.e., fitted.cmp), nobs (i.e., nobs.cmp), AIC (i.e., AIC.cmp) and residuals (i.e., residuals.cmp) can be used to extract various useful features of the value returned by glm.cmp.

An object class 'glm.cmp' is a list containing at least the following components:

coefficients

a named vector of coefficients

se_beta

approximate standard errors (using observed rather than expected information) for coefficients

residuals

the response residuals (i.e., observed-fitted)

fitted.values

the fitted mean values

rank

the numeric rank of the fitted linear model

linear.predictors

the linear fit on log scale

df.residuals

the residuals degrees of freedom

df.null

the residual degrees of freedom for the null model

null.deviance

The deviance for the null model. The null model will include only the intercept.

y

the y vector used.

x

the model matrix

model

the model frame

call

the matched call

formula

the formula supplied

terms

the terms object used

data

the data argument

offset

the offset vector used

Details

Fit a mean parametrizied COM-Poisson regression using maximum likelihood estimation via an iterative Fisher Scoring algorithm.

The COM-Poisson regression model is

Y_i ~ CMP(mu_i, nu),

where

E(Y_i) = mu_i = exp(x_i^T beta),

and nu > 0 is the dispersion parameter.

The fitted COM-Poisson distribution is over- or under-dispersed if nu < 1 and nu > 1 respectively.

References

Fung, T., Alwan, A., Wishart, J. and Huang, A. (2018). The mpcmp package for Mean-parametrized Conway-Maxwell-Poisson Regression.

Huang, A. (2017). Mean-parametrized Conway–Maxwell–Poisson regression models for dispersed counts. Statistical Modelling 17, 359--380.

See also

Examples

### Huang (2017) Page 368--370: Overdispersed Attendance data

# NOT RUN {
data(attendance)
M.attendance <- glm.cmp(daysabs~ gender+math+prog, data=attendance)
M.attendance
summary(M.attendance)
# }

### Huang (2017) Page 371--372: Underdispersed Takeover Bids data
data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
    M.bids

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + whtknght + 
#>     bidprem + insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Linear Model Coefficients:
#> (Intercept)     leglrest      rearest      finrest     whtknght      bidprem  
#>     0.99000      0.26800     -0.17300      0.06770      0.48100     -0.68500  
#>    insthold         size       sizesq     regulatn  
#>    -0.36800      0.17900     -0.00758     -0.03760  
#> 
#> Dispersion (nu): 1.757228
#> Degrees of Freedom: 125 Total (i.e. Null);  116 Residual
#> Null Deviance: 182.8135 
#> Residual Deviance: 131.4979 
#> AIC: 382.1754 
#> 

summary(M.bids)

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + whtknght + 
#>     bidprem + insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -2.71706  -0.70717  -0.07761   0.36132   3.05692  
#> 
#> Linear Model Coefficients:
#>             Estimate Std.Err Z value Pr(>|z|)    
#> (Intercept)   0.9897  0.4349    2.28  0.02285 *  
#> leglrest      0.2679  0.1227    2.18  0.02904 *  
#> rearest      -0.1732  0.1546   -1.12  0.26269    
#> finrest       0.0677  0.1742    0.39  0.69735    
#> whtknght      0.4813  0.1316    3.66  0.00025 ***
#> bidprem      -0.6849  0.3073   -2.23  0.02582 *  
#> insthold     -0.3679  0.3464   -1.06  0.28816    
#> size          0.1793  0.0476    3.77  0.00016 ***
#> sizesq       -0.0076  0.0025   -3.06  0.00225 ** 
#> regulatn     -0.0376  0.1301   -0.29  0.77269    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for Mean-CMP(1.757) family taken to be 1)
#> 
#>     Null deviance: 182.81  on 125 degrees of freedom
#> Residual deviance: 131.50  on 116 degrees of freedom
#> 
#> AIC: 382.1754 
#>