Non-randomized Probability Integral Transform

Functions to produce the non-randomzied probability integral transform (PIT) to check the adequacy of the distributional assumption of the COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

compPredProb(object)

compPIT(object, bins = 10)

Arguments

object
an object class "cmp", obtained from a call to glm.cmp.

bins
numeric; the number of bins shown in the PIT histogram or the PIT Q-Q plot.

Value

compPredprob returns a list with values:

upper

the predictive cumulative probabilities used as the upper bound for computing the non-randomzied PIT.

lower

the predictive cumulative probabilities used as the upper bound for computing the non-randomzied PIT.

compPIT returns a list with values:
conditionalPIT

the conditional probability integral transoformation given the observed counts.

PIT

the probability integral transformation.

Details

These functions are used to obtain the predictive probabilities and the probability integral transform for a fitted COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

References

Czado, C., Gneiting, T. and Held, L. (2009). Predictive model assessment for count data. Biometrics, 65, 1254--1261.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Seires Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
compPredProb(M.bids)

#> $lower
#>   [1] 0.16804162 0.00000000 0.05645573 0.23361641 0.22833977 0.66774650
#>   [7] 0.42366439 0.28308191 0.40317554 0.00000000 0.30744284 0.38382762
#>  [13] 0.28987423 0.16593630 0.45581761 0.34934593 0.60505643 0.40186691
#>  [19] 0.52608392 0.33607608 0.96179197 0.14755181 0.31034141 0.61602794
#>  [25] 0.08385275 0.13249111 0.14748410 0.30628345 0.90584522 0.26854345
#>  [31] 0.50068856 0.10063028 0.90466406 0.12769417 0.08635869 0.99660674
#>  [37] 0.28781977 0.19728280 0.12351865 0.17338972 0.27812389 0.48197779
#>  [43] 0.04072820 0.34544260 0.36009989 0.13281163 0.04262089 0.04551656
#>  [49] 0.33230543 0.63652016 0.16053844 0.70272243 0.51802467 0.44656351
#>  [55] 0.80590275 0.10653169 0.29049007 0.42297257 0.45958549 0.63624132
#>  [61] 0.00000000 0.76811344 0.00000000 0.05191588 0.99769462 0.40870376
#>  [67] 0.40457961 0.17455051 0.47342018 0.36768253 0.05310539 0.82962591
#>  [73] 0.82053772 0.98156240 0.02209898 0.40755419 0.32680245 0.27681796
#>  [79] 0.74102980 0.41223260 0.26078309 0.04414228 0.10466902 0.25148394
#>  [85] 0.30211753 0.72180573 0.94236312 0.64448568 0.09266157 0.37397233
#>  [91] 0.14434797 0.09828880 0.19796173 0.07220317 0.00000000 0.07480719
#>  [97] 0.00000000 0.13380158 0.06083210 0.23698620 0.24972104 0.03776903
#> [103] 0.25034431 0.12363664 0.35217322 0.71517354 0.01987096 0.17387919
#> [109] 0.24944522 0.00000000 0.00000000 0.28722754 0.27698698 0.23962374
#> [115] 0.19240610 0.10082380 0.59363921 0.99477678 0.49914700 0.84905947
#> [121] 0.74941069 0.94165832 0.59914897 0.59264093 0.00000000 0.00208884
#> 
#> $upper
#>   [1] 0.45733775 0.20061437 0.31036988 0.67242011 0.66549226 0.88734903
#>   [7] 0.76201157 0.73097056 0.83645493 0.24314879 0.75609816 0.72670223
#>  [13] 0.62828688 0.57095782 0.78812893 0.79455242 0.85002711 0.74311496
#>  [19] 0.83858218 0.78297833 0.99569019 0.53769899 0.65176012 0.83242818
#>  [25] 0.39360274 0.50811393 0.53757095 0.75495145 0.97870348 0.71487722
#>  [31] 0.82133797 0.43697355 0.97832717 0.49819530 0.40039286 0.99924624
#>  [37] 0.73603239 0.62158131 0.48935117 0.58364435 0.72557884 0.80792881
#>  [43] 0.25260909 0.68926911 0.80355077 0.39440708 0.26007555 0.27119968
#>  [49] 0.77959107 0.86940418 0.56149292 0.93178891 0.75928266 0.86508174
#>  [55] 0.95203750 0.45115998 0.73884707 0.70968391 0.79105866 0.90180412
#>  [61] 0.29371624 0.93668652 0.17897724 0.29462101 0.99972715 0.74914971
#>  [67] 0.74552139 0.58558163 0.80158951 0.80969875 0.29881297 0.94212740
#>  [73] 0.95750862 0.99638793 0.16839163 0.74814200 0.77456641 0.72414227
#>  [79] 0.94653035 0.69971935 0.70592474 0.26596386 0.33769217 0.69484673
#>  [85] 0.75079172 0.91547966 0.99219416 0.90583764 0.41696912 0.71742581
#>  [91] 0.53159304 0.43120024 0.62260293 0.36038569 0.05811363 0.36806102
#>  [97] 0.20381658 0.51078014 0.32494520 0.67676969 0.69270209 0.16567664
#> [103] 0.69346198 0.48960385 0.69609181 0.93677794 0.15656459 0.58446248
#> [109] 0.69236524 0.32392253 0.02552198 0.62515925 0.72432859 0.68013434
#> [115] 0.61415555 0.43744693 0.87946664 0.99911056 0.82025620 0.95836415
#> [121] 0.94951787 0.98905904 0.84624088 0.87891242 0.33510166 0.02970313
#> 

compPIT(M.bids)

#> $ConditionalPIT
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.4984688 0.1714921 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1104694 0.9969375 0.5653260 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.4561360 1.0000000 0.9591599 0.1512831 0.1639250 0.0000000 0.0000000
#>  [5,] 0.8018026 1.0000000 1.0000000 0.3791755 0.3926782 0.0000000 0.0000000
#>  [6,] 1.0000000 1.0000000 1.0000000 0.6070678 0.6214313 0.0000000 0.2256133
#>  [7,] 1.0000000 1.0000000 1.0000000 0.8349601 0.8501844 0.0000000 0.5211677
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.1468722 0.8167221
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.6022403 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>             [,8]      [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.0000000 0.4112708 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.8225416 0.0000000 0.00000000 0.00000000 0.08410343
#>  [4,] 0.03777298 0.0000000 1.0000000 0.0000000 0.00000000 0.02992138 0.33100389
#>  [5,] 0.26104276 0.0000000 1.0000000 0.2062990 0.04716703 0.32541860 0.57790435
#>  [6,] 0.48431254 0.2234689 1.0000000 0.4291873 0.33881884 0.62091582 0.82480481
#>  [7,] 0.70758231 0.4542668 1.0000000 0.6520755 0.63047065 0.91641305 1.00000000
#>  [8,] 0.93085209 0.6850648 1.0000000 0.8749638 0.92212245 1.00000000 1.00000000
#>  [9,] 1.00000000 0.9158628 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20] [,21]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [4,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [5,] 0.0000000 0.1137766 0.0000000 0.0000000 0.0000000 0.1430378     0
#>  [6,] 0.1329548 0.3383914 0.0000000 0.2875711 0.0000000 0.3668004     0
#>  [7,] 0.4338774 0.5630063 0.0000000 0.5806131 0.2365328 0.5905630     0
#>  [8,] 0.7348001 0.7876212 0.3875712 0.8736551 0.5565345 0.8143255     0
#>  [9,] 1.0000000 1.0000000 0.7957833 1.0000000 0.8765363 1.0000000     0
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     0
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     1
#>           [,22]     [,23]     [,24]      [,25]     [,26]     [,27]     [,28]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.05212993 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1344318 0.0000000 0.0000000 0.37497095 0.1797252 0.1346262 0.0000000
#>  [4,] 0.3907453 0.0000000 0.0000000 0.69781196 0.4459497 0.3909793 0.0000000
#>  [5,] 0.6470589 0.2626060 0.0000000 1.00000000 0.7121742 0.6473325 0.2088773
#>  [6,] 0.9033724 0.5555015 0.0000000 1.00000000 0.9783987 0.9036857 0.4317592
#>  [7,] 1.0000000 0.8483969 0.0000000 1.00000000 1.0000000 1.0000000 0.6546412
#>  [8,] 1.0000000 1.0000000 0.3880405 1.00000000 1.0000000 1.0000000 0.8775231
#>  [9,] 1.0000000 1.0000000 0.8501472 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>       [,29]      [,30]     [,31]     [,32] [,33]     [,34]      [,35] [,36]
#>  [1,]     0 0.00000000 0.0000000 0.0000000     0 0.0000000 0.00000000     0
#>  [2,]     0 0.00000000 0.0000000 0.0000000     0 0.0000000 0.04343892     0
#>  [3,]     0 0.00000000 0.0000000 0.2954414     0 0.1951568 0.36187561     0
#>  [4,]     0 0.07047764 0.0000000 0.5927567     0 0.4650616 0.68031230     0
#>  [5,]     0 0.29452521 0.0000000 0.8900720     0 0.7349663 0.99874899     0
#>  [6,]     0 0.51857279 0.0000000 1.0000000     0 1.0000000 1.00000000     0
#>  [7,]     0 0.74262037 0.3097197 1.0000000     0 1.0000000 1.00000000     0
#>  [8,]     0 0.96666795 0.6215868 1.0000000     0 1.0000000 1.00000000     0
#>  [9,]     0 1.00000000 0.9334539 1.0000000     0 1.0000000 1.00000000     0
#> [10,]     0 1.00000000 1.0000000 1.0000000     0 1.0000000 1.00000000     0
#> [11,]     1 1.00000000 1.0000000 1.0000000     1 1.0000000 1.00000000     1
#>            [,37]       [,38]     [,39]      [,40]      [,41]      [,42]
#>  [1,] 0.00000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.006403978 0.2090611 0.06486285 0.00000000 0.00000000
#>  [4,] 0.02717511 0.242087109 0.4824102 0.30861390 0.04889008 0.00000000
#>  [5,] 0.25028351 0.477770241 0.7557594 0.55236495 0.27237626 0.00000000
#>  [6,] 0.47339192 0.713453373 1.0000000 0.79611601 0.49586245 0.05529117
#>  [7,] 0.69650032 0.949136504 1.0000000 1.00000000 0.71934863 0.36208573
#>  [8,] 0.91960872 1.000000000 1.0000000 1.00000000 0.94283482 0.66888028
#>  [9,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 0.97567484
#> [10,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,43]     [,44]      [,45]     [,46]    [,47]     [,48]     [,49]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.000000 0.0000000 0.0000000
#>  [2,] 0.2797411 0.0000000 0.00000000 0.0000000 0.263867 0.2414156 0.0000000
#>  [3,] 0.7517044 0.0000000 0.00000000 0.2568408 0.723733 0.6845148 0.0000000
#>  [4,] 1.0000000 0.0000000 0.00000000 0.6391104 1.000000 1.0000000 0.0000000
#>  [5,] 1.0000000 0.1586771 0.08997639 1.0000000 1.000000 1.0000000 0.1513453
#>  [6,] 1.0000000 0.4495215 0.31548051 1.0000000 1.000000 1.0000000 0.3749161
#>  [7,] 1.0000000 0.7403658 0.54098462 1.0000000 1.000000 1.0000000 0.5984868
#>  [8,] 1.0000000 1.0000000 0.76648874 1.0000000 1.000000 1.0000000 0.8220576
#>  [9,] 1.0000000 1.0000000 0.99199286 1.0000000 1.000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.00000000 1.0000000 1.000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.000000 1.0000000 1.0000000
#>           [,50]      [,51]     [,52]     [,53]     [,54]     [,55]     [,56]
#>  [1,] 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.09841905 0.0000000 0.0000000 0.0000000 0.0000000 0.2712149
#>  [4,] 0.0000000 0.34782392 0.0000000 0.0000000 0.0000000 0.0000000 0.5613826
#>  [5,] 0.0000000 0.59722879 0.0000000 0.0000000 0.0000000 0.0000000 0.8515503
#>  [6,] 0.0000000 0.84663366 0.0000000 0.0000000 0.1276802 0.0000000 1.0000000
#>  [7,] 0.0000000 1.00000000 0.0000000 0.3397829 0.3666184 0.0000000 1.0000000
#>  [8,] 0.2725813 1.00000000 0.0000000 0.7542769 0.6055566 0.0000000 1.0000000
#>  [9,] 0.7019796 1.00000000 0.4246696 1.0000000 0.8444949 0.0000000 1.0000000
#> [10,] 1.0000000 1.00000000 0.8612241 1.0000000 1.0000000 0.6439074 1.0000000
#> [11,] 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>            [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.3404647 0.0000000 0.5587302
#>  [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.6809293 0.0000000 1.0000000
#>  [4,] 0.02121062 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [5,] 0.24424718 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [6,] 0.46728373 0.2686585 0.1219239 0.0000000 1.0000000 0.0000000 1.0000000
#>  [7,] 0.69032028 0.6174413 0.4236075 0.0000000 1.0000000 0.0000000 1.0000000
#>  [8,] 0.91335683 0.9662242 0.7252910 0.2400889 1.0000000 0.0000000 1.0000000
#>  [9,] 1.00000000 1.0000000 1.0000000 0.6166476 1.0000000 0.1891557 1.0000000
#> [10,] 1.00000000 1.0000000 1.0000000 0.9932064 1.0000000 0.7823702 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,64] [,65]     [,66]     [,67]      [,68]      [,69]      [,70]
#>  [1,] 0.0000000     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.1981175     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.6101400     0 0.0000000 0.0000000 0.06191621 0.00000000 0.00000000
#>  [4,] 1.0000000     0 0.0000000 0.0000000 0.30520679 0.00000000 0.00000000
#>  [5,] 1.0000000     0 0.0000000 0.0000000 0.54849737 0.00000000 0.07311377
#>  [6,] 1.0000000     0 0.2681666 0.2798730 0.79178796 0.08099421 0.29934981
#>  [7,] 1.0000000     0 0.5618990 0.5731782 1.00000000 0.38571496 0.52558585
#>  [8,] 1.0000000     0 0.8556314 0.8664834 1.00000000 0.69043570 0.75182190
#>  [9,] 1.0000000     0 1.0000000 1.0000000 1.00000000 0.99515644 0.97805794
#> [10,] 1.0000000     0 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.0000000     1 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,71]     [,72]     [,73] [,74]     [,75]     [,76]     [,77]
#>  [1,] 0.0000000 0.0000000 0.0000000     0 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1908554 0.0000000 0.0000000     0 0.5325013 0.0000000 0.0000000
#>  [3,] 0.5978432 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [4,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [5,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.1634735
#>  [6,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.2714302 0.3868055
#>  [7,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.5650402 0.6101374
#>  [8,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.8586503 0.8334694
#>  [9,] 1.0000000 0.0000000 0.0000000     0 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 0.6255392 0.5801399     0 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000     1 1.0000000 1.0000000 1.0000000
#>            [,78]     [,79]     [,80]      [,81]     [,82]     [,83]     [,84]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.00000000 0.2518137 0.0000000 0.0000000
#>  [3,] 0.00000000 0.0000000 0.0000000 0.00000000 0.7026265 0.4091052 0.0000000
#>  [4,] 0.05182378 0.0000000 0.0000000 0.08809984 1.0000000 0.8382471 0.1094275
#>  [5,] 0.27537524 0.0000000 0.0000000 0.31274744 1.0000000 1.0000000 0.3349764
#>  [6,] 0.49892669 0.0000000 0.3052920 0.53739503 1.0000000 1.0000000 0.5605253
#>  [7,] 0.72247814 0.0000000 0.6531341 0.76204262 1.0000000 1.0000000 0.7860742
#>  [8,] 0.94602959 0.0000000 1.0000000 0.98669021 1.0000000 1.0000000 1.0000000
#>  [9,] 1.00000000 0.2869588 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.00000000 0.7735756 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>           [,85]     [,86] [,87]     [,88]      [,89]      [,90]     [,91]
#>  [1,] 0.0000000 0.0000000     0 0.0000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.0000000 0.0000000     0 0.0000000 0.02262799 0.00000000 0.0000000
#>  [3,] 0.0000000 0.0000000     0 0.0000000 0.33097728 0.00000000 0.1437127
#>  [4,] 0.0000000 0.0000000     0 0.0000000 0.63932657 0.00000000 0.4019471
#>  [5,] 0.2181593 0.0000000     0 0.0000000 0.94767585 0.07578224 0.6601815
#>  [6,] 0.4410382 0.0000000     0 0.0000000 1.00000000 0.36694248 0.9184159
#>  [7,] 0.6639171 0.0000000     0 0.0000000 1.00000000 0.65810273 1.0000000
#>  [8,] 0.8867960 0.0000000     0 0.2124121 1.00000000 0.94926297 1.0000000
#>  [9,] 1.0000000 0.4037419     0 0.5950379 1.00000000 1.00000000 1.0000000
#> [10,] 1.0000000 0.9200736     0 0.9776637 1.00000000 1.00000000 1.0000000
#> [11,] 1.0000000 1.0000000     1 1.0000000 1.00000000 1.00000000 1.0000000
#>             [,92]       [,93]      [,94] [,95]      [,96]     [,97]     [,98]
#>  [1,] 0.000000000 0.000000000 0.00000000     0 0.00000000 0.0000000 0.0000000
#>  [2,] 0.005140112 0.000000000 0.09645564     1 0.08590788 0.4906372 0.0000000
#>  [3,] 0.305520297 0.004799974 0.44345795     1 0.42690939 0.9812744 0.1756026
#>  [4,] 0.605900482 0.240292905 0.79046025     1 0.76791090 1.0000000 0.4408697
#>  [5,] 0.906280668 0.475785835 1.00000000     1 1.00000000 1.0000000 0.7061368
#>  [6,] 1.000000000 0.711278766 1.00000000     1 1.00000000 1.0000000 0.9714038
#>  [7,] 1.000000000 0.946771696 1.00000000     1 1.00000000 1.0000000 1.0000000
#>  [8,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#>  [9,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#> [10,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#> [11,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#>           [,99]    [,100]    [,101]    [,102]    [,103]    [,104]    [,105]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1482997 0.0000000 0.0000000 0.4865306 0.0000000 0.0000000 0.0000000
#>  [3,] 0.5269254 0.0000000 0.0000000 1.0000000 0.0000000 0.2086618 0.0000000
#>  [4,] 0.9055511 0.1432837 0.1135014 1.0000000 0.1120598 0.4819103 0.0000000
#>  [5,] 1.0000000 0.3706683 0.3392447 1.0000000 0.3377335 0.7551588 0.1390643
#>  [6,] 1.0000000 0.5980529 0.5649880 1.0000000 0.5634072 1.0000000 0.4298307
#>  [7,] 1.0000000 0.8254375 0.7907313 1.0000000 0.7890809 1.0000000 0.7205972
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>          [,106]    [,107]    [,108]    [,109]    [,110] [,111]     [,112]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000      0 0.00000000
#>  [2,] 0.0000000 0.5861944 0.0000000 0.0000000 0.3087158      1 0.00000000
#>  [3,] 0.0000000 1.0000000 0.0636188 0.0000000 0.6174316      1 0.00000000
#>  [4,] 0.0000000 1.0000000 0.3071747 0.1141397 0.9261474      1 0.03779597
#>  [5,] 0.0000000 1.0000000 0.5507307 0.3399141 1.0000000      1 0.33371375
#>  [6,] 0.0000000 1.0000000 0.7942866 0.5656885 1.0000000      1 0.62963153
#>  [7,] 0.0000000 1.0000000 1.0000000 0.7914629 1.0000000      1 0.92554931
#>  [8,] 0.0000000 1.0000000 1.0000000 1.0000000 1.0000000      1 1.00000000
#>  [9,] 0.3827833 1.0000000 1.0000000 1.0000000 1.0000000      1 1.00000000
#> [10,] 0.8340379 1.0000000 1.0000000 1.0000000 1.0000000      1 1.00000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000      1 1.00000000
#>           [,113]    [,114]    [,115]    [,116]     [,117] [,118]      [,119]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000      0 0.000000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000      0 0.000000000
#>  [3,] 0.00000000 0.0000000 0.0180057 0.2946209 0.00000000      0 0.000000000
#>  [4,] 0.05144394 0.1370597 0.2551133 0.5916890 0.00000000      0 0.000000000
#>  [5,] 0.27498675 0.3640690 0.4922209 0.8887571 0.00000000      0 0.000000000
#>  [6,] 0.49852956 0.5910783 0.7293285 1.0000000 0.00000000      0 0.002656407
#>  [7,] 0.72207237 0.8180876 0.9664361 1.0000000 0.02225397      0 0.314076947
#>  [8,] 0.94561518 1.0000000 1.0000000 1.0000000 0.37211542      0 0.625497488
#>  [9,] 1.00000000 1.0000000 1.0000000 1.0000000 0.72197687      0 0.936918029
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000      0 1.000000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000      1 1.000000000
#>          [,120]    [,121] [,122]      [,123]     [,124]    [,125] [,126]
#>  [1,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.0000000      0
#>  [2,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.2984169      1
#>  [3,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.5968338      1
#>  [4,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.8952507      1
#>  [5,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [6,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [7,] 0.0000000 0.0000000      0 0.003444173 0.02570661 1.0000000      1
#>  [8,] 0.0000000 0.0000000      0 0.408151885 0.37502537 1.0000000      1
#>  [9,] 0.0000000 0.2528111      0 0.812859597 0.72434412 1.0000000      1
#> [10,] 0.4660416 0.7525433      0 1.000000000 1.00000000 1.0000000      1
#> [11,] 1.0000000 1.0000000      1 1.000000000 1.00000000 1.0000000      1
#> 
#> $PIT
#>       [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]    0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,]    0 0.4984688 0.3349804 0.2233203 0.1674902 0.1339922 0.1116601
#>  [3,]    0 1.1074070 0.8363665 0.5575777 0.4181832 0.3345466 0.2787888
#>  [4,]    0 1.4561360 1.2076479 0.8555263 0.6826260 0.5461008 0.4550840
#>  [5,]    0 1.8018026 1.4009013 1.0603260 0.8934140 0.7147312 0.5956094
#>  [6,]    0 2.0000000 1.5000000 1.2023559 1.0571248 0.8456998 0.7423521
#>  [7,]    0 2.0000000 1.5000000 1.2783200 1.1712861 0.9370289 0.8677187
#>  [8,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.0293744 0.9939324
#>  [9,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.1204481 1.1003734
#> [10,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#> [11,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#>            [,8]       [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0957087 0.08374511 0.1201369 0.1081232 0.09829379 0.09010264 0.08317167
#>  [3,] 0.2389619 0.20909162 0.2772527 0.2495275 0.22684315 0.20793955 0.19841369
#>  [4,] 0.3954681 0.34603463 0.4186974 0.3768277 0.34257064 0.31651653 0.31763095
#>  [5,] 0.5478141 0.47933737 0.5371888 0.5040998 0.46256046 0.45113197 0.46088369
#>  [6,] 0.7054893 0.64523672 0.6846549 0.6591081 0.62999090 0.62923464 0.64427850
#>  [7,] 0.8448421 0.79602017 0.8186846 0.8020237 0.78642796 0.79726005 0.81285543
#>  [8,] 0.9849209 0.94743889 0.9532790 0.9454475 0.94332704 0.94804978 0.95204595
#>  [9,] 1.0860343 1.06476289 1.0575670 1.0518103 1.04710028 1.04317526 1.03985409
#> [10,] 1.1428571 1.12500000 1.1111111 1.1000000 1.09090909 1.08333333 1.07692308
#> [11,] 1.1428571 1.12500000 1.1111111 1.1000000 1.09090909 1.08333333 1.07692308
#>            [,15]      [,16]      [,17]      [,18]      [,19]      [,20]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07723083 0.07208211 0.06757698 0.06360186 0.06006843 0.05690693
#>  [3,] 0.18424129 0.17195854 0.16121113 0.15172812 0.14329878 0.13575674
#>  [4,] 0.29494302 0.27528015 0.25807514 0.24289425 0.22940013 0.21732644
#>  [5,] 0.42796343 0.40701764 0.38157903 0.35913321 0.33918136 0.32885802
#>  [6,] 0.60775538 0.58979778 0.55293542 0.53732576 0.50747433 0.50007044
#>  [7,] 0.78578557 0.77093362 0.72275027 0.71438926 0.68784168 0.68272175
#>  [8,] 0.93652839 0.92660125 0.89291187 0.89177911 0.87315441 0.87005816
#>  [9,] 1.03700737 1.03454021 1.01961790 1.01846391 1.01057904 1.01002225
#> [10,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#> [11,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#>            [,21]      [,22]      [,23]      [,24]     [,25]      [,26]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.05406158 0.05148722 0.04914689 0.04701007 0.0472234 0.04533446
#>  [3,] 0.12896890 0.12922904 0.12335499 0.11799173 0.1286992 0.13074024
#>  [4,] 0.20646011 0.21523560 0.20545216 0.19651946 0.2174066 0.22654837
#>  [5,] 0.31241512 0.32835053 0.32536214 0.31121596 0.3399153 0.35480566
#>  [6,] 0.47506692 0.49546241 0.49819146 0.47653096 0.4983422 0.51754444
#>  [7,] 0.64858566 0.66531968 0.67364137 0.64435262 0.6591713 0.67280441
#>  [8,] 0.82655525 0.83481452 0.84232296 0.82257155 0.8299644 0.83676582
#>  [9,] 0.95952114 0.96144870 0.96320103 0.95828565 0.9600237 0.96162280
#> [10,] 1.00000000 1.00000000 1.00000000 1.00000000 1.0000000 1.00000000
#> [11,] 1.05000000 1.04761905 1.04545455 1.04347826 1.0416667 1.04000000
#>            [,27]      [,28]     [,29]      [,30]      [,31]      [,32]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04359083 0.04197636 0.0404772 0.03908144 0.03777872 0.03656005
#>  [3,] 0.13088970 0.12604193 0.1215404 0.11734939 0.11343774 0.11930882
#>  [4,] 0.23287264 0.22424773 0.2162389 0.21121263 0.20417221 0.21670719
#>  [5,] 0.36605669 0.36023523 0.3473697 0.34554746 0.33402921 0.35196608
#>  [6,] 0.53239602 0.52866873 0.5097877 0.51009064 0.49308762 0.50943963
#>  [7,] 0.68538885 0.68425005 0.6598125 0.66266799 0.65090305 0.66216424
#>  [8,] 0.84304406 0.84432106 0.8141667 0.81942540 0.81283078 0.81886850
#>  [9,] 0.96309884 0.96446555 0.9300204 0.93243345 0.93246746 0.93464593
#> [10,] 1.00000000 1.00000000 0.9642857 0.96551724 0.96666667 0.96774194
#> [11,] 1.03846154 1.03703704 1.0357143 1.03448276 1.03333333 1.03225806
#>            [,33]      [,34]      [,35]      [,36]     [,37]      [,38]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.03541755 0.03434429 0.03461178 0.03362287 0.0326889 0.03180542
#>  [3,] 0.11558042 0.11799183 0.12516488 0.12158874 0.1182113 0.11518946
#>  [4,] 0.20993509 0.21766620 0.23127344 0.22466562 0.2191798 0.21979889
#>  [5,] 0.34096714 0.35290650 0.37190187 0.36127610 0.3581930 0.36142479
#>  [6,] 0.49351964 0.50886753 0.52331261 0.50836082 0.5073895 0.51295875
#>  [7,] 0.64147161 0.65233610 0.66256151 0.64363118 0.6450998 0.65331698
#>  [8,] 0.79327886 0.79954313 0.80543892 0.78242638 0.7862370 0.79201438
#>  [9,] 0.90543824 0.90830375 0.91100070 0.88497211 0.8881673 0.89118983
#> [10,] 0.93750000 0.93939394 0.94117647 0.91428571 0.9166667 0.91891892
#> [11,] 1.03125000 1.03030303 1.02941176 1.02857143 1.0277778 1.02702703
#>            [,39]      [,40]      [,41]      [,42]      [,43]      [,44]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03096844 0.03017437 0.02942001 0.02870245 0.03467956 0.03387306
#>  [3,] 0.11765976 0.11630600 0.11339835 0.11063253 0.12589615 0.12296833
#>  [4,] 0.22670972 0.22880982 0.22431183 0.21884081 0.23743984 0.23191798
#>  [5,] 0.37180202 0.37643184 0.37383045 0.36471263 0.37983852 0.37469523
#>  [6,] 0.52577563 0.53270743 0.53178631 0.52016448 0.53158913 0.52968058
#>  [7,] 0.66244022 0.67109560 0.67230192 0.66473567 0.67271816 0.67429136
#>  [8,] 0.79748769 0.80268031 0.80618417 0.80283530 0.80752970 0.81200575
#>  [9,] 0.89405326 0.89676984 0.89935060 0.90121216 0.90356425 0.90580695
#> [10,] 0.92105263 0.92307692 0.92500000 0.92682927 0.92857143 0.93023256
#> [11,] 1.02631579 1.02564103 1.02500000 1.02439024 1.02380952 1.02325581
#>            [,45]      [,46]      [,47]      [,48]      [,49]      [,50]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03310322 0.03236759 0.03740019 0.04174094 0.04087134 0.04003723
#>  [3,] 0.12017360 0.12321065 0.13626548 0.14793036 0.14484848 0.14189239
#>  [4,] 0.22664712 0.23581297 0.25242573 0.26833157 0.26274133 0.25737926
#>  [5,] 0.36822435 0.38226381 0.39569286 0.40855046 0.40319202 0.39496361
#>  [6,] 0.52481240 0.53537212 0.54547273 0.55514352 0.55138878 0.54013595
#>  [7,] 0.67126166 0.67856696 0.68555463 0.69224496 0.69029166 0.67620408
#>  [8,] 0.81097127 0.81517191 0.81918991 0.82303694 0.82301653 0.81178316
#>  [9,] 0.90776572 0.90981537 0.91177590 0.91365301 0.91545191 0.91109533
#> [10,] 0.93181818 0.93333333 0.93478261 0.93617021 0.93750000 0.93877551
#> [11,] 1.02272727 1.02222222 1.02173913 1.02127660 1.02083333 1.02040816
#>            [,51]      [,52]      [,53]      [,54]      [,55]      [,56]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03923649 0.03846714 0.03772739 0.03701555 0.03633008 0.03566953
#>  [3,] 0.14102292 0.13825776 0.13559896 0.13304049 0.13057678 0.13313383
#>  [4,] 0.25918815 0.25410603 0.24921938 0.24451712 0.23998903 0.24583255
#>  [5,] 0.39900891 0.39118521 0.38366241 0.37642350 0.36945269 0.37821811
#>  [6,] 0.54626590 0.53555481 0.52525568 0.51775425 0.50816621 0.51710864
#>  [7,] 0.68268000 0.66929412 0.66295736 0.65736606 0.64519261 0.65164366
#>  [8,] 0.81554750 0.79955637 0.79868561 0.79504167 0.78031868 0.78431288
#>  [9,] 0.91287342 0.90330080 0.90516040 0.90401577 0.88727473 0.88932428
#> [10,] 0.94000000 0.93845537 0.93963892 0.94077781 0.93528021 0.93645694
#> [11,] 1.02000000 1.01960784 1.01923077 1.01886792 1.01851852 1.01818182
#>            [,57]      [,58]      [,59]      [,60]      [,61]      [,62]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03503258 0.03441797 0.03382456 0.03325126 0.03837148 0.03774244
#>  [3,] 0.13075644 0.12846247 0.12624760 0.12410781 0.13338817 0.13120148
#>  [4,] 0.24182144 0.23757896 0.23348277 0.22952544 0.24236668 0.23839345
#>  [5,] 0.37582577 0.36923233 0.36286626 0.35671598 0.36743738 0.36141382
#>  [6,] 0.51621891 0.51187575 0.50515244 0.49659053 0.50498069 0.49670232
#>  [7,] 0.65233431 0.65172215 0.64778914 0.63680967 0.64286284 0.63232410
#>  [8,] 0.78661724 0.78976824 0.78865656 0.77935880 0.78303616 0.77019950
#>  [9,] 0.89130064 0.89320764 0.89504889 0.89033022 0.89215805 0.88063343
#> [10,] 0.93759163 0.93868652 0.93974365 0.94064980 0.94163897 0.93902800
#> [11,] 1.01785714 1.01754386 1.01724138 1.01694915 1.01666667 1.01639344
#>            [,63]      [,64]      [,65]      [,66]      [,67]      [,68]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04614547 0.04855773 0.04779901 0.04706364 0.04635056 0.04565876
#>  [3,] 0.14521436 0.15259413 0.15020985 0.14789893 0.14565803 0.14440815
#>  [4,] 0.25067743 0.26257144 0.25846876 0.25449232 0.25063637 0.25145086
#>  [5,] 0.37171360 0.38168640 0.37572255 0.36994220 0.36433701 0.36708568
#>  [6,] 0.50482002 0.51268002 0.50466940 0.50103089 0.49768002 0.50206969
#>  [7,] 0.63825436 0.64399635 0.63393391 0.63282568 0.63192193 0.63741563
#>  [8,] 0.77390596 0.77749475 0.76534640 0.76673540 0.76824673 0.77170573
#>  [9,] 0.88255869 0.88442284 0.87060373 0.87259445 0.87452483 0.87639760
#> [10,] 0.94001142 0.94096362 0.92626107 0.92739551 0.92849558 0.92956281
#> [11,] 1.01612903 1.01587302 1.01562500 1.01538462 1.01515152 1.01492537
#>           [,69]      [,70]      [,71]      [,72]      [,73]      [,74]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0449873 0.04433531 0.04642846 0.04577454 0.04513878 0.04452044
#>  [3,] 0.1422845 0.14022241 0.14675985 0.14469281 0.14268319 0.14072862
#>  [4,] 0.2477531 0.24416243 0.25496011 0.25136912 0.24787788 0.24448229
#>  [5,] 0.3616874 0.35750513 0.36668363 0.36151907 0.35649797 0.35161444
#>  [6,] 0.4958774 0.49302918 0.50027162 0.49322554 0.48637518 0.47971251
#>  [7,] 0.6337142 0.63214708 0.63740212 0.62842462 0.61969650 0.61120751
#>  [8,] 0.7705106 0.77023974 0.77352203 0.76262735 0.75203530 0.74173345
#>  [9,] 0.8781441 0.87959208 0.88131219 0.86889934 0.85683130 0.84509388
#> [10,] 0.9305986 0.93160447 0.93258155 0.92825700 0.92342204 0.91077243
#> [11,] 1.0147059 1.01449275 1.01428571 1.01408451 1.01388889 1.01369863
#>            [,75]      [,76]      [,77]      [,78]     [,79]      [,80]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.05111477 0.05043324 0.04976965 0.04912329 0.0484935 0.04787966
#>  [3,] 0.15234040 0.15030919 0.14833144 0.14640506 0.1445281 0.14269860
#>  [4,] 0.25469199 0.25129610 0.24798957 0.24544196 0.2422953 0.23922824
#>  [5,] 0.36037641 0.35557139 0.35304378 0.35203510 0.3475218 0.34312282
#>  [6,] 0.48674342 0.48387258 0.48259538 0.48280747 0.4766176 0.47444895
#>  [7,] 0.61646146 0.61577585 0.61570166 0.61708836 0.6091770 0.60973339
#>  [8,] 0.74522354 0.74673589 0.74787712 0.75045053 0.7408294 0.74411001
#>  [9,] 0.84718721 0.84922471 0.85120860 0.85314095 0.8458822 0.84783307
#> [10,] 0.91197821 0.91315183 0.91429457 0.91540763 0.9135893 0.91468307
#> [11,] 1.01351351 1.01333333 1.01315789 1.01298701 1.0128205 1.01265823
#>            [,81]      [,82]      [,83]      [,84]      [,85]      [,86]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04728117 0.04980626 0.04919887 0.04860611 0.04802747 0.04746244
#>  [3,] 0.14091487 0.14784958 0.15103563 0.14921592 0.14743954 0.14570496
#>  [4,] 0.23733914 0.24675471 0.25396803 0.25222657 0.24922388 0.24629183
#>  [5,] 0.34274313 0.35085741 0.35877378 0.35848707 0.35681650 0.35261866
#>  [6,] 0.47523578 0.48171435 0.48803491 0.48890829 0.48833840 0.48259325
#>  [7,] 0.61163726 0.61643186 0.62110952 0.62309705 0.62358300 0.61624673
#>  [8,] 0.74714227 0.75026397 0.75330953 0.75628170 0.75783544 0.74891973
#>  [9,] 0.84973515 0.85159027 0.85340015 0.85516641 0.85689062 0.85155946
#> [10,] 0.91574953 0.91678966 0.91780442 0.91879473 0.91976146 0.91976513
#> [11,] 1.01250000 1.01234568 1.01219512 1.01204819 1.01190476 1.01176471
#>            [,87]      [,88]      [,89]      [,90]      [,91]      [,92]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04691055 0.04637135 0.04610153 0.04558354 0.04507706 0.04463819
#>  [3,] 0.14401071 0.14235542 0.14449885 0.14287526 0.14288457 0.14467178
#>  [4,] 0.24342797 0.24062995 0.24516059 0.24240598 0.24417866 0.24815362
#>  [5,] 0.34851844 0.34451248 0.35136661 0.34827016 0.35173584 0.35782974
#>  [6,] 0.47698170 0.47149915 0.47750484 0.47626257 0.48117538 0.48687675
#>  [7,] 0.60908107 0.60208014 0.60660196 0.60718062 0.61154528 0.61581401
#>  [8,] 0.74021136 0.73414471 0.73716579 0.73954890 0.74244280 0.74527310
#>  [9,] 0.84165761 0.83882290 0.84065445 0.84244485 0.84419547 0.84590760
#> [10,] 0.90907019 0.90985862 0.91088296 0.91188427 0.91286333 0.91382088
#> [11,] 1.01162791 1.01149425 1.01136364 1.01123596 1.01111111 1.01098901
#>            [,93]      [,94]      [,95]      [,96]      [,97]      [,98]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04415299 0.04471539 0.05487799 0.05520462 0.05974037 0.05912449
#>  [3,] 0.14315143 0.14638053 0.15546159 0.15831894 0.16689139 0.16698119
#>  [4,] 0.24806818 0.25390036 0.26183759 0.26716467 0.27479837 0.27651045
#>  [5,] 0.35911187 0.36600314 0.37274779 0.37935044 0.38581554 0.38911782
#>  [6,] 0.48931590 0.49480713 0.50018152 0.50544277 0.51059441 0.51534502
#>  [7,] 0.61941138 0.62350373 0.62750901 0.63142996 0.63526923 0.63902935
#>  [8,] 0.74804187 0.75075110 0.75340268 0.75599845 0.75854013 0.76102941
#>  [9,] 0.84758252 0.84922142 0.85082545 0.85239570 0.85393325 0.85543909
#> [10,] 0.91475761 0.91567419 0.91657128 0.91744947 0.91830938 0.91915155
#> [11,] 1.01086957 1.01075269 1.01063830 1.01052632 1.01041667 1.01030928
#>            [,99]     [,100]     [,101]     [,102]     [,103]     [,104]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06003445 0.05942804 0.05883376 0.06306838 0.06245006 0.06184375
#>  [3,] 0.17065409 0.16893032 0.16724101 0.17548615 0.17376570 0.17410450
#>  [4,] 0.28292923 0.28151867 0.27983850 0.28696881 0.28525402 0.28716330
#>  [5,] 0.39535132 0.39510199 0.39454342 0.40053804 0.39992231 0.40337121
#>  [6,] 0.52029048 0.52107596 0.52151508 0.52625255 0.52661681 0.53121277
#>  [7,] 0.64271272 0.64455842 0.64602015 0.64952490 0.65089310 0.65428249
#>  [8,] 0.76346788 0.76585709 0.76819852 0.77049359 0.77274365 0.77495002
#>  [9,] 0.85691420 0.85835951 0.85977592 0.86116428 0.86252541 0.86386012
#> [10,] 0.91997653 0.92078485 0.92157700 0.92235347 0.92311471 0.92386117
#> [11,] 1.01020408 1.01010101 1.01000000 1.00990099 1.00980392 1.00970874
#>          [,105]     [,106]     [,107]     [,108]     [,109]     [,110]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0612491 0.06066577 0.06562359 0.06501029 0.06440834 0.06664969
#>  [3,] 0.1724304 0.17078822 0.17861097 0.17753628 0.17589242 0.17994324
#>  [4,] 0.2844021 0.28169352 0.28847000 0.28864481 0.28702902 0.29289249
#>  [5,] 0.4008298 0.39701237 0.40270093 0.40408438 0.40349022 0.40896278
#>  [6,] 0.5302379 0.52518806 0.52966741 0.53214049 0.53245112 0.53674056
#>  [7,] 0.6549201 0.64868279 0.65199711 0.65524947 0.65651071 0.65966199
#>  [8,] 0.7771140 0.76971288 0.77188540 0.77401731 0.77610974 0.77816378
#>  [9,] 0.8651692 0.86057500 0.86189033 0.86318108 0.86444792 0.86569152
#> [10,] 0.9245933 0.92373084 0.92445036 0.92515643 0.92584943 0.92652971
#> [11,] 1.0096154 1.00952381 1.00943396 1.00934579 1.00925926 1.00917431
#>           [,111]     [,112]    [,113]     [,114]     [,115]     [,116]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07513469 0.07445781 0.0737930 0.07313997 0.07249839 0.07186797
#>  [3,] 0.18739830 0.18571003 0.1840519 0.18242313 0.18098087 0.18196904
#>  [4,] 0.29932074 0.29696466 0.2947725 0.29337683 0.29304118 0.29563812
#>  [5,] 0.41433585 0.41360952 0.4123718 0.41194436 0.41264854 0.41678862
#>  [6,] 0.54095201 0.54175092 0.5413650 0.54180496 0.54344990 0.54741990
#>  [7,] 0.66275597 0.66512347 0.6656319 0.66698111 0.66960791 0.67248088
#>  [8,] 0.78018048 0.78216083 0.7836202 0.78553511 0.78741638 0.78926493
#>  [9,] 0.86691250 0.86811149 0.8692891 0.87044580 0.87158224 0.87269892
#> [10,] 0.92719762 0.92785349 0.9284977 0.92913042 0.92975209 0.93036294
#> [11,] 1.00909091 1.00900901 1.0089286 1.00884956 1.00877193 1.00869565
#>           [,117]     [,118]     [,119]     [,120]     [,121]     [,122]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07124842 0.07063946 0.07004082 0.06945224 0.06887347 0.06830427
#>  [3,] 0.18040034 0.17885846 0.17734271 0.17585244 0.17438700 0.17294579
#>  [4,] 0.29308951 0.29058448 0.28812190 0.28570070 0.28331986 0.28097838
#>  [5,] 0.41319561 0.40966402 0.40619229 0.40277891 0.39942242 0.39612141
#>  [6,] 0.54270077 0.53806230 0.53352496 0.52904156 0.52463288 0.52029707
#>  [7,] 0.66687548 0.66117569 0.65823418 0.65270280 0.64726361 0.64191432
#>  [8,] 0.78566882 0.77895370 0.77765322 0.77111832 0.76469234 0.75837257
#>  [9,] 0.87139959 0.86395173 0.86457009 0.85730479 0.85226734 0.84522381
#> [10,] 0.93096326 0.92300631 0.92365880 0.91981327 0.91841936 0.91082911
#> [11,] 1.00862069 1.00854701 1.00847458 1.00840336 1.00833333 1.00826446
#>          [,123]     [,124]     [,125]     [,126]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0677444 0.06719363 0.06905833 0.07650587
#>  [3,] 0.1715282 0.17013366 0.17357479 0.18018619
#>  [4,] 0.2786753 0.27640962 0.28140028 0.28714907
#>  [5,] 0.3928745 0.38968041 0.39460235 0.39944553
#>  [6,] 0.5160323 0.51183695 0.51577375 0.51964756
#>  [7,] 0.6366810 0.63171369 0.63468374 0.63760627
#>  [8,] 0.7555019 0.75240860 0.75440530 0.75637006
#>  [9,] 0.8449585 0.84397793 0.84523617 0.84647428
#> [10,] 0.9115600 0.91227905 0.91298647 0.91368258
#> [11,] 1.0081967 1.00813008 1.00806452 1.00800000
#>