Cotton Bolls data set

This data set gives the observed number of bolls produced by the cotton plants at five growth stages: vegetative, flower-bud, blossom, fig and cotton boll; to examine the effect of five defoliation levels (0

data(cottonbolls)

Format

A data frame with 125 observations on 4 variables.

nc

number of bolls produced by two cotton plants at harvest

stages

growth stage

def

artificial defoliation level

def2

square of def

Source

Supplementary Content of Zevini et al. (2014): http://www.leg.ufpr.br/doku.php/publications:papercompanions:zeviani-jas2014

References

Zeviani, W.M., Riberio P.J. Jr., Bonat, W.H., Shimakura S.E. and Muniz J.A. (2014). The Gamma-count distribution in the analysis of experimental underdispersed data. Journal of Applied Statistics 41, 2616--26.

Examples

## For examples see example(glm.cmp)