Non-randomized Probability Integral Transform

Functions to produce the non-randomized probability integral transform (PIT) to check the adequacy of the distributional assumption of the COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

compPredProb(object)

compPIT(object, bins = 10)

Arguments

object
an object class "cmp", obtained from a call to glm.cmp.

bins
numeric; the number of bins shown in the PIT histogram or the PIT Q-Q plot.

Value

compPredprob returns a list with values:

upper

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

lower

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

compPIT returns a list with values:
conditionalPIT

the conditional probability integral transformation given the observed counts.

PIT

the probability integral transformation.

Details

These functions are used to obtain the predictive probabilities and the probability integral transform for a fitted COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

References

Czado, C., Gneiting, T. and Held, L. (2009). Predictive model assessment for count data. Biometrics, 65, 1254--1261.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Series Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
compPredProb(M.bids)

#> $lower
#>   [1] 0.167217742 0.000000000 0.055886848 0.232782730 0.227540790 0.668012399
#>   [7] 0.423343339 0.282350973 0.402598024 0.000000000 0.306708955 0.383375434
#>  [13] 0.289153465 0.165112896 0.455580042 0.348672724 0.605150278 0.401426879
#>  [19] 0.525973088 0.335383175 0.962118768 0.146746963 0.309505243 0.616238000
#>  [25] 0.083175463 0.131699468 0.146680518 0.305543497 0.906414302 0.267794738
#>  [31] 0.500332030 0.099888629 0.905221574 0.126905588 0.085614483 0.996675684
#>  [37] 0.287052652 0.196460405 0.122739186 0.172480563 0.277336647 0.481782247
#>  [43] 0.040235703 0.344917130 0.359475573 0.131973528 0.042114516 0.045006660
#>  [49] 0.331500648 0.636785173 0.159694829 0.703058183 0.517772165 0.446064407
#>  [55] 0.806440880 0.105711905 0.289729236 0.422599263 0.459218716 0.636342390
#>  [61] 0.000000000 0.768603532 0.000000000 0.051369299 0.997747904 0.408313552
#>  [67] 0.404141820 0.173763143 0.473161918 0.367071413 0.052537150 0.830196946
#>  [73] 0.820968711 0.981802405 0.021775417 0.407122875 0.326094209 0.276044214
#>  [79] 0.741350207 0.411839074 0.260023490 0.043643059 0.103927339 0.250677833
#>  [85] 0.301376082 0.722212432 0.942782112 0.644623695 0.091967086 0.373460419
#>  [91] 0.143547518 0.097551370 0.197028656 0.071551971 0.000000000 0.074104456
#>  [97] 0.000000000 0.133009304 0.060182287 0.236165505 0.248907731 0.037321681
#> [103] 0.249428330 0.122860488 0.351603287 0.715466593 0.019538142 0.173066265
#> [109] 0.248654406 0.000000000 0.000000000 0.286365986 0.276212810 0.238818808
#> [115] 0.191567659 0.100111036 0.593639620 0.994871230 0.498971694 0.849511233
#> [121] 0.749783423 0.942134843 0.599288765 0.592670872 0.000000000 0.002032666
#> 
#> $upper
#>   [1] 0.45701443 0.19977533 0.30974965 0.67260214 0.66571076 0.88790645
#>   [7] 0.76255215 0.73131703 0.83685029 0.24226483 0.75644070 0.72715773
#>  [13] 0.62844174 0.57094305 0.78870014 0.79492902 0.85058972 0.74357074
#>  [19] 0.83914301 0.78334695 0.99576739 0.53761376 0.65180572 0.83304453
#>  [25] 0.39317147 0.50794775 0.53748768 0.75528819 0.97897709 0.71519969
#>  [31] 0.82176715 0.43660505 0.97859954 0.49799636 0.39981827 0.99926672
#>  [37] 0.73634100 0.62169286 0.48913541 0.58352098 0.72586263 0.80849693
#>  [43] 0.25183556 0.68966518 0.80395885 0.39384034 0.25930742 0.27050538
#>  [49] 0.77986080 0.87001778 0.56141405 0.93225557 0.75966778 0.86546956
#>  [55] 0.95245318 0.45067498 0.73916277 0.71008818 0.79152523 0.90229419
#>  [61] 0.29299557 0.93716011 0.17812350 0.29396643 0.99973622 0.74964444
#>  [67] 0.74597734 0.58566744 0.80212264 0.81010985 0.29811212 0.94259096
#>  [73] 0.95785572 0.99646262 0.16764738 0.74860158 0.77492669 0.72444035
#>  [79] 0.94691588 0.70010940 0.70623024 0.26527015 0.33708433 0.69508858
#>  [85] 0.75112750 0.91600776 0.99231964 0.90633040 0.41661771 0.71782937
#>  [91] 0.53149530 0.43081291 0.62254990 0.35983266 0.05753309 0.36740386
#>  [97] 0.20301328 0.51062288 0.32415740 0.67697407 0.69293316 0.16490799
#> [103] 0.69356860 0.48939626 0.69644250 0.93720292 0.15565218 0.58450273
#> [109] 0.69262349 0.32310618 0.02512772 0.62514391 0.72462627 0.68036278
#> [115] 0.61422660 0.43715178 0.87998124 0.99913409 0.82081432 0.95870780
#> [121] 0.94990580 0.98923340 0.84684242 0.87944513 0.33430601 0.02932721
#> 

compPIT(M.bids)

#> $ConditionalPIT
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.5005623 0.1737677 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1131216 1.0000000 0.5676812 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.4581911 1.0000000 0.9615948 0.1528292 0.1653678 0.0000000 0.0000000
#>  [5,] 0.8032606 1.0000000 1.0000000 0.3801953 0.3935897 0.0000000 0.0000000
#>  [6,] 1.0000000 1.0000000 1.0000000 0.6075613 0.6218117 0.0000000 0.2259866
#>  [7,] 1.0000000 1.0000000 1.0000000 0.8349274 0.8500336 0.0000000 0.5207903
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.1454682 0.8155940
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.6002327 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>             [,8]      [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.0000000 0.4127714 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.8255429 0.0000000 0.00000000 0.00000000 0.08596479
#>  [4,] 0.03931038 0.0000000 1.0000000 0.0000000 0.00000000 0.03196849 0.33237329
#>  [5,] 0.26204437 0.0000000 1.0000000 0.2074371 0.04835783 0.32670311 0.57878179
#>  [6,] 0.48477835 0.2242981 1.0000000 0.4297919 0.33923959 0.62143773 0.82519030
#>  [7,] 0.70751234 0.4545790 1.0000000 0.6521466 0.63012136 0.91617235 1.00000000
#>  [8,] 0.93024632 0.6848599 1.0000000 0.8745014 0.92100312 1.00000000 1.00000000
#>  [9,] 1.00000000 0.9151408 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20] [,21]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [4,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [5,] 0.0000000 0.1150175 0.0000000 0.0000000 0.0000000 0.1442456     0
#>  [6,] 0.1333452 0.3391040 0.0000000 0.2881043 0.0000000 0.3674780     0
#>  [7,] 0.4335372 0.5631904 0.0000000 0.5803790 0.2363794 0.5907103     0
#>  [8,] 0.7337292 0.7872769 0.3864486 0.8726538 0.5556948 0.8139426     0
#>  [9,] 1.0000000 1.0000000 0.7938810 1.0000000 0.8750103 1.0000000     0
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     0
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     1
#>           [,22]     [,23]     [,24]     [,25]     [,26]     [,27]     [,28]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0542734 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1362434 0.0000000 0.0000000 0.3768582 0.1815305 0.1364342 0.0000000
#>  [4,] 0.3920851 0.0000000 0.0000000 0.6994430 0.4473124 0.3923149 0.0000000
#>  [5,] 0.6479267 0.2643723 0.0000000 1.0000000 0.7130944 0.6481956 0.2100225
#>  [6,] 0.9037683 0.5565133 0.0000000 1.0000000 0.9788763 0.9040763 0.4323709
#>  [7,] 1.0000000 0.8486543 0.0000000 1.0000000 1.0000000 1.0000000 0.6547192
#>  [8,] 1.0000000 1.0000000 0.3863445 1.0000000 1.0000000 1.0000000 0.8770676
#>  [9,] 1.0000000 1.0000000 0.8475852 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>       [,29]      [,30]     [,31]        [,32] [,33]     [,34]      [,35] [,36]
#>  [1,]     0 0.00000000 0.0000000 0.0000000000     0 0.0000000 0.00000000     0
#>  [2,]     0 0.00000000 0.0000000 0.0003307572     0 0.0000000 0.04578403     0
#>  [3,]     0 0.00000000 0.0000000 0.2973165719     0 0.1969718 0.36404882     0
#>  [4,]     0 0.07198235 0.0000000 0.5943023867     0 0.4664476 0.68231360     0
#>  [5,]     0 0.29549351 0.0000000 0.8912882014     0 0.7359235 1.00000000     0
#>  [6,]     0 0.51900467 0.0000000 1.0000000000     0 1.0000000 1.00000000     0
#>  [7,]     0 0.74251583 0.3100718 1.0000000000     0 1.0000000 1.00000000     0
#>  [8,]     0 0.96602699 0.6211766 1.0000000000     0 1.0000000 1.00000000     0
#>  [9,]     0 1.00000000 0.9322814 1.0000000000     0 1.0000000 1.00000000     0
#> [10,]     0 1.00000000 1.0000000 1.0000000000     0 1.0000000 1.00000000     0
#> [11,]     1 1.00000000 1.0000000 1.0000000000     1 1.0000000 1.00000000     1
#>            [,37]       [,38]     [,39]      [,40]      [,41]      [,42]
#>  [1,] 0.00000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.008323906 0.2108668 0.06695068 0.00000000 0.00000000
#>  [4,] 0.02881746 0.243489396 0.4837954 0.31023576 0.05052852 0.00000000
#>  [5,] 0.25139167 0.478654887 0.7567240 0.55352084 0.27348104 0.00000000
#>  [6,] 0.47396588 0.713820377 1.0000000 0.79680592 0.49643357 0.05576044
#>  [7,] 0.69654010 0.948985868 1.0000000 1.00000000 0.71938610 0.36183789
#>  [8,] 0.91911431 1.000000000 1.0000000 1.00000000 0.94233862 0.66791535
#>  [9,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 0.97399281
#> [10,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,43]     [,44]      [,45]     [,46]     [,47]     [,48]     [,49]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.2824402 0.0000000 0.00000000 0.0000000 0.2665165 0.2438743 0.0000000
#>  [3,] 0.7550303 0.0000000 0.00000000 0.2597751 0.7269367 0.6873358 0.0000000
#>  [4,] 1.0000000 0.0000000 0.00000000 0.6416486 1.0000000 1.0000000 0.0000000
#>  [5,] 1.0000000 0.1597772 0.09117199 1.0000000 1.0000000 1.0000000 0.1527775
#>  [6,] 1.0000000 0.4498441 0.31615233 1.0000000 1.0000000 1.0000000 0.3758125
#>  [7,] 1.0000000 0.7399110 0.54113267 1.0000000 1.0000000 1.0000000 0.5988475
#>  [8,] 1.0000000 1.0000000 0.76611302 1.0000000 1.0000000 1.0000000 0.8218825
#>  [9,] 1.0000000 1.0000000 0.99109336 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,50]     [,51]     [,52]     [,53]     [,54]     [,55]     [,56]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.1003317 0.0000000 0.0000000 0.0000000 0.0000000 0.2733281
#>  [4,] 0.0000000 0.3492618 0.0000000 0.0000000 0.0000000 0.0000000 0.5632142
#>  [5,] 0.0000000 0.5981919 0.0000000 0.0000000 0.0000000 0.0000000 0.8531003
#>  [6,] 0.0000000 0.8471219 0.0000000 0.0000000 0.1286002 0.0000000 1.0000000
#>  [7,] 0.0000000 1.0000000 0.0000000 0.3399311 0.3670331 0.0000000 1.0000000
#>  [8,] 0.2710377 1.0000000 0.0000000 0.7533325 0.6054661 0.0000000 1.0000000
#>  [9,] 0.6997942 1.0000000 0.4229621 1.0000000 0.8438990 0.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 0.8592673 1.0000000 1.0000000 0.6407619 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>            [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.3413021 0.0000000 0.5614082
#>  [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.6826042 0.0000000 1.0000000
#>  [4,] 0.02285269 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [5,] 0.24535500 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [6,] 0.46785731 0.2692303 0.1227219 0.0000000 1.0000000 0.0000000 1.0000000
#>  [7,] 0.69035963 0.6170698 0.4236489 0.0000000 1.0000000 0.0000000 1.0000000
#>  [8,] 0.91286194 0.9649093 0.7245759 0.2393577 1.0000000 0.0000000 1.0000000
#>  [9,] 1.00000000 1.0000000 1.0000000 0.6153657 1.0000000 0.1862666 1.0000000
#> [10,] 1.00000000 1.0000000 1.0000000 0.9913737 1.0000000 0.7795393 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,64] [,65]     [,66]     [,67]      [,68]      [,69]      [,70]
#>  [1,] 0.0000000     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.2004587     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.6126647     0 0.0000000 0.0000000 0.06369649 0.00000000 0.00000000
#>  [4,] 1.0000000     0 0.0000000 0.0000000 0.30647133 0.00000000 0.00000000
#>  [5,] 1.0000000     0 0.0000000 0.0000000 0.54924617 0.00000000 0.07432445
#>  [6,] 1.0000000     0 0.2686146 0.2804219 0.79202101 0.08158446 0.30003850
#>  [7,] 1.0000000     0 0.5615854 0.5729603 1.00000000 0.38557212 0.52575255
#>  [8,] 1.0000000     0 0.8545563 0.8654987 1.00000000 0.68955978 0.75146660
#>  [9,] 1.0000000     0 1.0000000 1.0000000 1.00000000 0.99354744 0.97718064
#> [10,] 1.0000000     0 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.0000000     1 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,71]     [,72]     [,73] [,74]     [,75]     [,76]     [,77]
#>  [1,] 0.0000000 0.0000000 0.0000000     0 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1932723 0.0000000 0.0000000     0 0.5362551 0.0000000 0.0000000
#>  [3,] 0.6004800 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [4,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [5,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.1646623
#>  [6,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.2719851 0.3874626
#>  [7,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.5648291 0.6102629
#>  [8,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.8576732 0.8330631
#>  [9,] 1.0000000 0.0000000 0.0000000     0 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 0.6210567 0.5773469     0 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000     1 1.0000000 1.0000000 1.0000000
#>           [,78]     [,79]     [,80]     [,81]     [,82]     [,83]     [,84]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.2542872 0.0000000 0.0000000
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.7054956 0.4120514 0.0000000
#>  [4,] 0.0534255 0.0000000 0.0000000 0.0895919 1.0000000 0.8409469 0.1109833
#>  [5,] 0.2764426 0.0000000 0.0000000 0.3137033 1.0000000 1.0000000 0.3360004
#>  [6,] 0.4994597 0.0000000 0.3058273 0.5378146 1.0000000 1.0000000 0.5610174
#>  [7,] 0.7224768 0.0000000 0.6527239 0.7619260 1.0000000 1.0000000 0.7860345
#>  [8,] 0.9454938 0.0000000 0.9996205 0.9860373 1.0000000 1.0000000 1.0000000
#>  [9,] 1.0000000 0.2853093 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 0.7717718 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,85]     [,86] [,87]     [,88]      [,89]      [,90]     [,91]
#>  [1,] 0.0000000 0.0000000     0 0.0000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.0000000 0.0000000     0 0.0000000 0.02474326 0.00000000 0.0000000
#>  [3,] 0.0000000 0.0000000     0 0.0000000 0.33276669 0.00000000 0.1455157
#>  [4,] 0.0000000 0.0000000     0 0.0000000 0.64079012 0.00000000 0.4032823
#>  [5,] 0.2192854 0.0000000     0 0.0000000 0.94881356 0.07706729 0.6610490
#>  [6,] 0.4416304 0.0000000     0 0.0000000 1.00000000 0.36745351 0.9188156
#>  [7,] 0.6639755 0.0000000     0 0.0000000 1.00000000 0.65783973 1.0000000
#>  [8,] 0.8863205 0.0000000     0 0.2115968 1.00000000 0.94822596 1.0000000
#>  [9,] 1.0000000 0.4013903     0 0.5937040 1.00000000 1.00000000 1.0000000
#> [10,] 1.0000000 0.9173986     0 0.9758111 1.00000000 1.00000000 1.0000000
#> [11,] 1.0000000 1.0000000     1 1.0000000 1.00000000 1.00000000 1.0000000
#>             [,92]       [,93]     [,94] [,95]      [,96]     [,97]     [,98]
#>  [1,] 0.000000000 0.000000000 0.0000000     0 0.00000000 0.0000000 0.0000000
#>  [2,] 0.007347472 0.000000000 0.0986817     1 0.08829048 0.4925786 0.0000000
#>  [3,] 0.307412098 0.006982833 0.4455658     1 0.42923900 0.9851572 0.1774054
#>  [4,] 0.607476723 0.241988723 0.7924500     1 0.77018753 1.0000000 0.4422264
#>  [5,] 0.907541349 0.476994614 1.0000000     1 1.00000000 1.0000000 0.7070474
#>  [6,] 1.000000000 0.712000505 1.0000000     1 1.00000000 1.0000000 0.9718684
#>  [7,] 1.000000000 0.947006396 1.0000000     1 1.00000000 1.0000000 1.0000000
#>  [8,] 1.000000000 1.000000000 1.0000000     1 1.00000000 1.0000000 1.0000000
#>  [9,] 1.000000000 1.000000000 1.0000000     1 1.00000000 1.0000000 1.0000000
#> [10,] 1.000000000 1.000000000 1.0000000     1 1.00000000 1.0000000 1.0000000
#> [11,] 1.000000000 1.000000000 1.0000000     1 1.00000000 1.0000000 1.0000000
#>           [,99]    [,100]    [,101]    [,102]    [,103]    [,104]    [,105]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1508389 0.0000000 0.0000000 0.4912621 0.0000000 0.0000000 0.0000000
#>  [3,] 0.5296625 0.0000000 0.0000000 1.0000000 0.0000000 0.2104556 0.0000000
#>  [4,] 0.9084861 0.1448123 0.1150661 1.0000000 0.1138642 0.4832803 0.0000000
#>  [5,] 1.0000000 0.3716681 0.3402784 1.0000000 0.3390183 0.7561049 0.1403457
#>  [6,] 1.0000000 0.5985240 0.5654907 1.0000000 0.5641724 1.0000000 0.4303360
#>  [7,] 1.0000000 0.8253798 0.7907031 1.0000000 0.7893265 1.0000000 0.7203262
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>          [,106]    [,107]     [,108]    [,109]    [,110] [,111]     [,112]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000      0 0.00000000
#>  [2,] 0.0000000 0.5911356 0.00000000 0.0000000 0.3094958      1 0.00000000
#>  [3,] 0.0000000 1.0000000 0.06546268 0.0000000 0.6189916      1 0.00000000
#>  [4,] 0.0000000 1.0000000 0.30851357 0.1156513 0.9284873      1 0.04024469
#>  [5,] 0.0000000 1.0000000 0.55156446 0.3408922 1.0000000      1 0.33542332
#>  [6,] 0.0000000 1.0000000 0.79461535 0.5661331 1.0000000      1 0.63060194
#>  [7,] 0.0000000 1.0000000 1.00000000 0.7913740 1.0000000      1 0.92578056
#>  [8,] 0.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>  [9,] 0.3812339 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [10,] 0.8322200 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>           [,113]    [,114]    [,115]    [,116]     [,117] [,118]      [,119]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000      0 0.000000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000      0 0.000000000
#>  [3,] 0.00000000 0.0000000 0.0199507 0.2963706 0.00000000      0 0.000000000
#>  [4,] 0.05304745 0.1385619 0.2565481 0.5930706 0.00000000      0 0.000000000
#>  [5,] 0.27605592 0.3650400 0.4931455 0.8897706 0.00000000      0 0.000000000
#>  [6,] 0.49906439 0.5915180 0.7297429 1.0000000 0.00000000      0 0.003195058
#>  [7,] 0.72207287 0.8179960 0.9663402 1.0000000 0.02221256      0 0.313905921
#>  [8,] 0.94508134 1.0000000 1.0000000 1.0000000 0.37144576      0 0.624616783
#>  [9,] 1.00000000 1.0000000 1.0000000 1.0000000 0.72067896      0 0.935327646
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000      0 1.000000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000      1 1.000000000
#>          [,120]    [,121] [,122]      [,123]     [,124]    [,125] [,126]
#>  [1,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.0000000      0
#>  [2,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.2991271      1
#>  [3,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.5982543      1
#>  [4,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.8973814      1
#>  [5,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [6,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [7,] 0.0000000 0.0000000      0 0.002873055 0.02555713 1.0000000      1
#>  [8,] 0.0000000 0.0000000      0 0.406825891 0.37426346 1.0000000      1
#>  [9,] 0.0000000 0.2509293      0 0.810778727 0.72296980 1.0000000      1
#> [10,] 0.4623659 0.7506236      0 1.000000000 1.00000000 1.0000000      1
#> [11,] 1.0000000 1.0000000      1 1.000000000 1.00000000 1.0000000      1
#> 
#> $PIT
#>       [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]    0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,]    0 0.5005623 0.3371650 0.2247767 0.1685825 0.1348660 0.1123883
#>  [3,]    0 1.1131216 0.8404014 0.5602676 0.4202007 0.3361606 0.2801338
#>  [4,]    0 1.4581911 1.2098929 0.8575384 0.6844957 0.5475966 0.4563305
#>  [5,]    0 1.8032606 1.4016303 1.0611520 0.8942614 0.7154091 0.5961743
#>  [6,]    0 2.0000000 1.5000000 1.2025204 1.0573433 0.8458746 0.7425599
#>  [7,]    0 2.0000000 1.5000000 1.2783091 1.1712403 0.9369922 0.8676252
#>  [8,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.0290936 0.9935104
#>  [9,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.1200465 1.1000388
#> [10,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#> [11,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#>             [,8]       [,9]     [,10]     [,11]     [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000
#>  [2,] 0.09633286 0.08429125 0.1207890 0.1087101 0.0988274 0.09059178 0.08362319
#>  [3,] 0.24011469 0.21010035 0.2784829 0.2506346 0.2278496 0.20886214 0.19940850
#>  [4,] 0.39675619 0.34716166 0.4196993 0.3777293 0.3433903 0.31743848 0.31858732
#>  [5,] 0.54844143 0.47988625 0.5376767 0.5046527 0.4631714 0.45179900 0.46156691
#>  [6,] 0.70573400 0.64555452 0.6849374 0.6594228 0.6303152 0.62957545 0.64462274
#>  [7,] 0.84475195 0.79598034 0.8186492 0.8019989 0.7863737 0.79719025 0.81279100
#>  [8,] 0.98447265 0.94702106 0.9529076 0.9450670 0.9428794 0.94763942 0.95166715
#>  [9,] 1.08574753 1.06442169 1.0572637 1.0515374 1.0468521 1.04294779 1.03964412
#> [10,] 1.14285714 1.12500000 1.1111111 1.1000000 1.0909091 1.08333333 1.07692308
#> [11,] 1.14285714 1.12500000 1.1111111 1.1000000 1.0909091 1.08333333 1.07692308
#>           [,15]      [,16]      [,17]      [,18]      [,19]      [,20]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0776501 0.07247343 0.06794384 0.06394714 0.06039452 0.05721586
#>  [3,] 0.1851650 0.17282070 0.16201940 0.15248885 0.14401725 0.13643739
#>  [4,] 0.2958311 0.27610901 0.25885219 0.24362559 0.23009084 0.21798080
#>  [5,] 0.4285978 0.40769249 0.38221171 0.35972867 0.33974374 0.32945437
#>  [6,] 0.6081029 0.59016965 0.55328405 0.53768524 0.50781384 0.50042774
#>  [7,] 0.7857014 0.77086738 0.72268817 0.71431704 0.68776495 0.68265681
#>  [8,] 0.9361002 0.92617861 0.89244548 0.89128126 0.87263757 0.86954837
#>  [9,] 1.0368124 1.03435823 1.01932841 1.01819144 1.01023694 1.00969815
#> [10,] 1.0714286 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#> [11,] 1.0714286 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#>            [,21]      [,22]     [,23]      [,24]      [,25]      [,26]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.05435507 0.05176673 0.0494137 0.04726528 0.04755728 0.04565499
#>  [3,] 0.12961552 0.12993114 0.1240252 0.11863278 0.12939217 0.13147770
#>  [4,] 0.20708176 0.21589144 0.2060782 0.19711827 0.21804847 0.22721902
#>  [5,] 0.31298165 0.32893141 0.3259969 0.31182313 0.34049717 0.35540105
#>  [6,] 0.47540635 0.49580454 0.4985640 0.47688733 0.49868370 0.51789140
#>  [7,] 0.64852397 0.66526092 0.6735970 0.64431016 0.65913057 0.67276535
#>  [8,] 0.82607095 0.83435328 0.8418827 0.82207667 0.82949014 0.83631054
#>  [9,] 0.95921324 0.96115547 0.9629211 0.95790652 0.95966042 0.96127400
#> [10,] 1.00000000 1.00000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.05000000 1.04761905 1.0454545 1.04347826 1.04166667 1.04000000
#>            [,27]      [,28]      [,29]      [,30]      [,31]      [,32]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04389903 0.04227314 0.04076339 0.03935775 0.03804583 0.03682921
#>  [3,] 0.13166834 0.12679173 0.12226346 0.11804748 0.11411256 0.12002237
#>  [4,] 0.23356887 0.22491817 0.21688538 0.21188872 0.20482576 0.21738952
#>  [5,] 0.36666238 0.36086090 0.34797302 0.34616338 0.33462460 0.35258149
#>  [6,] 0.53274466 0.52902712 0.51013329 0.51043920 0.49342456 0.50976570
#>  [7,] 0.68535129 0.68421677 0.65978046 0.66263340 0.65088135 0.66214324
#>  [8,] 0.84260628 0.84388263 0.81374397 0.81899510 0.81240115 0.81845273
#>  [9,] 0.96276346 0.96414259 0.92970893 0.93213276 0.93213771 0.93432682
#> [10,] 1.00000000 1.00000000 0.96428571 0.96551724 0.96666667 0.96774194
#> [11,] 1.03846154 1.03703704 1.03571429 1.03448276 1.03333333 1.03225806
#>           [,33]      [,34]      [,35]      [,36]      [,37]      [,38]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0356783 0.03459714 0.03492616 0.03392827 0.03298582 0.03209431
#>  [3,] 0.1162717 0.11871713 0.12593277 0.12233469 0.11893650 0.11594697
#>  [4,] 0.2105961 0.21834918 0.23199519 0.22536676 0.21990705 0.22054442
#>  [5,] 0.3415633 0.35351362 0.37252793 0.36188427 0.35881504 0.36205395
#>  [6,] 0.4938355 0.50917384 0.52360991 0.50864962 0.50768619 0.51325738
#>  [7,] 0.6414513 0.65231638 0.66254237 0.64361259 0.64508280 0.65329639
#>  [8,] 0.7928761 0.79915256 0.80505984 0.78205813 0.78586525 0.79165267
#>  [9,] 0.9051291 0.90800398 0.91070975 0.88468947 0.88789254 0.89092247
#> [10,] 0.9375000 0.93939394 0.94117647 0.91428571 0.91666667 0.91891892
#> [11,] 1.0312500 1.03030303 1.02941176 1.02857143 1.02777778 1.02702703
#>            [,39]      [,40]      [,41]      [,42]      [,43]      [,44]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03124973 0.03044845 0.02968724 0.02896316 0.03499833 0.03418441
#>  [3,] 0.11844486 0.11712450 0.11419639 0.11141111 0.12673537 0.12378804
#>  [4,] 0.22747207 0.22959422 0.22511758 0.21962690 0.23820722 0.23266751
#>  [5,] 0.37244000 0.37708310 0.37449305 0.36535907 0.38046957 0.37533719
#>  [6,] 0.52606640 0.53300844 0.53209406 0.52047617 0.53189340 0.52998528
#>  [7,] 0.66242017 0.67107606 0.67228382 0.66471196 0.67269501 0.67425817
#>  [8,] 0.79713550 0.80233715 0.80583719 0.80247324 0.80717626 0.81166053
#>  [9,] 0.89379293 0.89651619 0.89910328 0.90092986 0.90328867 0.90553777
#> [10,] 0.92105263 0.92307692 0.92500000 0.92682927 0.92857143 0.93023256
#> [11,] 1.02631579 1.02564103 1.02500000 1.02439024 1.02380952 1.02325581
#>            [,45]      [,46]      [,47]      [,48]      [,49]     [,50]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.03340749 0.03266511 0.03774883 0.04213448 0.04125668 0.0404147
#>  [3,] 0.12097467 0.12405913 0.13716516 0.14887092 0.14576944 0.1427946
#>  [4,] 0.22737962 0.23658559 0.25318156 0.26907131 0.26346566 0.2580888
#>  [5,] 0.36887889 0.38290380 0.39631894 0.40916322 0.40382185 0.3955806
#>  [6,] 0.52512544 0.53567821 0.54577216 0.55543658 0.55169441 0.5404353
#>  [7,] 0.67123259 0.67853854 0.68552683 0.69221775 0.69027253 0.6761853
#>  [8,] 0.81062536 0.81483369 0.81885904 0.82271310 0.82269580 0.8114375
#>  [9,] 0.90748222 0.90953817 0.91150473 0.91338761 0.91519203 0.9107962
#> [10,] 0.93181818 0.93333333 0.93478261 0.93617021 0.93750000 0.9387755
#> [11,] 1.02272727 1.02222222 1.02173913 1.02127660 1.02083333 1.0204082
#>            [,51]      [,52]      [,53]      [,54]     [,55]      [,56]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.03960641 0.03882981 0.03808309 0.03736454 0.0366726 0.03600583
#>  [3,] 0.14194530 0.13916206 0.13648586 0.13391066 0.1314308 0.13401078
#>  [4,] 0.25991227 0.25481595 0.24991564 0.24520025 0.2406595 0.24652414
#>  [5,] 0.39963281 0.39179687 0.38426232 0.37701209 0.3700304 0.37881347
#>  [6,] 0.54656908 0.53585204 0.52554719 0.51805762 0.5084640 0.51740098
#>  [7,] 0.68266163 0.66927611 0.66294255 0.65735936 0.6451860 0.65163720
#>  [8,] 0.81520872 0.79922424 0.79834170 0.79470254 0.7799858 0.78398608
#>  [9,] 0.91258024 0.90297988 0.90484565 0.90369572 0.8869606 0.88901587
#> [10,] 0.94000000 0.93841701 0.93960129 0.94074089 0.9351857 0.93636417
#> [11,] 1.02000000 1.01960784 1.01923077 1.01886792 1.0185185 1.01818182
#>            [,57]      [,58]      [,59]      [,60]      [,61]      [,62]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03536287 0.03474247 0.03414346 0.03356475 0.03869371 0.03805939
#>  [3,] 0.13161773 0.12930865 0.12707919 0.12492531 0.13421995 0.13201963
#>  [4,] 0.24253001 0.23827509 0.23416690 0.23019797 0.24302800 0.23904394
#>  [5,] 0.37643028 0.36982624 0.36344993 0.35728976 0.36800160 0.36196878
#>  [6,] 0.51651627 0.51217792 0.50546317 0.49689599 0.50528106 0.49699776
#>  [7,] 0.65232867 0.65171009 0.64777800 0.63679872 0.64285207 0.63231351
#>  [8,] 0.78628744 0.78942116 0.78830313 0.77899897 0.78268232 0.76985147
#>  [9,] 0.89099773 0.89291005 0.89475643 0.89002099 0.89185398 0.88028697
#> [10,] 0.93750052 0.93859700 0.93965568 0.94053225 0.94152338 0.93886790
#> [11,] 1.01785714 1.01754386 1.01724138 1.01694915 1.01666667 1.01639344
#>           [,63]      [,64]      [,65]     [,66]      [,67]      [,68]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#>  [2,] 0.0465005 0.04894428 0.04817952 0.0474383 0.04671954 0.04602223
#>  [3,] 0.1460193 0.15342638 0.15102909 0.1487056 0.14645245 0.14521729
#>  [4,] 0.2513174 0.26320127 0.25908875 0.2551028 0.25123758 0.25206196
#>  [5,] 0.3722596 0.38222374 0.37625150 0.3704630 0.36484994 0.36760212
#>  [6,] 0.5051107 0.51296609 0.50495099 0.5013150 0.49796818 0.50235703
#>  [7,] 0.6382439 0.64398610 0.63392382 0.6328109 0.63190409 0.63739806
#>  [8,] 0.7735635 0.77715777 0.76501468 0.7663922 0.76789385 0.77135813
#>  [9,] 0.8822178 0.88408738 0.87027352 0.8722693 0.87420462 0.87608217
#> [10,] 0.9398539 0.94080860 0.92610847 0.9272453 0.92834761 0.92941705
#> [11,] 1.0161290 1.01587302 1.01562500 1.0153846 1.01515152 1.01492537
#>            [,69]      [,70]      [,71]      [,72]      [,73]      [,74]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04534543 0.04468825 0.04681088 0.04615158 0.04551058 0.04488715
#>  [3,] 0.14308174 0.14100809 0.14757198 0.14549350 0.14347275 0.14150737
#>  [4,] 0.24835517 0.24475582 0.25554502 0.25194580 0.24844655 0.24504317
#>  [5,] 0.36219621 0.35802415 0.36719524 0.36202347 0.35699537 0.35210502
#>  [6,] 0.49616920 0.49332673 0.50056492 0.49351471 0.48666033 0.47999375
#>  [7,] 0.63369474 0.63213036 0.63738564 0.62840837 0.61968048 0.61119171
#>  [8,] 0.77015521 0.76988436 0.77317173 0.76228198 0.75169473 0.74139754
#>  [9,] 0.87780960 0.87924976 0.88097476 0.86856667 0.85650324 0.84477032
#> [10,] 0.93045503 0.93146293 0.93244203 0.92805632 0.92318536 0.91053898
#> [11,] 1.01470588 1.01449275 1.01428571 1.01408451 1.01388889 1.01369863
#>            [,75]      [,76]      [,77]     [,78]      [,79]      [,80]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000
#>  [2,] 0.05152726 0.05084023 0.05017128 0.0495197 0.04888483 0.04826604
#>  [3,] 0.15310863 0.15106718 0.14907945 0.1471434 0.14525690 0.14341821
#>  [4,] 0.25524529 0.25184202 0.24852831 0.2459945 0.24284073 0.23976680
#>  [5,] 0.36086036 0.35604889 0.35353064 0.3525295 0.34800989 0.34360470
#>  [6,] 0.48702087 0.48415372 0.48288147 0.4830968 0.47690322 0.47473771
#>  [7,] 0.61644587 0.61575765 0.61568535 0.6170723 0.60916107 0.60971250
#>  [8,] 0.74489217 0.74639592 0.74753628 0.7501072 0.74049040 0.74377052
#>  [9,] 0.84686802 0.84890978 0.85089781 0.8528342 0.84555824 0.84751320
#> [10,] 0.91174791 0.91292461 0.91407034 0.9151863 0.91334766 0.91444453
#> [11,] 1.01351351 1.01333333 1.01315789 1.0129870 1.01282051 1.01265823
#>            [,81]      [,82]      [,83]      [,84]      [,85]      [,86]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04766271 0.05021363 0.04960127 0.04900366 0.04842029 0.04785064
#>  [3,] 0.14162548 0.14858684 0.15179982 0.14997091 0.14818554 0.14644218
#>  [4,] 0.23788961 0.24729838 0.25453800 0.25280842 0.24979880 0.24685999
#>  [5,] 0.34323093 0.35133919 0.35924969 0.35896958 0.35730667 0.35310306
#>  [6,] 0.47552617 0.48200115 0.48831821 0.48919411 0.48862787 0.48287931
#>  [7,] 0.61161516 0.61641004 0.62108797 0.62307527 0.62356218 0.61622615
#>  [8,] 0.74679886 0.74992480 0.75297450 0.75595071 0.75750273 0.74859093
#>  [9,] 0.84941928 0.85127830 0.85309198 0.85486196 0.85658979 0.85123450
#> [10,] 0.91551397 0.91655701 0.91757460 0.91856768 0.91953711 0.91951195
#> [11,] 1.01250000 1.01234568 1.01219512 1.01204819 1.01190476 1.01176471
#>            [,87]      [,88]      [,89]      [,90]      [,91]     [,92]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.04729423 0.04675062 0.04650054 0.04597806 0.04546719 0.0450483
#>  [3,] 0.14473936 0.14307569 0.14523127 0.14359946 0.14362075 0.1454207
#>  [4,] 0.24398953 0.24118505 0.24572602 0.24296505 0.24474635 0.2487324
#>  [5,] 0.34899721 0.34498575 0.35184743 0.34876001 0.35222989 0.3583322
#>  [6,] 0.47726443 0.47177863 0.47778115 0.47654151 0.48145567 0.4871540
#>  [7,] 0.60906073 0.60206004 0.60658208 0.60715801 0.61152292 0.6157919
#>  [8,] 0.73988639 0.73381409 0.73683893 0.73921407 0.74211169 0.7449456
#>  [9,] 0.84133643 0.83849008 0.84032542 0.84211952 0.84387374 0.8455894
#> [10,] 0.90881996 0.90958997 0.91061736 0.91162165 0.91260364 0.9135640
#> [11,] 1.01162791 1.01149425 1.01136364 1.01123596 1.01111111 1.0109890
#>            [,93]      [,94]      [,95]      [,96]      [,97]      [,98]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04455864 0.04514061 0.05529869 0.05564597 0.06019735 0.05957676
#>  [3,] 0.14391590 0.14715945 0.15623222 0.15910597 0.16771067 0.16781062
#>  [4,] 0.24865910 0.25450631 0.26243710 0.26778184 0.27540911 0.27712888
#>  [5,] 0.35962202 0.36650781 0.37324709 0.37984449 0.38630444 0.38961107
#>  [6,] 0.48959794 0.49508614 0.50045756 0.50571590 0.51086470 0.51561731
#>  [7,] 0.61939206 0.62348462 0.62749010 0.63141126 0.63525072 0.63901102
#>  [8,] 0.74771796 0.75043067 0.75308566 0.75568476 0.75822971 0.76072219
#>  [9,] 0.84726779 0.84891007 0.85051741 0.85209091 0.85363163 0.85514059
#> [10,] 0.91450356 0.91542287 0.91632263 0.91720344 0.91806591 0.91891059
#> [11,] 1.01086957 1.01075269 1.01063830 1.01052632 1.01041667 1.01030928
#>            [,99]     [,100]     [,101]     [,102]     [,103]     [,104]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06050801 0.05989681 0.05929785 0.06357472 0.06295144 0.06234026
#>  [3,] 0.17150299 0.16977063 0.16807293 0.17630983 0.17458130 0.17492959
#>  [4,] 0.28357130 0.28216969 0.28049866 0.28762243 0.28591892 0.28783505
#>  [5,] 0.39583953 0.39559537 0.39504220 0.40103188 0.40042391 0.40387712
#>  [6,] 0.52055999 0.52134751 0.52178894 0.52652370 0.52689281 0.53148608
#>  [7,] 0.64269458 0.64453989 0.64600152 0.64950646 0.65087724 0.65426678
#>  [8,] 0.76316380 0.76555608 0.76790052 0.77019854 0.77245149 0.77466070
#>  [9,] 0.85661874 0.85806704 0.85948637 0.86087759 0.86224154 0.86357900
#> [10,] 0.91973803 0.92054876 0.92134327 0.92212205 0.92288556 0.92363425
#> [11,] 1.01020408 1.01010101 1.01000000 1.00990099 1.00980392 1.00970874
#>           [,105]     [,106]     [,107]     [,108]     [,109]     [,110]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06174083 0.06115283 0.06615266 0.06553441 0.06492761 0.06717136
#>  [3,] 0.17324758 0.17159760 0.17941272 0.17834777 0.17669640 0.18075415
#>  [4,] 0.28506740 0.28235248 0.28912274 0.28930396 0.28769606 0.29357488
#>  [5,] 0.40134316 0.39752085 0.40320461 0.40459115 0.40400135 0.40946923
#>  [6,] 0.53051348 0.52546097 0.52993776 0.53241138 0.53272362 0.53701056
#>  [7,] 0.65490197 0.64866481 0.65197929 0.65523182 0.65649240 0.65964384
#>  [8,] 0.77682742 0.76942907 0.77160426 0.77373880 0.77583382 0.77789039
#>  [9,] 0.86489074 0.86028448 0.86160255 0.86289599 0.86416547 0.86541166
#> [10,] 0.92436853 0.92349093 0.92421271 0.92492100 0.92561618 0.92629860
#> [11,] 1.00961538 1.00952381 1.00943396 1.00934579 1.00925926 1.00917431
#>           [,111]     [,112]    [,113]     [,114]     [,115]     [,116]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07565162 0.07497007 0.0743007 0.07364317 0.07299718 0.07236242
#>  [3,] 0.18820184 0.18650633 0.1848411 0.18320533 0.18177327 0.18276977
#>  [4,] 0.29999693 0.29765682 0.2954728 0.29408421 0.29375495 0.29635770
#>  [5,] 0.41483769 0.41412224 0.4128895 0.41246606 0.41317378 0.41731810
#>  [6,] 0.54121955 0.54202480 0.5416412 0.54208261 0.54372875 0.54769633
#>  [7,] 0.66273799 0.66510774 0.6656164 0.66696485 0.66959095 0.67246407
#>  [8,] 0.77990956 0.78189236 0.7833494 0.78526667 0.78715029 0.78900116
#>  [9,] 0.86663519 0.86783667 0.8690167 0.87017585 0.87131466 0.87243366
#> [10,] 0.92696861 0.92762655 0.9282727 0.92890750 0.92953112 0.93014389
#> [11,] 1.00909091 1.00900901 1.0089286 1.00884956 1.00877193 1.00869565
#>          [,117]     [,118]    [,119]     [,120]     [,121]    [,122]     [,123]
#>  [1,] 0.0000000 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.0717386 0.07112545 0.0705227 0.06993007 0.06934732 0.0687742 0.06821048
#>  [3,] 0.1811942 0.17964550 0.1781231 0.17662625 0.17515436 0.1737068 0.17228298
#>  [4,] 0.2938029 0.29129175 0.2888232 0.28639609 0.28400946 0.2816623 0.27935357
#>  [5,] 0.4137205 0.41018445 0.4067083 0.40329060 0.39992984 0.3966246 0.39337362
#>  [6,] 0.5429748 0.53833400 0.5337989 0.52931322 0.52490227 0.5205642 0.51629732
#>  [7,] 0.6668585 0.66115881 0.6582160 0.65268476 0.64724572 0.6418966 0.63665869
#>  [8,] 0.7854015 0.77868871 0.7773830 0.77085039 0.76442663 0.7581091 0.75522969
#>  [9,] 0.8711254 0.86367991 0.8642871 0.85702418 0.85197339 0.8449323 0.84465234
#> [10,] 0.9307461 0.92279100 0.9234453 0.91957070 0.91816281 0.9105747 0.91130768
#> [11,] 1.0086207 1.00854701 1.0084746 1.00840336 1.00833333 1.0082645 1.00819672
#>           [,124]     [,125]     [,126]
#>  [1,] 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06765592 0.06952262 0.07696644
#>  [3,] 0.17088231 0.17432885 0.18093422
#>  [4,] 0.27708240 0.28208481 0.28782813
#>  [5,] 0.39017546 0.39509340 0.39993265
#>  [6,] 0.51209978 0.51603446 0.51990618
#>  [7,] 0.63169038 0.63466062 0.63758334
#>  [8,] 0.75213240 0.75413133 0.75609828
#>  [9,] 0.84366305 0.84492383 0.84616444
#> [10,] 0.91202875 0.91273820 0.91343629
#> [11,] 1.00813008 1.00806452 1.00800000
#>