Non-randomized Probability Integral Transform

Functions to produce the non-randomzied probability integral transform (PIT) to check the adequacy of the distributional assumption of the COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

compPredProb(object)

compPIT(object, bins = 10)

Arguments

object
an object class "cmp", obtained from a call to glm.cmp.

bins
numeric; the number of bins shown in the PIT histogram or the PIT Q-Q plot.

Value

compPredprob returns a list with values:

upper

the predictive cumulative probabilities used as the upper bound for computing the non-randomzied PIT.

lower

the predictive cumulative probabilities used as the upper bound for computing the non-randomzied PIT.

compPIT returns a list with values:
conditionalPIT

the conditional probability integral transoformation given the observed counts.

PIT

the probability integral transformation.

Details

These functions are used to obtain the predictive probabilities and the probability integral transform for a fitted COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

References

Czado, C., Gneiting, T. and Held, L. (2009). Predictive model assessment for count data. Biometrics, 65, 1254--1261.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Seires Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
compPredProb(M.bids)

#> $lower
#>   [1] 0.166805415 0.000000000 0.055599393 0.232373512 0.227137557 0.668144192
#>   [7] 0.423152672 0.281975122 0.402309908 0.000000000 0.306343006 0.383133009
#>  [13] 0.288795208 0.164704574 0.455432198 0.348338393 0.605197227 0.401196004
#>  [19] 0.525906353 0.335037091 0.962279201 0.146341240 0.309147694 0.616311596
#>  [25] 0.082833091 0.131301009 0.146275277 0.305175375 0.906687578 0.267409891
#>  [31] 0.500205504 0.099519857 0.905491762 0.126510976 0.085261554 0.996711866
#>  [37] 0.286671329 0.196047851 0.122346507 0.172056139 0.276951501 0.481669348
#>  [43] 0.039991341 0.344635399 0.359159723 0.131581073 0.041863493 0.044749544
#>  [49] 0.331137981 0.636894630 0.159279009 0.703208053 0.517699114 0.445815159
#>  [55] 0.806701058 0.105328668 0.289350276 0.422406222 0.459044972 0.636401879
#>  [61] 0.000000000 0.768839537 0.000000000 0.051093151 0.997773976 0.408094138
#>  [67] 0.403911765 0.173358976 0.473023166 0.366756234 0.052255019 0.830475124
#>  [73] 0.821219802 0.981921799 0.021610698 0.406896105 0.325741904 0.275658299
#>  [79] 0.741509870 0.411629886 0.259633938 0.043388970 0.103549408 0.250274932
#>  [85] 0.301000686 0.722404826 0.942985760 0.644693006 0.091613074 0.373194032
#>  [91] 0.143143158 0.097184287 0.196602397 0.071224924 0.000000000 0.073771055
#>  [97] 0.000000000 0.132609646 0.059874764 0.235756111 0.248507373 0.037093797
#> [103] 0.249013602 0.122468325 0.351307873 0.715610569 0.019381139 0.172655126
#> [109] 0.248257264 0.000000000 0.000000000 0.285991605 0.275825850 0.238415116
#> [115] 0.191148112 0.099744675 0.593647454 0.994920178 0.498869920 0.849778970
#> [121] 0.749958016 0.942362966 0.599337335 0.592682196 0.000000000 0.002004289
#> 
#> $upper
#>   [1] 0.45684925 0.19935618 0.30942285 0.67270190 0.66581387 0.88818369
#>   [7] 0.76279590 0.73147819 0.83704796 0.24185607 0.75661213 0.72736921
#>  [13] 0.62852044 0.57093921 0.78896242 0.79511902 0.85086994 0.74378788
#>  [19] 0.83941500 0.78353145 0.99580615 0.53756215 0.65189503 0.83333059
#>  [25] 0.39294058 0.50785613 0.53743678 0.75545758 0.97910970 0.71534804
#>  [31] 0.82201716 0.43642163 0.97873238 0.49789293 0.39957738 0.99927782
#>  [37] 0.73649680 0.62174522 0.48901781 0.58350749 0.72601290 0.80876909
#>  [43] 0.25144873 0.68984272 0.80415954 0.39360421 0.25892429 0.27014285
#>  [49] 0.78003151 0.87031028 0.56138275 0.93248051 0.75990283 0.86566309
#>  [55] 0.95265692 0.45049138 0.73932127 0.71028346 0.79176537 0.90254306
#>  [61] 0.29262627 0.93739182 0.17771200 0.29362231 0.99974077 0.74986944
#>  [67] 0.74619437 0.58569079 0.80238128 0.81030735 0.29776193 0.94281822
#>  [73] 0.95804151 0.99650028 0.16725186 0.74882055 0.77510820 0.72458952
#>  [79] 0.94710744 0.70029178 0.70637058 0.26489879 0.33676123 0.69520863
#>  [85] 0.75128987 0.91626537 0.99238082 0.90657625 0.41641993 0.71802014
#>  [91] 0.53143566 0.43061879 0.62258248 0.35954606 0.05723997 0.36712203
#>  [97] 0.20260585 0.51053400 0.32381374 0.67707634 0.69305532 0.16450488
#> [103] 0.69367353 0.48928032 0.69660626 0.93741341 0.15523856 0.58451289
#> [109] 0.69274944 0.32273754 0.02494186 0.62520158 0.72477432 0.68047434
#> [115] 0.61426018 0.43697567 0.88024166 0.99914624 0.82108258 0.95889526
#> [121] 0.95009460 0.98931835 0.84712832 0.87970834 0.33394692 0.02913135
#> 

compPIT(M.bids)

#> $ConditionalPIT
#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.5016147 0.1749271 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1144468 1.0000000 0.5689017 0.0000000 0.0000000 0.0000000 0.0000000
#>  [4,] 0.4592223 1.0000000 0.9628764 0.1535819 0.1660961 0.0000000 0.0000000
#>  [5,] 0.8039977 1.0000000 1.0000000 0.3806852 0.3940547 0.0000000 0.0000000
#>  [6,] 1.0000000 1.0000000 1.0000000 0.6077884 0.6220132 0.0000000 0.2262590
#>  [7,] 1.0000000 1.0000000 1.0000000 0.8348916 0.8499717 0.0000000 0.5206856
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.1447731 0.8151122
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.5992370 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>             [,8]      [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.0000000 0.4134691 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.8269381 0.0000000 0.00000000 0.00000000 0.08688434
#>  [4,] 0.04009957 0.0000000 1.0000000 0.0000000 0.00000000 0.03298193 0.33304749
#>  [5,] 0.26256746 0.0000000 1.0000000 0.2080023 0.04899831 0.32733745 0.57921065
#>  [6,] 0.48503535 0.2247102 1.0000000 0.4300917 0.33949652 0.62169298 0.82537380
#>  [7,] 0.70750324 0.4547338 1.0000000 0.6521811 0.62999473 0.91604850 1.00000000
#>  [8,] 0.92997114 0.6847574 1.0000000 0.8742705 0.92049295 1.00000000 1.00000000
#>  [9,] 1.00000000 0.9147810 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20] [,21]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [4,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [5,] 0.0000000 0.1156308 0.0000000 0.0000000 0.0000000 0.1448467     0
#>  [6,] 0.1336245 0.3394543 0.0000000 0.2884015 0.0000000 0.3678149     0
#>  [7,] 0.4334474 0.5632778 0.0000000 0.5802940 0.2363369 0.5907831     0
#>  [8,] 0.7332703 0.7871013 0.3858905 0.8721865 0.5553073 0.8137514     0
#>  [9,] 1.0000000 1.0000000 0.7929361 1.0000000 0.8742778 1.0000000     0
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     0
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     1
#>           [,22]     [,23]     [,24]      [,25]     [,26]     [,27]     [,28]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.05535793 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1371572 0.0000000 0.0000000 0.37782676 0.1824407 0.1373467 0.0000000
#>  [4,] 0.3927672 0.0000000 0.0000000 0.70029560 0.4480061 0.3929955 0.0000000
#>  [5,] 0.6483773 0.2650708 0.0000000 1.00000000 0.7135715 0.6486444 0.2105893
#>  [6,] 0.9039874 0.5568309 0.0000000 1.00000000 0.9791368 0.9042933 0.4326723
#>  [7,] 1.0000000 0.8485910 0.0000000 1.00000000 1.0000000 1.0000000 0.6547552
#>  [8,] 1.0000000 1.0000000 0.3856271 1.00000000 1.0000000 1.0000000 0.8768382
#>  [9,] 1.0000000 1.0000000 0.8464163 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>       [,29]      [,30]     [,31]       [,32] [,33]     [,34]      [,35] [,36]
#>  [1,]     0 0.00000000 0.0000000 0.000000000     0 0.0000000 0.00000000     0
#>  [2,]     0 0.00000000 0.0000000 0.001425171     0 0.0000000 0.04689056     0
#>  [3,]     0 0.00000000 0.0000000 0.298247597     0 0.1978799 0.36504190     0
#>  [4,]     0 0.07275582 0.0000000 0.595070023     0 0.4671445 0.68319323     0
#>  [5,]     0 0.29600093 0.0000000 0.891892449     0 0.7364090 1.00000000     0
#>  [6,]     0 0.51924603 0.0000000 1.000000000     0 1.0000000 1.00000000     0
#>  [7,]     0 0.74249114 0.3101022 1.000000000     0 1.0000000 1.00000000     0
#>  [8,]     0 0.96573624 0.6208430 1.000000000     0 1.0000000 1.00000000     0
#>  [9,]     0 1.00000000 0.9315837 1.000000000     0 1.0000000 1.00000000     0
#> [10,]     0 1.00000000 1.0000000 1.000000000     0 1.0000000 1.00000000     0
#> [11,]     1 1.00000000 1.0000000 1.000000000     1 1.0000000 1.00000000     1
#>            [,37]      [,38]     [,39]      [,40]      [,41]      [,42]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.00928394 0.2117796 0.06791535 0.00000000 0.00000000
#>  [4,] 0.02963076 0.24419260 0.4845034 0.31095745 0.05132594 0.00000000
#>  [5,] 0.25193920 0.47910127 0.7572272 0.55399955 0.27401264 0.00000000
#>  [6,] 0.47424765 0.71400993 1.0000000 0.79704165 0.49669933 0.05603995
#>  [7,] 0.69655609 0.94891859 1.0000000 1.00000000 0.71938603 0.36175709
#>  [8,] 0.91886454 1.00000000 1.0000000 1.00000000 0.94207272 0.66747424
#>  [9,] 1.00000000 1.00000000 1.0000000 1.00000000 1.00000000 0.97319139
#> [10,] 1.00000000 1.00000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.00000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,43]     [,44]      [,45]     [,46]     [,47]     [,48]     [,49]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.2837861 0.0000000 0.00000000 0.0000000 0.2678351 0.2451291 0.0000000
#>  [3,] 0.7566946 0.0000000 0.00000000 0.2611179 0.7285355 0.6887980 0.0000000
#>  [4,] 1.0000000 0.0000000 0.00000000 0.6427636 1.0000000 1.0000000 0.0000000
#>  [5,] 1.0000000 0.1603807 0.09177594 1.0000000 1.0000000 1.0000000 0.1534039
#>  [6,] 1.0000000 0.4500617 0.31649513 1.0000000 1.0000000 1.0000000 0.3761739
#>  [7,] 1.0000000 0.7397427 0.54121432 1.0000000 1.0000000 1.0000000 0.5989439
#>  [8,] 1.0000000 1.0000000 0.76593352 1.0000000 1.0000000 1.0000000 0.8217138
#>  [9,] 1.0000000 1.0000000 0.99065271 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,50]     [,51]     [,52]     [,53]     [,54]     [,55]     [,56]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.1012699 0.0000000 0.0000000 0.0000000 0.0000000 0.2742803
#>  [4,] 0.0000000 0.3499619 0.0000000 0.0000000 0.0000000 0.0000000 0.5639987
#>  [5,] 0.0000000 0.5986539 0.0000000 0.0000000 0.0000000 0.0000000 0.8537172
#>  [6,] 0.0000000 0.8473460 0.0000000 0.0000000 0.1290583 0.0000000 1.0000000
#>  [7,] 0.0000000 1.0000000 0.0000000 0.3398003 0.3672397 0.0000000 1.0000000
#>  [8,] 0.2703562 1.0000000 0.0000000 0.7526759 0.6054212 0.0000000 1.0000000
#>  [9,] 0.6987765 1.0000000 0.4221700 1.0000000 0.8436027 0.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 0.8583323 1.0000000 1.0000000 0.6392271 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>            [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.3417328 0.0000000 0.5627082
#>  [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.6834656 0.0000000 1.0000000
#>  [4,] 0.02366758 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [5,] 0.24590413 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [6,] 0.46814067 0.2695377 0.1230914 0.0000000 1.0000000 0.0000000 1.0000000
#>  [7,] 0.69037722 0.6169080 0.4236441 0.0000000 1.0000000 0.0000000 1.0000000
#>  [8,] 0.91261376 0.9642783 0.7241967 0.2389638 1.0000000 0.0000000 1.0000000
#>  [9,] 1.00000000 1.0000000 1.0000000 0.6147043 1.0000000 0.1848712 1.0000000
#> [10,] 1.00000000 1.0000000 1.0000000 0.9904447 1.0000000 0.7781589 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,64] [,65]     [,66]     [,67]      [,68]     [,69]      [,70]
#>  [1,] 0.0000000     0 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.2016535     0 0.0000000 0.0000000 0.00000000 0.0000000 0.00000000
#>  [3,] 0.6139750     0 0.0000000 0.0000000 0.06461064 0.0000000 0.00000000
#>  [4,] 1.0000000     0 0.0000000 0.0000000 0.30713377 0.0000000 0.00000000
#>  [5,] 1.0000000     0 0.0000000 0.0000000 0.54965689 0.0000000 0.07494912
#>  [6,] 1.0000000     0 0.2689073 0.2807278 0.79218001 0.0819073 0.30040228
#>  [7,] 1.0000000     0 0.5614972 0.5728840 1.00000000 0.3855282 0.52585543
#>  [8,] 1.0000000     0 0.8540871 0.8650403 1.00000000 0.6891491 0.75130859
#>  [9,] 1.0000000     0 1.0000000 1.0000000 1.00000000 0.9927700 0.97676175
#> [10,] 1.0000000     0 1.0000000 1.0000000 1.00000000 1.0000000 1.00000000
#> [11,] 1.0000000     1 1.0000000 1.0000000 1.00000000 1.0000000 1.00000000
#>           [,71]    [,72]     [,73] [,74]     [,75]     [,76]     [,77]
#>  [1,] 0.0000000 0.000000 0.0000000     0 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1944751 0.000000 0.0000000     0 0.5382359 0.0000000 0.0000000
#>  [3,] 0.6017956 0.000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [4,] 1.0000000 0.000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [5,] 1.0000000 0.000000 0.0000000     0 1.0000000 0.0000000 0.1652507
#>  [6,] 1.0000000 0.000000 0.0000000     0 1.0000000 0.2722938 0.3877863
#>  [7,] 1.0000000 0.000000 0.0000000     0 1.0000000 0.5647560 0.6103219
#>  [8,] 1.0000000 0.000000 0.0000000     0 1.0000000 0.8572183 0.8328575
#>  [9,] 1.0000000 0.000000 0.0000000     0 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 0.618862 0.5757873     0 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.000000 1.0000000     1 1.0000000 1.0000000 1.0000000
#>            [,78]     [,79]     [,80]      [,81]     [,82]     [,83]     [,84]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.00000000 0.2555689 0.0000000 0.0000000
#>  [3,] 0.00000000 0.0000000 0.0000000 0.00000000 0.7070162 0.4135751 0.0000000
#>  [4,] 0.05422145 0.0000000 0.0000000 0.09035763 1.0000000 0.8423698 0.1117584
#>  [5,] 0.27697272 0.0000000 0.0000000 0.31420316 1.0000000 1.0000000 0.3365110
#>  [6,] 0.49972399 0.0000000 0.3061371 0.53804869 1.0000000 1.0000000 0.5612635
#>  [7,] 0.72247526 0.0000000 0.6525631 0.76189421 1.0000000 1.0000000 0.7860161
#>  [8,] 0.94522653 0.0000000 0.9989892 0.98573974 1.0000000 1.0000000 1.0000000
#>  [9,] 1.00000000 0.2844884 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.00000000 0.7708755 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>           [,85]     [,86] [,87]     [,88]      [,89]      [,90]     [,91]
#>  [1,] 0.0000000 0.0000000     0 0.0000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.0000000 0.0000000     0 0.0000000 0.02582127 0.00000000 0.0000000
#>  [3,] 0.0000000 0.0000000     0 0.0000000 0.33369654 0.00000000 0.1464279
#>  [4,] 0.0000000 0.0000000     0 0.0000000 0.64157181 0.00000000 0.4039657
#>  [5,] 0.2198572 0.0000000     0 0.0000000 0.94944708 0.07773764 0.6615035
#>  [6,] 0.4419367 0.0000000     0 0.0000000 1.00000000 0.36773888 0.9190413
#>  [7,] 0.6640162 0.0000000     0 0.0000000 1.00000000 0.65774013 1.0000000
#>  [8,] 0.8860957 0.0000000     0 0.2111895 1.00000000 0.94774137 1.0000000
#>  [9,] 1.0000000 0.4002629     0 0.5930391 1.00000000 1.00000000 1.0000000
#> [10,] 1.0000000 0.9160976     0 0.9748886 1.00000000 1.00000000 1.0000000
#> [11,] 1.0000000 1.0000000     1 1.0000000 1.00000000 1.00000000 1.0000000
#>             [,92]       [,93]      [,94] [,95]      [,96]     [,97]     [,98]
#>  [1,] 0.000000000 0.000000000 0.00000000     0 0.00000000 0.0000000 0.0000000
#>  [2,] 0.008444576 0.000000000 0.09980217     1 0.08941148 0.4935692 0.0000000
#>  [3,] 0.308353552 0.007975966 0.44663766     1 0.43030007 0.9871383 0.1783170
#>  [4,] 0.608262528 0.242728725 0.79347314     1 0.77118866 1.0000000 0.4429203
#>  [5,] 0.908171504 0.477481483 1.00000000     1 1.00000000 1.0000000 0.7075235
#>  [6,] 1.000000000 0.712234241 1.00000000     1 1.00000000 1.0000000 0.9721267
#>  [7,] 1.000000000 0.946987000 1.00000000     1 1.00000000 1.0000000 1.0000000
#>  [8,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#>  [9,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#> [10,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#> [11,] 1.000000000 1.000000000 1.00000000     1 1.00000000 1.0000000 1.0000000
#>           [,99]    [,100]    [,101]    [,102]    [,103]    [,104]    [,105]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1520247 0.0000000 0.0000000 0.4937263 0.0000000 0.0000000 0.0000000
#>  [3,] 0.5309001 0.0000000 0.0000000 1.0000000 0.0000000 0.2113662 0.0000000
#>  [4,] 0.9097756 0.1455720 0.1158314 1.0000000 0.1146638 0.4839855 0.0000000
#>  [5,] 1.0000000 0.3721649 0.3407790 1.0000000 0.3395548 0.7566047 0.1410146
#>  [6,] 1.0000000 0.5987577 0.5657267 1.0000000 0.5644457 1.0000000 0.4306192
#>  [7,] 1.0000000 0.8253505 0.7906743 1.0000000 0.7893367 1.0000000 0.7202238
#>  [8,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [9,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>          [,106]    [,107]     [,108]    [,109]    [,110] [,111]     [,112]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000      0 0.00000000
#>  [2,] 0.0000000 0.5934079 0.00000000 0.0000000 0.3098493      1 0.00000000
#>  [3,] 0.0000000 1.0000000 0.06639398 0.0000000 0.6196986      1 0.00000000
#>  [4,] 0.0000000 1.0000000 0.30919625 0.1164087 0.9295479      1 0.04129712
#>  [5,] 0.0000000 1.0000000 0.55199852 0.3413845 1.0000000      1 0.33609977
#>  [6,] 0.0000000 1.0000000 0.79480079 0.5663603 1.0000000      1 0.63090242
#>  [7,] 0.0000000 1.0000000 1.00000000 0.7913362 1.0000000      1 0.92570507
#>  [8,] 0.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>  [9,] 0.3804705 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [10,] 0.8313213 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>           [,113]    [,114]     [,115]       [,116]     [,117] [,118]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.0000000000 0.00000000      0
#>  [2,] 0.00000000 0.0000000 0.00000000 0.0007571225 0.00000000      0
#>  [3,] 0.00000000 0.0000000 0.02092091 0.2972897725 0.00000000      0
#>  [4,] 0.05384616 0.1393137 0.25726491 0.5938224225 0.00000000      0
#>  [5,] 0.27658887 0.3655277 0.49360891 0.8903550724 0.00000000      0
#>  [6,] 0.49933158 0.5917417 0.72995292 1.0000000000 0.00000000      0
#>  [7,] 0.72207429 0.8179558 0.96629692 1.0000000000 0.02216565      0
#>  [8,] 0.94481700 1.0000000 1.00000000 1.0000000000 0.37109105      0
#>  [9,] 1.00000000 1.0000000 1.00000000 1.0000000000 0.72001645      0
#> [10,] 1.00000000 1.0000000 1.00000000 1.0000000000 1.00000000      0
#> [11,] 1.00000000 1.0000000 1.00000000 1.0000000000 1.00000000      1
#>            [,119]    [,120]    [,121] [,122]      [,123]     [,124]    [,125]
#>  [1,] 0.000000000 0.0000000 0.0000000      0 0.000000000 0.00000000 0.0000000
#>  [2,] 0.000000000 0.0000000 0.0000000      0 0.000000000 0.00000000 0.2994488
#>  [3,] 0.000000000 0.0000000 0.0000000      0 0.000000000 0.00000000 0.5988976
#>  [4,] 0.000000000 0.0000000 0.0000000      0 0.000000000 0.00000000 0.8983464
#>  [5,] 0.000000000 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000
#>  [6,] 0.003507249 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000
#>  [7,] 0.313861287 0.0000000 0.0000000      0 0.002674289 0.02549525 1.0000000
#>  [8,] 0.624215326 0.0000000 0.0000000      0 0.406240228 0.37389557 1.0000000
#>  [9,] 0.934569364 0.0000000 0.2500392      0 0.809806168 0.72229588 1.0000000
#> [10,] 1.000000000 0.4602524 0.7496979      0 1.000000000 1.00000000 1.0000000
#> [11,] 1.000000000 1.0000000 1.0000000      1 1.000000000 1.00000000 1.0000000
#>       [,126]
#>  [1,]      0
#>  [2,]      1
#>  [3,]      1
#>  [4,]      1
#>  [5,]      1
#>  [6,]      1
#>  [7,]      1
#>  [8,]      1
#>  [9,]      1
#> [10,]      1
#> [11,]      1
#> 
#> $PIT
#>       [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]    0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,]    0 0.5016147 0.3382709 0.2255140 0.1691355 0.1353084 0.1127570
#>  [3,]    0 1.1144468 0.8416743 0.5611162 0.4208371 0.3366697 0.2805581
#>  [4,]    0 1.4592223 1.2110493 0.8585602 0.6854442 0.5483553 0.4569628
#>  [5,]    0 1.8039977 1.4019989 1.0615610 0.8946844 0.7157475 0.5964563
#>  [6,]    0 2.0000000 1.5000000 1.2025961 1.0574504 0.8459603 0.7426768
#>  [7,]    0 2.0000000 1.5000000 1.2782972 1.1712158 0.9369727 0.8675915
#>  [8,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.0289546 0.9933142
#>  [9,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.1198474 1.0998728
#> [10,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#> [11,]    0 2.0000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#>             [,8]       [,9]     [,10]     [,11]     [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000
#>  [2,] 0.09664884 0.08456773 0.1211123 0.1090011 0.0990919 0.09083424 0.08384699
#>  [3,] 0.24047836 0.21041857 0.2789207 0.2510287 0.2282079 0.20919055 0.19978238
#>  [4,] 0.39741090 0.34773453 0.4202085 0.3781876 0.3438069 0.31790485 0.31906967
#>  [5,] 0.54875786 0.48016313 0.5379228 0.5049307 0.4634823 0.45213692 0.46191182
#>  [6,] 0.70587084 0.64572577 0.6850896 0.6595898 0.6304904 0.62975727 0.64480470
#>  [7,] 0.84472173 0.79597324 0.8186429 0.8019967 0.7863602 0.79716752 0.81277002
#>  [8,] 0.98426521 0.94682673 0.9527349 0.9448884 0.9426707 0.94744810 0.95149056
#>  [9,] 1.08560529 1.06425225 1.0571131 1.0514018 1.0467289 1.04283483 1.03953984
#> [10,] 1.14285714 1.12500000 1.1111111 1.1000000 1.0909091 1.08333333 1.07692308
#> [11,] 1.14285714 1.12500000 1.1111111 1.1000000 1.0909091 1.08333333 1.07692308
#>            [,15]      [,16]      [,17]      [,18]      [,19]     [,20]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.07785792 0.07266739 0.06812568 0.06411829 0.06055616 0.0573690
#>  [3,] 0.18551221 0.17314473 0.16232319 0.15277476 0.14428728 0.1366932
#>  [4,] 0.29627898 0.27652705 0.25924411 0.24399445 0.23043920 0.2183108
#>  [5,] 0.42891812 0.40803230 0.38253028 0.36002850 0.34002692 0.3297543
#>  [6,] 0.60829183 0.59036933 0.55347124 0.53787890 0.50799674 0.5006187
#>  [7,] 0.78567554 0.77084903 0.72267096 0.71429585 0.68774257 0.6826394
#>  [8,] 0.93590339 0.92598325 0.89222746 0.89104858 0.87239628 0.8693097
#>  [9,] 1.03671557 1.03426786 1.01918463 1.01805612 1.01006844 1.0095385
#> [10,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.0526316
#> [11,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.0526316
#>            [,21]      [,22]      [,23]      [,24]     [,25]      [,26]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.05450055 0.05190528 0.04954595 0.04739178 0.0477237 0.04581475
#>  [3,] 0.12985855 0.13020610 0.12428764 0.11888383 0.1296731 0.13178383
#>  [4,] 0.20739528 0.21622252 0.20639422 0.19742056 0.2183737 0.22755899
#>  [5,] 0.31326656 0.32922421 0.32630815 0.31212084 0.3407825 0.35569403
#>  [6,] 0.47558781 0.49598779 0.49875339 0.47706846 0.4988573 0.51806845
#>  [7,] 0.64850747 0.66524521 0.67357911 0.64429306 0.6591142 0.67274961
#>  [8,] 0.82584423 0.83413736 0.84167657 0.82184833 0.8292713 0.83610046
#>  [9,] 0.95906160 0.96101104 0.96278327 0.95772383 0.9594853 0.96110593
#> [10,] 1.00000000 1.00000000 1.00000000 1.00000000 1.0000000 1.00000000
#> [11,] 1.05000000 1.04761905 1.04545455 1.04347826 1.0416667 1.04000000
#>            [,27]      [,28]      [,29]      [,30]      [,31]      [,32]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04405265 0.04242107 0.04090603 0.03949548 0.03817896 0.03699336
#>  [3,] 0.13199778 0.12710897 0.12256937 0.11834284 0.11439808 0.12032871
#>  [4,] 0.23392193 0.22525815 0.21721322 0.21223193 0.20515753 0.21773536
#>  [5,] 0.36696135 0.36116979 0.34827087 0.34646846 0.33491951 0.35288638
#>  [6,] 0.53292325 0.52921026 0.51030989 0.51061803 0.49359743 0.50993300
#>  [7,] 0.68533617 0.68420354 0.65976770 0.66262023 0.65086963 0.66213190
#>  [8,] 0.84240429 0.84367962 0.81354821 0.81879607 0.81219763 0.81825577
#>  [9,] 0.96260185 0.96398697 0.92955886 0.93198787 0.93197440 0.93416877
#> [10,] 1.00000000 1.00000000 0.96428571 0.96551724 0.96666667 0.96774194
#> [11,] 1.03846154 1.03703704 1.03571429 1.03448276 1.03333333 1.03225806
#>            [,33]      [,34]      [,35]      [,36]      [,37]      [,38]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03583731 0.03475133 0.03510837 0.03410527 0.03315791 0.03226175
#>  [3,] 0.11656843 0.11903242 0.12626799 0.12266033 0.11925310 0.11628096
#>  [4,] 0.21093113 0.21869517 0.23235687 0.22571811 0.22027123 0.22091776
#>  [5,] 0.34185868 0.35381475 0.37282020 0.36216820 0.35910628 0.36234939
#>  [6,] 0.49399759 0.50933100 0.52376244 0.50879780 0.50783807 0.51341028
#>  [7,] 0.64144028 0.65230572 0.66253203 0.64360254 0.64507347 0.65328550
#>  [8,] 0.79268528 0.79896754 0.80488026 0.78188368 0.78568871 0.79148091
#>  [9,] 0.90497600 0.90785551 0.91056564 0.88454948 0.88775644 0.89079005
#> [10,] 0.93750000 0.93939394 0.94117647 0.91428571 0.91666667 0.91891892
#> [11,] 1.03125000 1.03030303 1.02941176 1.02857143 1.02777778 1.02702703
#>            [,39]     [,40]      [,41]      [,42]      [,43]      [,44]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03141275 0.0306073 0.02984211 0.02911426 0.03517787 0.03435978
#>  [3,] 0.11879408 0.1174895 0.11455226 0.11175831 0.12711393 0.12415779
#>  [4,] 0.22785422 0.2299851 0.22551860 0.22001814 0.23858914 0.23304055
#>  [5,] 0.37274091 0.3773886 0.37480417 0.36566260 0.38076587 0.37564064
#>  [6,] 0.52621528 0.5331595 0.53224804 0.52063321 0.53204670 0.53014007
#>  [7,] 0.66240957 0.6710657 0.67227374 0.66470016 0.67268349 0.67424301
#>  [8,] 0.79696825 0.8021742 0.80567166 0.80230099 0.80700811 0.81149629
#>  [9,] 0.89366400 0.8963906 0.89898080 0.90079081 0.90315294 0.90540519
#> [10,] 0.92105263 0.9230769 0.92500000 0.92682927 0.92857143 0.93023256
#> [11,] 1.02631579 1.0256410 1.02500000 1.02439024 1.02380952 1.02325581
#>            [,45]      [,46]      [,47]      [,48]      [,49]      [,50]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03357888 0.03283268 0.03794143 0.04234968 0.04146739 0.04062112
#>  [3,] 0.12133603 0.12444229 0.13757475 0.14930291 0.14619243 0.14320891
#>  [4,] 0.22774418 0.23696683 0.25355451 0.26943633 0.26382307 0.25843893
#>  [5,] 0.36918917 0.38320719 0.39661573 0.40945369 0.40411932 0.39587199
#>  [6,] 0.52528451 0.53583374 0.54592431 0.55558550 0.55184775 0.54058555
#>  [7,] 0.67121963 0.67852586 0.68551443 0.69220561 0.69026266 0.67617567
#>  [8,] 0.81046077 0.81467276 0.81870161 0.82255902 0.82254141 0.81127233
#>  [9,] 0.90734264 0.90940169 0.91137122 0.91325694 0.91506408 0.91065005
#> [10,] 0.93181818 0.93333333 0.93478261 0.93617021 0.93750000 0.93877551
#> [11,] 1.02272727 1.02222222 1.02173913 1.02127660 1.02083333 1.02040816
#>           [,51]      [,52]      [,53]      [,54]     [,55]      [,56]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.0398087 0.03902813 0.03827759 0.03755537 0.0368599 0.03618972
#>  [3,] 0.1423701 0.13957856 0.13689435 0.13431144 0.1318242 0.13441430
#>  [4,] 0.2602694 0.25516607 0.25025903 0.24553716 0.2409902 0.24686306
#>  [5,] 0.3999276 0.39208591 0.38454579 0.37729021 0.3703034 0.37909270
#>  [6,] 0.5467208 0.53600075 0.52569304 0.51820937 0.5086129 0.51754721
#>  [7,] 0.6826522 0.66926682 0.66293092 0.65735184 0.6451787 0.65162996
#>  [8,] 0.8150469 0.79906557 0.79817346 0.79453662 0.7798230 0.78382620
#>  [9,] 0.9124370 0.90282397 0.90469274 0.90354010 0.8868079 0.88886591
#> [10,] 0.9400000 0.93839867 0.93958331 0.94072325 0.9351400 0.93631926
#> [11,] 1.0200000 1.01960784 1.01923077 1.01886792 1.0185185 1.01818182
#>            [,57]      [,58]      [,59]      [,60]      [,61]      [,62]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03554348 0.03491991 0.03431784 0.03373618 0.03886946 0.03823226
#>  [3,] 0.13201405 0.12969801 0.12746184 0.12530147 0.13460421 0.13239758
#>  [4,] 0.24287742 0.23861641 0.23450234 0.23052772 0.24335226 0.23936288
#>  [5,] 0.37671433 0.37010531 0.36372418 0.35755937 0.36826671 0.36222955
#>  [6,] 0.51666495 0.51232938 0.50561838 0.49704858 0.50543110 0.49714535
#>  [7,] 0.65232187 0.65170058 0.64776857 0.63678944 0.64284295 0.63230454
#>  [8,] 0.78612598 0.78925146 0.78812983 0.77882193 0.78250823 0.76968023
#>  [9,] 0.89085045 0.89276535 0.89461423 0.88986999 0.89170549 0.88011804
#> [10,] 0.93745642 0.93855367 0.93961309 0.94047464 0.94146673 0.93878956
#> [11,] 1.01785714 1.01754386 1.01724138 1.01694915 1.01666667 1.01639344
#>            [,63]      [,64]      [,65]      [,66]      [,67]      [,68]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04669155 0.04915126 0.04838327 0.04763891 0.04691711 0.04621686
#>  [3,] 0.14639117 0.15381313 0.15140980 0.14908042 0.14682163 0.14559460
#>  [4,] 0.25163122 0.26351009 0.25939274 0.25540209 0.25153236 0.25236223
#>  [5,] 0.37251617 0.38247623 0.37650004 0.37070773 0.36509095 0.36784566
#>  [6,] 0.50525591 0.51310899 0.50509166 0.50145805 0.49811366 0.50250271
#>  [7,] 0.63823511 0.64397741 0.63391526 0.63280114 0.63189330 0.63738743
#>  [8,] 0.77339506 0.77699196 0.76485146 0.76622432 0.76772153 0.77118837
#>  [9,] 0.88205162 0.88392382 0.87011251 0.87211078 0.87404849 0.87592837
#> [10,] 0.93977682 0.94073274 0.92603380 0.92717174 0.92827520 0.92934572
#> [11,] 1.01612903 1.01587302 1.01562500 1.01538462 1.01515152 1.01492537
#>           [,69]      [,70]      [,71]      [,72]     [,73]      [,74]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.0455372 0.04487724 0.04701435 0.04635218 0.0457084 0.04508225
#>  [3,] 0.1434535 0.14137446 0.14795191 0.14586808 0.1438421 0.14187169
#>  [4,] 0.2486510 0.24504738 0.25583242 0.25222915 0.2487260 0.24531876
#>  [5,] 0.3624362 0.35826969 0.36743726 0.36226209 0.3572307 0.35233710
#>  [6,] 0.4963175 0.49347813 0.50071415 0.49366184 0.4868054 0.48013686
#>  [7,] 0.6336836 0.63212089 0.63737631 0.62839918 0.6196714 0.61118276
#>  [8,] 0.7699819 0.76971129 0.77300112 0.76211378 0.7515289 0.74123395
#>  [9,] 0.8776466 0.87908308 0.88081046 0.86840468 0.8563435 0.84461277
#> [10,] 0.9303847 0.93139367 0.93237376 0.92795810 0.9230668 0.91042209
#> [11,] 1.0147059 1.01449275 1.01428571 1.01408451 1.0138889 1.01369863
#>            [,75]      [,76]      [,77]      [,78]      [,79]     [,80]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.05174649 0.05105654 0.05038474 0.04973039 0.04909282 0.0484714
#>  [3,] 0.15346802 0.15142178 0.14942939 0.14748875 0.14559787 0.1437549
#>  [4,] 0.25551715 0.25211026 0.24879302 0.24626611 0.24310886 0.2400315
#>  [5,] 0.36108930 0.35627478 0.35376131 0.35276405 0.34824144 0.3438333
#>  [6,] 0.48716204 0.48429713 0.48302725 0.48324409 0.47704865 0.4748852
#>  [7,] 0.61643705 0.61574797 0.61567657 0.61706357 0.60915250 0.6097020
#>  [8,] 0.74473079 0.74623063 0.74737045 0.74994001 0.74032540 0.7435996
#>  [9,] 0.84671260 0.84875643 0.85074648 0.85268483 0.84540027 0.8473572
#> [10,] 0.91163260 0.91281083 0.91395806 0.91507548 0.91322677 0.9143252
#> [11,] 1.01351351 1.01333333 1.01315789 1.01298701 1.01282051 1.0126582
#>           [,81]      [,82]      [,83]      [,84]      [,85]      [,86]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0478655 0.05042974 0.04981475 0.04921457 0.04862868 0.04805658
#>  [3,] 0.1419579 0.14893395 0.15216128 0.15032801 0.14853839 0.14679088
#>  [4,] 0.2381606 0.24756603 0.25481973 0.25309610 0.25008306 0.24714090
#>  [5,] 0.3434629 0.35156833 0.35947604 0.35919935 0.35754052 0.35333416
#>  [6,] 0.4756748 0.48214791 0.48846318 0.48934029 0.48877596 0.48302566
#>  [7,] 0.6116044 0.61639941 0.62107747 0.62306468 0.62355220 0.61621629
#>  [8,] 0.7466264 0.74975444 0.75280622 0.75578446 0.75733578 0.74842595
#>  [9,] 0.8492653 0.85112618 0.85294172 0.85471350 0.85644310 0.85107628
#> [10,] 0.9153961 0.91644059 0.91745961 0.91845407 0.91942485 0.91938571
#> [11,] 1.0125000 1.01234568 1.01219512 1.01204819 1.01190476 1.01176471
#>            [,87]      [,88]      [,89]      [,90]      [,91]      [,92]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04749778 0.04695183 0.04671171 0.04618686 0.04567367 0.04526456
#>  [3,] 0.14508401 0.14341638 0.14557865 0.14394294 0.14397055 0.14577695
#>  [4,] 0.24426717 0.24145950 0.24600623 0.24324212 0.24502793 0.24901952
#>  [5,] 0.34922562 0.34521153 0.35207784 0.34899537 0.35246768 0.35857432
#>  [6,] 0.47740908 0.47192162 0.47792251 0.47668449 0.48159957 0.48729628
#>  [7,] 0.60905098 0.60205040 0.60657255 0.60714747 0.61151250 0.61578159
#>  [8,] 0.73972332 0.73364822 0.73667495 0.73904648 0.74194596 0.74478172
#>  [9,] 0.84118004 0.83832785 0.84016503 0.84196093 0.84371692 0.84543431
#> [10,] 0.90869518 0.90945602 0.91048493 0.91149072 0.91247416 0.91343598
#> [11,] 1.01162791 1.01149425 1.01136364 1.01123596 1.01111111 1.01098901
#>            [,93]      [,94]      [,95]      [,96]      [,97]      [,98]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04477256 0.04536427 0.05551997 0.05587672 0.06043602 0.05981297
#>  [3,] 0.14427912 0.14753028 0.15659911 0.15948017 0.16810161 0.16820693
#>  [4,] 0.24895115 0.25480622 0.26273381 0.26808597 0.27571007 0.27743389
#>  [5,] 0.35986679 0.36674994 0.37348664 0.38008152 0.38653900 0.38984812
#>  [6,] 0.48974125 0.49522791 0.50059782 0.50585469 0.51100203 0.51575590
#>  [7,] 0.61938165 0.62347432 0.62747991 0.63140118 0.63524075 0.63900115
#>  [8,] 0.74755583 0.75027029 0.75292699 0.75552775 0.75807434 0.76056842
#>  [9,] 0.84711438 0.84875831 0.85036726 0.85194234 0.85348461 0.85499508
#> [10,] 0.91437689 0.91529757 0.91619866 0.91708078 0.91794452 0.91879045
#> [11,] 1.01086957 1.01075269 1.01063830 1.01052632 1.01041667 1.01030928
#>           [,99]     [,100]     [,101]     [,102]     [,103]     [,104]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0607539 0.06014023 0.05953883 0.06383771 0.06321185 0.06259815
#>  [3,] 0.1719079 0.17017143 0.16846972 0.17670269 0.17497031 0.17532367
#>  [4,] 0.2838864 0.28248924 0.28082266 0.28794323 0.28624441 0.28816423
#>  [5,] 0.3960742 0.39583265 0.39528212 0.40126942 0.40066438 0.40412011
#>  [6,] 0.5206972 0.52148565 0.52192806 0.52666145 0.52703188 0.53162381
#>  [7,] 0.6426848 0.64452992 0.64599137 0.64949640 0.65086738 0.65425702
#>  [8,] 0.7630116 0.76540542 0.76775137 0.77005086 0.77230526 0.77451589
#>  [9,] 0.8564747 0.85792447 0.85934523 0.86073785 0.86210316 0.86344197
#> [10,] 0.9196191 0.92043105 0.92122674 0.92200667 0.92277131 0.92352111
#> [11,] 1.0102041 1.01010101 1.01000000 1.00990099 1.00980392 1.00970874
#>           [,105]    [,106]     [,107]    [,108]    [,109]     [,110]     [,111]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.0000000 0.0000000 0.00000000 0.00000000
#>  [2,] 0.06199624 0.0614058 0.06642469 0.0658039 0.0651946 0.06743914 0.07591696
#>  [3,] 0.17363787 0.1719842 0.17979564 0.1787358 0.1770809 0.18114157 0.18858574
#>  [4,] 0.28539342 0.2826754 0.28944260 0.2896272 0.2880233 0.29390889 0.30032790
#>  [5,] 0.40159025 0.3977656 0.40344704 0.4048354 0.4042479 0.40971348 0.41507972
#>  [6,] 0.53065261 0.5255988 0.53007426 0.5325483 0.5328614 0.53714709 0.54135484
#>  [7,] 0.65489132 0.6486543 0.65196884 0.6552215 0.6564818 0.65963333 0.66272758
#>  [8,] 0.77668401 0.7692870 0.77146355 0.7735994 0.7756957 0.77775355 0.77977397
#>  [9,] 0.86475503 0.8601428 0.86146220 0.8627569 0.8640277 0.86527517 0.86649994
#> [10,] 0.92425648 0.9233714 0.92409429 0.9248037 0.9255000 0.92618344 0.92685450
#> [11,] 1.00961538 1.0095238 1.00943396 1.0093458 1.0092593 1.00917431 1.00909091
#>           [,112]    [,113]     [,114]     [,115]     [,116]     [,117]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07523303 0.0745613 0.07390147 0.07325321 0.07262281 0.07199675
#>  [3,] 0.18688676 0.1852181 0.18357903 0.18215221 0.18315340 0.18157450
#>  [4,] 0.29799429 0.2958144 0.29442943 0.29410342 0.29670968 0.29415183
#>  [5,] 0.41436819 0.4131380 0.41271668 0.41342626 0.41757347 0.41397370
#>  [6,] 0.54216158 0.5417792 0.54222131 0.54386808 0.54783444 0.54311173
#>  [7,] 0.66509674 0.6656055 0.66695370 0.66957952 0.67245274 0.66684682
#>  [8,] 0.78175799 0.7832139 0.78513233 0.78701714 0.78886916 0.78526763
#>  [9,] 0.86770264 0.8688839 0.87004419 0.87118415 0.87230429 0.87099146
#> [10,] 0.92751347 0.9281607 0.92879642 0.92942101 0.93003474 0.93063789
#> [11,] 1.00900901 1.0089286 1.00884956 1.00877193 1.00869565 1.00862069
#>          [,118]     [,119]     [,120]     [,121]     [,122]     [,123]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0713814 0.07077647 0.07018171 0.06959686 0.06902168 0.06845593
#>  [3,] 0.1800226 0.17849696 0.17699699 0.17552201 0.17407142 0.17264460
#>  [4,] 0.2916377 0.28916621 0.28673624 0.28434677 0.28199680 0.27968535
#>  [5,] 0.4104355 0.40695720 0.40353739 0.40017458 0.39686735 0.39361434
#>  [6,] 0.5384698 0.53393617 0.52944931 0.52503724 0.52069809 0.51643007
#>  [7,] 0.6611473 0.65820417 0.65267304 0.64723410 0.64188506 0.63664563
#>  [8,] 0.7785559 0.77724797 0.77071647 0.76429383 0.75797736 0.75509426
#>  [9,] 0.8635471 0.86414897 0.85688722 0.85183015 0.84479023 0.84450348
#> [10,] 0.9226837 0.92333894 0.91944746 0.91803288 0.91044583 0.91117988
#> [11,] 1.0085470 1.00847458 1.00840336 1.00833333 1.00826446 1.00819672
#>           [,124]     [,125]     [,126]
#>  [1,] 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06789938 0.06976671 0.07720858
#>  [3,] 0.17124099 0.17468983 0.18129231
#>  [4,] 0.27741149 0.28241903 0.28815967
#>  [5,] 0.39041422 0.39533024 0.40016759
#>  [6,] 0.51223145 0.51616507 0.52003575
#>  [7,] 0.63167692 0.63464727 0.63757009
#>  [8,] 0.75199509 0.75399513 0.75596317
#>  [9,] 0.84350992 0.84477194 0.84601376
#> [10,] 0.91190200 0.91261246 0.91331156
#> [11,] 1.00813008 1.00806452 1.00800000
#>