Non-randomized Probability Integral Transform

Functions to produce the non-randomized probability integral transform (PIT) to check the adequacy of the distributional assumption of the COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

compPredProb(object)

compPIT(object, bins = 10)

Arguments

object
an object class "cmp", obtained from a call to glm.cmp.

bins
numeric; the number of bins shown in the PIT histogram or the PIT Q-Q plot.

Value

compPredprob returns a list with values:

upper

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

lower

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

compPIT returns a list with values:
conditionalPIT

the conditional probability integral transformation given the observed counts.

PIT

the probability integral transformation.

Details

These functions are used to obtain the predictive probabilities and the probability integral transform for a fitted COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

References

Czado, C., Gneiting, T. and Held, L. (2009). Predictive model assessment for count data. Biometrics, 65, 1254--1261.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Series Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght 
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
compPredProb(M.bids)

#> $lower
#>   [1] 0.167217581 0.000000000 0.055885961 0.232781936 0.227539216 0.668012652
#>   [7] 0.423340098 0.282348026 0.402595909 0.000000000 0.306708026 0.383373685
#>  [13] 0.289153449 0.165111576 0.455576529 0.348671053 0.605151216 0.401425570
#>  [19] 0.525971226 0.335381114 0.962118350 0.146745397 0.309509001 0.616236955
#>  [25] 0.083174412 0.131698288 0.146679200 0.305541744 0.906413568 0.267792239
#>  [31] 0.500334858 0.099887950 0.905221321 0.126904782 0.085614956 0.996675900
#>  [37] 0.287050814 0.196458851 0.122738172 0.172481330 0.277335034 0.481779811
#>  [43] 0.040235482 0.344914325 0.359472622 0.131974553 0.042114222 0.045005961
#>  [49] 0.331501537 0.636783191 0.159694010 0.703054642 0.517776159 0.446061837
#>  [55] 0.806440239 0.105713030 0.289728027 0.422598703 0.459219303 0.636341829
#>  [61] 0.000000000 0.768603026 0.000000000 0.051368570 0.997747917 0.408311095
#>  [67] 0.404140696 0.173760444 0.473160206 0.367068316 0.052536826 0.830197917
#>  [73] 0.820970946 0.981802752 0.021774877 0.407121409 0.326092606 0.276042716
#>  [79] 0.741348841 0.411838591 0.260021131 0.043642330 0.103927062 0.250676309
#>  [85] 0.301374168 0.722211829 0.942781690 0.644623657 0.091965709 0.373459237
#>  [91] 0.143546001 0.097550592 0.197030370 0.071551419 0.000000000 0.074104871
#>  [97] 0.000000000 0.133007722 0.060182936 0.236163975 0.248906370 0.037321198
#> [103] 0.249429460 0.122859045 0.351602517 0.715465751 0.019538443 0.173064487
#> [109] 0.248652647 0.000000000 0.000000000 0.286369030 0.276211615 0.238817281
#> [115] 0.191566955 0.100109237 0.593639637 0.994871520 0.498969760 0.849514997
#> [121] 0.749781875 0.942134161 0.599287612 0.592669365 0.000000000 0.002032585
#> 
#> $upper
#>   [1] 0.45701496 0.19977464 0.30974762 0.67260207 0.66570964 0.88790690
#>   [7] 0.76255003 0.73131472 0.83684941 0.24226559 0.75644059 0.72715677
#>  [13] 0.62844248 0.57094183 0.78869798 0.79492832 0.85059070 0.74357021
#>  [19] 0.83914227 0.78334586 0.99576735 0.53761188 0.65181067 0.83304418
#>  [25] 0.39316966 0.50794643 0.53748627 0.75528726 0.97897693 0.71519774
#>  [31] 0.82176964 0.43660448 0.97859953 0.49799577 0.39982063 0.99926678
#>  [37] 0.73633990 0.62169154 0.48913434 0.58352333 0.72586173 0.80849569
#>  [43] 0.25183558 0.68966302 0.80395713 0.39384308 0.25930718 0.27050366
#>  [49] 0.77986237 0.87001695 0.56141367 0.93225438 0.75967147 0.86546851
#>  [55] 0.95245310 0.45067875 0.73916234 0.71008824 0.79152625 0.90229425
#>  [61] 0.29299291 0.93716009 0.17812309 0.29396481 0.99973622 0.74964292
#>  [67] 0.74597698 0.58566400 0.80212191 0.81010806 0.29811197 0.94259156
#>  [73] 0.95785668 0.99646271 0.16764526 0.74860093 0.77492598 0.72443957
#>  [79] 0.94691559 0.70010952 0.70622838 0.26526827 0.33708448 0.69508765
#>  [85] 0.75112640 0.91600772 0.99231959 0.90633070 0.41661524 0.71782892
#>  [91] 0.53149347 0.43081207 0.62255351 0.35983206 0.05753223 0.36740614
#>  [97] 0.20301143 0.51062076 0.32416055 0.67697306 0.69293243 0.16490694
#> [103] 0.69357091 0.48939427 0.69644243 0.93720283 0.15565451 0.58450082
#> [109] 0.69262227 0.32310686 0.02512811 0.62514829 0.72462582 0.68036179
#> [115] 0.61422655 0.43714847 0.87998164 0.99913415 0.82081348 0.95870941
#> [121] 0.94990546 0.98923327 0.84684206 0.87944469 0.33430664 0.02932648
#> 

compPIT(M.bids)

#> $ConditionalPIT
#>            [,1]     [,2]      [,3]      [,4]      [,5]     [,6]      [,7]
#>  [1,] 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000 0.0000000
#>  [2,] 0.0000000 0.500564 0.1737720 0.0000000 0.0000000 0.000000 0.0000000
#>  [3,] 0.1131219 1.000000 0.5676873 0.0000000 0.0000000 0.000000 0.0000000
#>  [4,] 0.4581905 1.000000 0.9616026 0.1528308 0.1653712 0.000000 0.0000000
#>  [5,] 0.8032592 1.000000 1.0000000 0.3801965 0.3935929 0.000000 0.0000000
#>  [6,] 1.0000000 1.000000 1.0000000 0.6075622 0.6218146 0.000000 0.2259954
#>  [7,] 1.0000000 1.000000 1.0000000 0.8349278 0.8500363 0.000000 0.5207981
#>  [8,] 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 0.145467 0.8156008
#>  [9,] 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 0.600231 1.0000000
#> [10,] 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.000000 1.0000000
#> [11,] 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.000000 1.0000000
#>             [,8]      [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.00000000 0.0000000 0.4127701 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.00000000 0.0000000 0.8255403 0.0000000 0.00000000 0.00000000 0.08596802
#>  [4,] 0.03931689 0.0000000 1.0000000 0.0000000 0.00000000 0.03196847 0.33237646
#>  [5,] 0.26205056 0.0000000 1.0000000 0.2074388 0.04836281 0.32670243 0.57878490
#>  [6,] 0.48478423 0.2243024 1.0000000 0.4297931 0.33924390 0.62143639 0.82519334
#>  [7,] 0.70751791 0.4545826 1.0000000 0.6521475 0.63012500 0.91617035 1.00000000
#>  [8,] 0.93025158 0.6848629 1.0000000 0.8745019 0.92100610 1.00000000 1.00000000
#>  [9,] 1.00000000 0.9151431 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [10,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,15]     [,16]     [,17]     [,18]     [,19]     [,20] [,21]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [2,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [3,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [4,] 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>  [5,] 0.0000000 0.1150210 0.0000000 0.0000000 0.0000000 0.1442499     0
#>  [6,] 0.1333552 0.3391070 0.0000000 0.2881075 0.0000000 0.3674818     0
#>  [7,] 0.4335460 0.5631930 0.0000000 0.5803815 0.2363845 0.5907136     0
#>  [8,] 0.7337368 0.7872789 0.3864447 0.8726556 0.5556988 0.8139455     0
#>  [9,] 1.0000000 1.0000000 0.7938771 1.0000000 0.8750131 1.0000000     0
#> [10,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     0
#> [11,] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000     1
#>           [,22]     [,23]     [,24]      [,25]     [,26]     [,27]     [,28]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.05427692 0.0000000 0.0000000 0.0000000
#>  [3,] 0.1362476 0.0000000 0.0000000 0.37686251 0.1815337 0.1364377 0.0000000
#>  [4,] 0.3920894 0.0000000 0.0000000 0.69944810 0.4473157 0.3923184 0.0000000
#>  [5,] 0.6479312 0.2643604 0.0000000 1.00000000 0.7130978 0.6481991 0.2100260
#>  [6,] 0.9037731 0.5565004 0.0000000 1.00000000 0.9788798 0.9040798 0.4323740
#>  [7,] 1.0000000 0.8486403 0.0000000 1.00000000 1.0000000 1.0000000 0.6547219
#>  [8,] 1.0000000 1.0000000 0.3863480 1.00000000 1.0000000 1.0000000 0.8770699
#>  [9,] 1.0000000 1.0000000 0.8475873 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>       [,29]      [,30]     [,31]        [,32] [,33]     [,34]      [,35] [,36]
#>  [1,]     0 0.00000000 0.0000000 0.0000000000     0 0.0000000 0.00000000     0
#>  [2,]     0 0.00000000 0.0000000 0.0003327711     0 0.0000000 0.04578225     0
#>  [3,]     0 0.00000000 0.0000000 0.2973184882     0 0.1969738 0.36404513     0
#>  [4,]     0 0.07198785 0.0000000 0.5943042053     0 0.4664495 0.68230800     0
#>  [5,]     0 0.29549874 0.0000000 0.8912899224     0 0.7359252 1.00000000     0
#>  [6,]     0 0.51900963 0.0000000 1.0000000000     0 1.0000000 1.00000000     0
#>  [7,]     0 0.74252051 0.3100633 1.0000000000     0 1.0000000 1.00000000     0
#>  [8,]     0 0.96603140 0.6211684 1.0000000000     0 1.0000000 1.00000000     0
#>  [9,]     0 1.00000000 0.9322735 1.0000000000     0 1.0000000 1.00000000     0
#> [10,]     0 1.00000000 1.0000000 1.0000000000     0 1.0000000 1.00000000     0
#> [11,]     1 1.00000000 1.0000000 1.0000000000     1 1.0000000 1.00000000     1
#>           [,37]       [,38]     [,39]      [,40]      [,41]      [,42]
#>  [1,] 0.0000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0000000 0.000000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.0000000 0.008327556 0.2108696 0.06694856 0.00000000 0.00000000
#>  [4,] 0.0288215 0.243492918 0.4837983 0.31023270 0.05053203 0.00000000
#>  [5,] 0.2513953 0.478658279 0.7567269 0.55351684 0.27348421 0.00000000
#>  [6,] 0.4739692 0.713823641 1.0000000 0.79680099 0.49643638 0.05576769
#>  [7,] 0.6965430 0.948989002 1.0000000 1.00000000 0.71938855 0.36184403
#>  [8,] 0.9191169 1.000000000 1.0000000 1.00000000 0.94234072 0.66792037
#>  [9,] 1.0000000 1.000000000 1.0000000 1.00000000 1.00000000 0.97399671
#> [10,] 1.0000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.0000000 1.000000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,43]     [,44]      [,45]     [,46]     [,47]     [,48]     [,49]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.2824409 0.0000000 0.00000000 0.0000000 0.2665177 0.2438785 0.0000000
#>  [3,] 0.7550304 0.0000000 0.00000000 0.2597695 0.7269378 0.6873420 0.0000000
#>  [4,] 1.0000000 0.0000000 0.00000000 0.6416405 1.0000000 1.0000000 0.0000000
#>  [5,] 1.0000000 0.1597850 0.09117838 1.0000000 1.0000000 1.0000000 0.1527753
#>  [6,] 1.0000000 0.4498514 0.31615810 1.0000000 1.0000000 1.0000000 0.3758100
#>  [7,] 1.0000000 0.7399177 0.54113782 1.0000000 1.0000000 1.0000000 0.5988446
#>  [8,] 1.0000000 1.0000000 0.76611754 1.0000000 1.0000000 1.0000000 0.8218793
#>  [9,] 1.0000000 1.0000000 0.99109726 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,50]     [,51]     [,52]    [,53]     [,54]     [,55]     [,56]
#>  [1,] 0.0000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.0000000 0.0000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000
#>  [3,] 0.0000000 0.1003336 0.0000000 0.000000 0.0000000 0.0000000 0.2733227
#>  [4,] 0.0000000 0.3492634 0.0000000 0.000000 0.0000000 0.0000000 0.5632066
#>  [5,] 0.0000000 0.5981933 0.0000000 0.000000 0.0000000 0.0000000 0.8530905
#>  [6,] 0.0000000 0.8471231 0.0000000 0.000000 0.1286059 0.0000000 1.0000000
#>  [7,] 0.0000000 1.0000000 0.0000000 0.339915 0.3670379 0.0000000 1.0000000
#>  [8,] 0.2710448 1.0000000 0.0000000 0.753317 0.6054700 0.0000000 1.0000000
#>  [9,] 0.6997992 1.0000000 0.4229733 1.000000 0.8439021 0.0000000 1.0000000
#> [10,] 1.0000000 1.0000000 0.8592739 1.000000 1.0000000 0.6407639 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000 1.000000 1.0000000 1.0000000 1.0000000
#>            [,57]     [,58]     [,59]     [,60]     [,61]     [,62]     [,63]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.0000000 0.3413052 0.0000000 0.5614095
#>  [3,] 0.00000000 0.0000000 0.0000000 0.0000000 0.6826104 0.0000000 1.0000000
#>  [4,] 0.02285534 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [5,] 0.24535726 0.0000000 0.0000000 0.0000000 1.0000000 0.0000000 1.0000000
#>  [6,] 0.46785919 0.2692317 0.1227200 0.0000000 1.0000000 0.0000000 1.0000000
#>  [7,] 0.69036111 0.6170705 0.4236466 0.0000000 1.0000000 0.0000000 1.0000000
#>  [8,] 0.91286303 0.9649092 0.7245732 0.2393592 1.0000000 0.0000000 1.0000000
#>  [9,] 1.00000000 1.0000000 1.0000000 0.6153664 1.0000000 0.1862691 1.0000000
#> [10,] 1.00000000 1.0000000 1.0000000 0.9913735 1.0000000 0.7795400 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>           [,64] [,65]     [,66]     [,67]      [,68]      [,69]      [,70]
#>  [1,] 0.0000000     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.2004624     0 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [3,] 0.6126700     0 0.0000000 0.0000000 0.06370316 0.00000000 0.00000000
#>  [4,] 1.0000000     0 0.0000000 0.0000000 0.30647843 0.00000000 0.00000000
#>  [5,] 1.0000000     0 0.0000000 0.0000000 0.54925371 0.00000000 0.07433122
#>  [6,] 1.0000000     0 0.2686210 0.2804246 0.79202899 0.08158942 0.30004460
#>  [7,] 1.0000000     0 0.5615911 0.5729623 1.00000000 0.38557617 0.52575799
#>  [8,] 1.0000000     0 0.8545611 0.8655000 1.00000000 0.68956292 0.75147137
#>  [9,] 1.0000000     0 1.0000000 1.0000000 1.00000000 0.99354967 0.97718476
#> [10,] 1.0000000     0 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.0000000     1 1.0000000 1.0000000 1.00000000 1.00000000 1.00000000
#>           [,71]     [,72]     [,73] [,74]     [,75]     [,76]     [,77]
#>  [1,] 0.0000000 0.0000000 0.0000000     0 0.0000000 0.0000000 0.0000000
#>  [2,] 0.1932735 0.0000000 0.0000000     0 0.5362646 0.0000000 0.0000000
#>  [3,] 0.6004809 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [4,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.0000000
#>  [5,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.0000000 0.1646656
#>  [6,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.2719888 0.3874654
#>  [7,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.5648321 0.6102652
#>  [8,] 1.0000000 0.0000000 0.0000000     0 1.0000000 0.8576754 0.8330650
#>  [9,] 1.0000000 0.0000000 0.0000000     0 1.0000000 1.0000000 1.0000000
#> [10,] 1.0000000 0.6210501 0.5773359     0 1.0000000 1.0000000 1.0000000
#> [11,] 1.0000000 1.0000000 1.0000000     1 1.0000000 1.0000000 1.0000000
#>            [,78]     [,79]     [,80]      [,81]     [,82]     [,83]     [,84]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.00000000 0.0000000 0.0000000 0.00000000 0.2542919 0.0000000 0.0000000
#>  [3,] 0.00000000 0.0000000 0.0000000 0.00000000 0.7055026 0.4120518 0.0000000
#>  [4,] 0.05342875 0.0000000 0.0000000 0.08959708 1.0000000 0.8409466 0.1109866
#>  [5,] 0.27644548 0.0000000 0.0000000 0.31370819 1.0000000 1.0000000 0.3360033
#>  [6,] 0.49946221 0.0000000 0.3058283 0.53781930 1.0000000 1.0000000 0.5610201
#>  [7,] 0.72247894 0.0000000 0.6527242 0.76193040 1.0000000 1.0000000 0.7860368
#>  [8,] 0.94549567 0.0000000 0.9996201 0.98604151 1.0000000 1.0000000 1.0000000
#>  [9,] 1.00000000 0.2853144 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.00000000 0.7717744 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.00000000 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000 1.0000000
#>           [,85]     [,86] [,87]     [,88]      [,89]      [,90]     [,91]
#>  [1,] 0.0000000 0.0000000     0 0.0000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.0000000 0.0000000     0 0.0000000 0.02474758 0.00000000 0.0000000
#>  [3,] 0.0000000 0.0000000     0 0.0000000 0.33277205 0.00000000 0.1455197
#>  [4,] 0.0000000 0.0000000     0 0.0000000 0.64079653 0.00000000 0.4032866
#>  [5,] 0.2192893 0.0000000     0 0.0000000 0.94882100 0.07707056 0.6610534
#>  [6,] 0.4416339 0.0000000     0 0.0000000 1.00000000 0.36745616 0.9188203
#>  [7,] 0.6639785 0.0000000     0 0.0000000 1.00000000 0.65784177 1.0000000
#>  [8,] 0.8863232 0.0000000     0 0.2115967 1.00000000 0.94822738 1.0000000
#>  [9,] 1.0000000 0.4013923     0 0.5937033 1.00000000 1.00000000 1.0000000
#> [10,] 1.0000000 0.9173991     0 0.9758100 1.00000000 1.00000000 1.0000000
#> [11,] 1.0000000 1.0000000     1 1.0000000 1.00000000 1.00000000 1.0000000
#>             [,92]       [,93]      [,94] [,95]     [,96]     [,97]     [,98]
#>  [1,] 0.000000000 0.000000000 0.00000000     0 0.0000000 0.0000000 0.0000000
#>  [2,] 0.007349808 0.000000000 0.09868363     1 0.0882885 0.4925831 0.0000000
#>  [3,] 0.307414492 0.006978774 0.44556783     1 0.4292349 0.9851662 0.1774099
#>  [4,] 0.607479176 0.241983619 0.79245202     1 0.7701812 1.0000000 0.4422312
#>  [5,] 0.907543860 0.476988465 1.00000000     1 1.0000000 1.0000000 0.7070526
#>  [6,] 1.000000000 0.711993310 1.00000000     1 1.0000000 1.0000000 0.9718740
#>  [7,] 1.000000000 0.946998156 1.00000000     1 1.0000000 1.0000000 1.0000000
#>  [8,] 1.000000000 1.000000000 1.00000000     1 1.0000000 1.0000000 1.0000000
#>  [9,] 1.000000000 1.000000000 1.00000000     1 1.0000000 1.0000000 1.0000000
#> [10,] 1.000000000 1.000000000 1.00000000     1 1.0000000 1.0000000 1.0000000
#> [11,] 1.000000000 1.000000000 1.00000000     1 1.0000000 1.0000000 1.0000000
#>          [,99]    [,100]   [,101]    [,102]    [,103]    [,104]    [,105]
#>  [1,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,] 0.150835 0.0000000 0.000000 0.4912681 0.0000000 0.0000000 0.0000000
#>  [3,] 0.529655 0.0000000 0.000000 1.0000000 0.0000000 0.2104599 0.0000000
#>  [4,] 0.908475 0.1448156 0.115069 1.0000000 0.1138613 0.4832849 0.0000000
#>  [5,] 1.000000 0.3716712 0.340281 1.0000000 0.3390148 0.7561100 0.1403477
#>  [6,] 1.000000 0.5985267 0.565493 1.0000000 0.5641683 1.0000000 0.4303373
#>  [7,] 1.000000 0.8253823 0.790705 1.0000000 0.7893218 1.0000000 0.7203269
#>  [8,] 1.000000 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [9,] 1.000000 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [10,] 1.000000 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [11,] 1.000000 1.0000000 1.000000 1.0000000 1.0000000 1.0000000 1.0000000
#>          [,106]    [,107]     [,108]    [,109]    [,110] [,111]     [,112]
#>  [1,] 0.0000000 0.0000000 0.00000000 0.0000000 0.0000000      0 0.00000000
#>  [2,] 0.0000000 0.5911246 0.00000000 0.0000000 0.3094951      1 0.00000000
#>  [3,] 0.0000000 1.0000000 0.06546703 0.0000000 0.6189903      1 0.00000000
#>  [4,] 0.0000000 1.0000000 0.30851800 0.1156551 0.9284854      1 0.04023555
#>  [5,] 0.0000000 1.0000000 0.55156897 0.3408957 1.0000000      1 0.33541301
#>  [6,] 0.0000000 1.0000000 0.79461994 0.5661364 1.0000000      1 0.63059047
#>  [7,] 0.0000000 1.0000000 1.00000000 0.7913770 1.0000000      1 0.92576793
#>  [8,] 0.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>  [9,] 0.3812364 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [10,] 0.8322210 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#> [11,] 1.0000000 1.0000000 1.00000000 1.0000000 1.0000000      1 1.00000000
#>           [,113]    [,114]     [,115]    [,116]     [,117] [,118]      [,119]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000      0 0.000000000
#>  [2,] 0.00000000 0.0000000 0.00000000 0.0000000 0.00000000      0 0.000000000
#>  [3,] 0.00000000 0.0000000 0.01995233 0.2963773 0.00000000      0 0.000000000
#>  [4,] 0.05305003 0.1385652 0.25654935 0.5930786 0.00000000      0 0.000000000
#>  [5,] 0.27605812 0.3650430 0.49314637 0.8897800 0.00000000      0 0.000000000
#>  [6,] 0.49906622 0.5915207 0.72974340 1.0000000 0.00000000      0 0.003201056
#>  [7,] 0.72207432 0.8179984 0.96634042 1.0000000 0.02221247      0 0.313910863
#>  [8,] 0.94508242 1.0000000 1.00000000 1.0000000 0.37144520      0 0.624620670
#>  [9,] 1.00000000 1.0000000 1.00000000 1.0000000 0.72067794      0 0.935330478
#> [10,] 1.00000000 1.0000000 1.00000000 1.0000000 1.00000000      0 1.000000000
#> [11,] 1.00000000 1.0000000 1.00000000 1.0000000 1.00000000      1 1.000000000
#>          [,120]    [,121] [,122]      [,123]     [,124]    [,125] [,126]
#>  [1,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.0000000      0
#>  [2,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.2991266      1
#>  [3,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.5982531      1
#>  [4,] 0.0000000 0.0000000      0 0.000000000 0.00000000 0.8973797      1
#>  [5,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [6,] 0.0000000 0.0000000      0 0.000000000 0.00000000 1.0000000      1
#>  [7,] 0.0000000 0.0000000      0 0.002877701 0.02556229 1.0000000      1
#>  [8,] 0.0000000 0.0000000      0 0.406829234 0.37426733 1.0000000      1
#>  [9,] 0.0000000 0.2509356      0 0.810780767 0.72297237 1.0000000      1
#> [10,] 0.4623405 0.7506268      0 1.000000000 1.00000000 1.0000000      1
#> [11,] 1.0000000 1.0000000      1 1.000000000 1.00000000 1.0000000      1
#> 
#> $PIT
#>       [,1]     [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,]    0 0.000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>  [2,]    0 0.500564 0.3371680 0.2247787 0.1685840 0.1348672 0.1123893
#>  [3,]    0 1.113122 0.8404046 0.5602697 0.4202023 0.3361618 0.2801349
#>  [4,]    0 1.458191 1.2098966 0.8575413 0.6844988 0.5475990 0.4563325
#>  [5,]    0 1.803259 1.4016296 1.0611519 0.8942622 0.7154097 0.5961748
#>  [6,]    0 2.000000 1.5000000 1.2025207 1.0573442 0.8458754 0.7425620
#>  [7,]    0 2.000000 1.5000000 1.2783093 1.1712410 0.9369928 0.8676270
#>  [8,]    0 2.000000 1.5000000 1.3333333 1.2500000 1.0290934 0.9935113
#>  [9,]    0 2.000000 1.5000000 1.3333333 1.2500000 1.1200462 1.1000385
#> [10,]    0 2.000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#> [11,]    0 2.000000 1.5000000 1.3333333 1.2500000 1.2000000 1.1666667
#>             [,8]      [,9]     [,10]     [,11]      [,12]      [,13]      [,14]
#>  [1,] 0.00000000 0.0000000 0.0000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.09633371 0.0842920 0.1207896 0.1087106 0.09882783 0.09059218 0.08362355
#>  [3,] 0.24011559 0.2101011 0.2784833 0.2506349 0.22784995 0.20886245 0.19940903
#>  [4,] 0.39675887 0.3471640 0.4197013 0.3777312 0.34339201 0.31744005 0.31858900
#>  [5,] 0.54844274 0.4798874 0.5376777 0.5046538 0.46317280 0.45180027 0.46156832
#>  [6,] 0.70573664 0.6455574 0.6849399 0.6594252 0.63031781 0.62957769 0.64462505
#>  [7,] 0.84475431 0.7959829 0.8186514 0.8020010 0.78637594 0.79719214 0.81279275
#>  [8,] 0.98447419 0.9470228 0.9529091 0.9450684 0.94288093 0.94764085 0.95166848
#>  [9,] 1.08574729 1.0644218 1.0572638 1.0515374 1.04685219 1.04294784 1.03964416
#> [10,] 1.14285714 1.1250000 1.1111111 1.1000000 1.09090909 1.08333333 1.07692308
#> [11,] 1.14285714 1.1250000 1.1111111 1.1000000 1.09090909 1.08333333 1.07692308
#>            [,15]      [,16]      [,17]      [,18]      [,19]      [,20]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07765044 0.07247374 0.06794413 0.06394742 0.06039479 0.05721611
#>  [3,] 0.18516553 0.17282116 0.16201984 0.15248926 0.14401764 0.13643776
#>  [4,] 0.29583265 0.27611047 0.25885356 0.24362688 0.23009206 0.21798195
#>  [5,] 0.42859915 0.40769394 0.38221307 0.35972995 0.33974495 0.32945574
#>  [6,] 0.60810577 0.59017252 0.55328674 0.53768796 0.50781640 0.50043037
#>  [7,] 0.78570369 0.77086964 0.72269029 0.71431919 0.68776726 0.68265917
#>  [8,] 0.93610193 0.92618040 0.89244691 0.89128272 0.87263917 0.86955003
#>  [9,] 1.03681244 1.03435827 1.01932820 1.01819125 1.01023691 1.00969812
#> [10,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#> [11,] 1.07142857 1.06666667 1.06250000 1.05882353 1.05555556 1.05263158
#>            [,21]      [,22]      [,23]      [,24]      [,25]      [,26]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.05435531 0.05176696 0.04941392 0.04726548 0.04755763 0.04565532
#>  [3,] 0.12961587 0.12993167 0.12402568 0.11863326 0.12939281 0.13147845
#>  [4,] 0.20708285 0.21589269 0.20607938 0.19711941 0.21804977 0.22722041
#>  [5,] 0.31298295 0.32893287 0.32599776 0.31182394 0.34049794 0.35540194
#>  [6,] 0.47540885 0.49580715 0.49856593 0.47688915 0.49868544 0.51789321
#>  [7,] 0.64852621 0.66526306 0.67359839 0.64431151 0.65913186 0.67276658
#>  [8,] 0.82607253 0.83435479 0.84188411 0.82207820 0.82949161 0.83631194
#>  [9,] 0.95921322 0.96115544 0.96292111 0.95790659 0.95966048 0.96127406
#> [10,] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [11,] 1.05000000 1.04761905 1.04545455 1.04347826 1.04166667 1.04000000
#>            [,27]      [,28]      [,29]      [,30]     [,31]      [,32]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.00000000
#>  [2,] 0.04389935 0.04227345 0.04076368 0.03935804 0.0380461 0.03682954
#>  [3,] 0.13166919 0.12679255 0.12226424 0.11804824 0.1141133 0.12002314
#>  [4,] 0.23357033 0.22491958 0.21688674 0.21189022 0.2048272 0.21739099
#>  [5,] 0.36666337 0.36086198 0.34797406 0.34616456 0.3346257 0.35258265
#>  [6,] 0.53274654 0.52902904 0.51013515 0.51044116 0.4934265 0.50976754
#>  [7,] 0.68535249 0.68421802 0.65978166 0.66263473 0.6508823 0.66214421
#>  [8,] 0.84260764 0.84388402 0.81374530 0.81899655 0.8124023 0.81845382
#>  [9,] 0.96276352 0.96414265 0.92970899 0.93213281 0.9321375 0.93432662
#> [10,] 1.00000000 1.00000000 0.96428571 0.96551724 0.9666667 0.96774194
#> [11,] 1.03846154 1.03703704 1.03571429 1.03448276 1.0333333 1.03225806
#>            [,33]      [,34]      [,35]      [,36]      [,37]      [,38]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03567862 0.03459745 0.03492641 0.03392852 0.03298606 0.03209454
#>  [3,] 0.11627242 0.11871791 0.12593342 0.12233532 0.11893712 0.11594767
#>  [4,] 0.21059752 0.21835061 0.23199642 0.22536795 0.21990833 0.22054575
#>  [5,] 0.34156444 0.35351477 0.37252904 0.36188536 0.35881619 0.36205516
#>  [6,] 0.49383730 0.50917557 0.52361158 0.50865125 0.50768786 0.51325910
#>  [7,] 0.64145220 0.65231729 0.66254325 0.64361344 0.64508371 0.65329736
#>  [8,] 0.79287713 0.79915358 0.80506083 0.78205909 0.78586625 0.79165365
#>  [9,] 0.90512891 0.90800379 0.91070956 0.88468929 0.88789237 0.89092230
#> [10,] 0.93750000 0.93939394 0.94117647 0.91428571 0.91666667 0.91891892
#> [11,] 1.03125000 1.03030303 1.02941176 1.02857143 1.02777778 1.02702703
#>            [,39]      [,40]      [,41]      [,42]      [,43]      [,44]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03124995 0.03044867 0.02968745 0.02896337 0.03499855 0.03418463
#>  [3,] 0.11844562 0.11712518 0.11419705 0.11141176 0.12673601 0.12378866
#>  [4,] 0.22747345 0.22959548 0.22511889 0.21962819 0.23820847 0.23266874
#>  [5,] 0.37244126 0.37708423 0.37449423 0.36536022 0.38047069 0.37533847
#>  [6,] 0.52606807 0.53300994 0.53209560 0.52047784 0.53189504 0.52998705
#>  [7,] 0.66242112 0.67107699 0.67228478 0.66471305 0.67269607 0.67425937
#>  [8,] 0.79713645 0.80233808 0.80583815 0.80247430 0.80717729 0.81166154
#>  [9,] 0.89379277 0.89651603 0.89910313 0.90092980 0.90328862 0.90553772
#> [10,] 0.92105263 0.92307692 0.92500000 0.92682927 0.92857143 0.93023256
#> [11,] 1.02631579 1.02564103 1.02500000 1.02439024 1.02380952 1.02325581
#>           [,45]      [,46]      [,47]      [,48]      [,49]     [,50]     [,51]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.0000000 0.0000000
#>  [2,] 0.0334077 0.03266531 0.03774906 0.04213479 0.04125698 0.0404150 0.0396067
#>  [3,] 0.1209753 0.12405960 0.13716565 0.14887153 0.14577004 0.1427951 0.1419459
#>  [4,] 0.2273808 0.23658658 0.25318252 0.26907226 0.26346659 0.2580897 0.2599132
#>  [5,] 0.3688803 0.38290516 0.39632027 0.40916452 0.40382308 0.3955818 0.3996340
#>  [6,] 0.5251273 0.53568002 0.54577394 0.55543832 0.55169606 0.5404370 0.5465707
#>  [7,] 0.6712339 0.67853979 0.68552806 0.69221895 0.69027365 0.6761864 0.6826627
#>  [8,] 0.8106265 0.81483475 0.81886008 0.82271412 0.82269673 0.8114385 0.8152098
#>  [9,] 0.9074823 0.90953820 0.91150476 0.91338764 0.91519207 0.9107963 0.9125804
#> [10,] 0.9318182 0.93333333 0.93478261 0.93617021 0.93750000 0.9387755 0.9400000
#> [11,] 1.0227273 1.02222222 1.02173913 1.02127660 1.02083333 1.0204082 1.0200000
#>           [,52]      [,53]      [,54]      [,55]      [,56]      [,57]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0388301 0.03808337 0.03736482 0.03667287 0.03600609 0.03536313
#>  [3,] 0.1391627 0.13648645 0.13391123 0.13143140 0.13401124 0.13161818
#>  [4,] 0.2548169 0.24991653 0.24520112 0.24066036 0.24652484 0.24253074
#>  [5,] 0.3917981 0.38426348 0.37701323 0.37003150 0.37881439 0.37643123
#>  [6,] 0.5358536 0.52554873 0.51805924 0.50846556 0.51740255 0.51651784
#>  [7,] 0.6692772 0.66294327 0.65736015 0.64518682 0.65163797 0.65232945
#>  [8,] 0.7992253 0.79834240 0.79470330 0.77998657 0.78398682 0.78628818
#>  [9,] 0.9029802 0.90484599 0.90369611 0.88696099 0.88901625 0.89099810
#> [10,] 0.9384171 0.93960142 0.94074102 0.93518589 0.93636432 0.93750067
#> [11,] 1.0196078 1.01923077 1.01886792 1.01851852 1.01818182 1.01785714
#>            [,58]      [,59]     [,60]      [,61]      [,62]      [,63]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.03474272 0.03414371 0.0335650 0.03869401 0.03805968 0.04650081
#>  [3,] 0.12930909 0.12707962 0.1249257 0.13422048 0.13202014 0.14601982
#>  [4,] 0.23827582 0.23416761 0.2301987 0.24302869 0.23904461 0.25131809
#>  [5,] 0.36982717 0.36345084 0.3572907 0.36800248 0.36196965 0.37226046
#>  [6,] 0.51217949 0.50546467 0.4968975 0.50528251 0.49699919 0.50511211
#>  [7,] 0.65171087 0.64777873 0.6367994 0.64285277 0.63231420 0.63824462
#>  [8,] 0.78942188 0.78830380 0.7789997 0.78268299 0.76985212 0.77356419
#>  [9,] 0.89291042 0.89475679 0.8900214 0.89185433 0.88028736 0.88221821
#> [10,] 0.93859715 0.93965582 0.9405324 0.94152352 0.93886805 0.93985405
#> [11,] 1.01754386 1.01724138 1.0169492 1.01666667 1.01639344 1.01612903
#>            [,64]      [,65]      [,66]      [,67]      [,68]      [,69]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04894464 0.04817988 0.04743865 0.04671988 0.04602257 0.04534577
#>  [3,] 0.15342696 0.15102966 0.14870613 0.14645301 0.14521794 0.14308238
#>  [4,] 0.26320193 0.25908940 0.25510341 0.25123820 0.25206269 0.24835588
#>  [5,] 0.38222458 0.37625232 0.37046383 0.36485074 0.36760302 0.36219709
#>  [6,] 0.51296747 0.50495236 0.50131649 0.49796964 0.50235859 0.49617081
#>  [7,] 0.64398677 0.63392447 0.63281165 0.63190484 0.63739880 0.63369553
#>  [8,] 0.77715840 0.76501530 0.76639293 0.76789455 0.77135882 0.77015593
#>  [9,] 0.88408776 0.87027389 0.87226968 0.87420499 0.87608252 0.87780998
#> [10,] 0.94080875 0.92610861 0.92724540 0.92834775 0.92941718 0.93045517
#> [11,] 1.01587302 1.01562500 1.01538462 1.01515152 1.01492537 1.01470588
#>            [,70]      [,71]      [,72]      [,73]      [,74]      [,75]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04468858 0.04681123 0.04615191 0.04551091 0.04488748 0.05152771
#>  [3,] 0.14100872 0.14757261 0.14549412 0.14347337 0.14150798 0.15310922
#>  [4,] 0.24475652 0.25554571 0.25194648 0.24844722 0.24504383 0.25524595
#>  [5,] 0.35802512 0.36719619 0.36202442 0.35699630 0.35210594 0.36086127
#>  [6,] 0.49332840 0.50056656 0.49351633 0.48666194 0.47999534 0.48702243
#>  [7,] 0.63213121 0.63738648 0.62840921 0.61968130 0.61119252 0.61644667
#>  [8,] 0.76988514 0.77317250 0.76228275 0.75169548 0.74139829 0.74489290
#>  [9,] 0.87925020 0.88097519 0.86856709 0.85650366 0.84477073 0.84686843
#> [10,] 0.93146306 0.93244216 0.92805636 0.92318524 0.91053887 0.91174780
#> [11,] 1.01449275 1.01428571 1.01408451 1.01388889 1.01369863 1.01351351
#>            [,76]      [,77]      [,78]      [,79]      [,80]      [,81]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.05084067 0.05017172 0.04952014 0.04888526 0.04826646 0.04766313
#>  [3,] 0.15106777 0.14908003 0.14714393 0.14525747 0.14341877 0.14162603
#>  [4,] 0.25184267 0.24852895 0.24599518 0.24284139 0.23976745 0.23789032
#>  [5,] 0.35604978 0.35353157 0.35253045 0.34801083 0.34360563 0.34323191
#>  [6,] 0.48415531 0.48288307 0.48309839 0.47690482 0.47473929 0.47552779
#>  [7,] 0.61575848 0.61568620 0.61707312 0.60916192 0.60971334 0.61161606
#>  [8,] 0.74639667 0.74753704 0.75010794 0.74049117 0.74377128 0.74679966
#>  [9,] 0.84891018 0.85089821 0.85283459 0.84555869 0.84751365 0.84941973
#> [10,] 0.91292450 0.91407023 0.91518620 0.91334759 0.91444445 0.91551390
#> [11,] 1.01333333 1.01315789 1.01298701 1.01282051 1.01265823 1.01250000
#>           [,82]      [,83]      [,84]      [,85]      [,86]      [,87]
#>  [1,] 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.0502141 0.04960174 0.04900412 0.04842074 0.04785109 0.04729468
#>  [3,] 0.1485875 0.15180045 0.14997153 0.14818615 0.14644279 0.14473996
#>  [4,] 0.2472991 0.25453869 0.25280914 0.24979951 0.24686069 0.24399022
#>  [5,] 0.3513402 0.35925064 0.35897056 0.35730768 0.35310406 0.34899820
#>  [6,] 0.4820028 0.48831980 0.48919570 0.48862949 0.48288091 0.47726602
#>  [7,] 0.6164109 0.62108884 0.62307616 0.62356310 0.61622706 0.60906163
#>  [8,] 0.7499256 0.75297528 0.75595148 0.75750352 0.74859172 0.73988716
#>  [9,] 0.8512787 0.85309241 0.85486239 0.85659021 0.85123494 0.84133686
#> [10,] 0.9165569 0.91757453 0.91856761 0.91953704 0.91951189 0.90881989
#> [11,] 1.0123457 1.01219512 1.01204819 1.01190476 1.01176471 1.01162791
#>            [,88]      [,89]      [,90]      [,91]      [,92]      [,93]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04675106 0.04650102 0.04597854 0.04546767 0.04504879 0.04455913
#>  [3,] 0.14307629 0.14523192 0.14360010 0.14362143 0.14542135 0.14391654
#>  [4,] 0.24118573 0.24572677 0.24296579 0.24474713 0.24873320 0.24865983
#>  [5,] 0.34498673 0.35184848 0.34876109 0.35223100 0.35833334 0.35962307
#>  [6,] 0.47178020 0.47778270 0.47654307 0.48145727 0.48715554 0.48959943
#>  [7,] 0.60206092 0.60658295 0.60715890 0.61152380 0.61579277 0.61939283
#>  [8,] 0.73381486 0.73683969 0.73921483 0.74211244 0.74494637 0.74771869
#>  [9,] 0.83849050 0.84032584 0.84211993 0.84387415 0.84558982 0.84726819
#> [10,] 0.90958989 0.91061728 0.91162158 0.91260356 0.91356396 0.91450349
#> [11,] 1.01149425 1.01136364 1.01123596 1.01111111 1.01098901 1.01086957
#>            [,94]      [,95]      [,96]      [,97]      [,98]      [,99]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.04514111 0.05529918 0.05564644 0.06019786 0.05957727 0.06050847
#>  [3,] 0.14716011 0.15623287 0.15910658 0.16771136 0.16781135 0.17150363
#>  [4,] 0.25450706 0.26243784 0.26778251 0.27540977 0.27712958 0.28357188
#>  [5,] 0.36650885 0.37324811 0.37984550 0.38630545 0.38961212 0.39584056
#>  [6,] 0.49508760 0.50045901 0.50571734 0.51086612 0.51561877 0.52056144
#>  [7,] 0.62348538 0.62749085 0.63141200 0.63525146 0.63901175 0.64269531
#>  [8,] 0.75043140 0.75308638 0.75568547 0.75823042 0.76072289 0.76316449
#>  [9,] 0.84891047 0.85051780 0.85209130 0.85363202 0.85514096 0.85661912
#> [10,] 0.91542280 0.91632256 0.91720338 0.91806584 0.91891052 0.91973797
#> [11,] 1.01075269 1.01063830 1.01052632 1.01041667 1.01030928 1.01020408
#>           [,100]    [,101]     [,102]     [,103]     [,104]     [,105]
#>  [1,] 0.00000000 0.0000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.05989727 0.0592983 0.06357523 0.06295194 0.06234076 0.06174133
#>  [3,] 0.16977127 0.1680736 0.17631045 0.17458192 0.17493025 0.17324823
#>  [4,] 0.28217030 0.2804993 0.28762306 0.28591951 0.28783568 0.28506803
#>  [5,] 0.39559643 0.3950433 0.40103295 0.40042492 0.40387818 0.40134423
#>  [6,] 0.52134897 0.5217904 0.52652516 0.52689421 0.53148747 0.53051487
#>  [7,] 0.64454063 0.6460023 0.64950720 0.65087793 0.65426747 0.65490265
#>  [8,] 0.76555677 0.7679012 0.77019921 0.77245216 0.77466136 0.77682808
#>  [9,] 0.85806741 0.8594867 0.86087796 0.86224190 0.86357936 0.86489109
#> [10,] 0.92054869 0.9213432 0.92212199 0.92288550 0.92363418 0.92436847
#> [11,] 1.01010101 1.0100000 1.00990099 1.00980392 1.00970874 1.00961538
#>           [,106]     [,107]     [,108]     [,109]     [,110]     [,111]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06115331 0.06615304 0.06553479 0.06492799 0.06717172 0.07565198
#>  [3,] 0.17159824 0.17941335 0.17834844 0.17669706 0.18075480 0.18820248
#>  [4,] 0.28235310 0.28912335 0.28930461 0.28769674 0.29357554 0.29999758
#>  [5,] 0.39752190 0.40320566 0.40459223 0.40400245 0.40947032 0.41483877
#>  [6,] 0.52546235 0.52993912 0.53241277 0.53272503 0.53701195 0.54122093
#>  [7,] 0.64866549 0.65197996 0.65523249 0.65649308 0.65964452 0.66273866
#>  [8,] 0.76942971 0.77160490 0.77373944 0.77583444 0.77789101 0.77991018
#>  [9,] 0.86028486 0.86160292 0.86289636 0.86416583 0.86541202 0.86663555
#> [10,] 0.92349087 0.92421266 0.92492095 0.92561613 0.92629855 0.92696856
#> [11,] 1.00952381 1.00943396 1.00934579 1.00925926 1.00917431 1.00909091
#>           [,112]     [,113]     [,114]     [,115]     [,116]     [,117]
#>  [1,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>  [2,] 0.07497043 0.07430105 0.07364352 0.07299752 0.07236276 0.07173895
#>  [3,] 0.18650696 0.18484172 0.18320595 0.18177391 0.18277046 0.18119485
#>  [4,] 0.29765738 0.29547339 0.29408482 0.29375556 0.29635837 0.29380356
#>  [5,] 0.41412322 0.41289050 0.41246707 0.41317478 0.41731918 0.41372160
#>  [6,] 0.54202607 0.54164250 0.54208390 0.54373003 0.54769760 0.54297607
#>  [7,] 0.66510830 0.66561692 0.66696543 0.66959153 0.67246464 0.66685902
#>  [8,] 0.78189297 0.78335002 0.78526728 0.78715090 0.78900176 0.78540213
#>  [9,] 0.86783703 0.86901705 0.87017619 0.87131500 0.87243400 0.87112576
#> [10,] 0.92762650 0.92827269 0.92890745 0.92953107 0.93014384 0.93074605
#> [11,] 1.00900901 1.00892857 1.00884956 1.00877193 1.00869565 1.00862069
#>           [,118]     [,119]    [,120]     [,121]     [,122]    [,123]
#>  [1,] 0.00000000 0.00000000 0.0000000 0.00000000 0.00000000 0.0000000
#>  [2,] 0.07112579 0.07052303 0.0699304 0.06934765 0.06877453 0.0682108
#>  [3,] 0.17964617 0.17812375 0.1766269 0.17515502 0.17370746 0.1722836
#>  [4,] 0.29129241 0.28882383 0.2863967 0.28401010 0.28166291 0.2793542
#>  [5,] 0.41018551 0.40670937 0.4032916 0.39993088 0.39662566 0.3933746
#>  [6,] 0.53833524 0.53380021 0.5293145 0.52490354 0.52056549 0.5162986
#>  [7,] 0.66115937 0.65821659 0.6526854 0.64724631 0.64189717 0.6366593
#>  [8,] 0.77868929 0.77738363 0.7708510 0.76442723 0.75810965 0.7552303
#>  [9,] 0.86368024 0.86428744 0.8570245 0.85197378 0.84493268 0.8446527
#> [10,] 0.92279096 0.92344527 0.9195704 0.91816258 0.91057446 0.9113075
#> [11,] 1.00854701 1.00847458 1.0084034 1.00833333 1.00826446 1.0081967
#>           [,124]     [,125]     [,126]
#>  [1,] 0.00000000 0.00000000 0.00000000
#>  [2,] 0.06765624 0.06952294 0.07696675
#>  [3,] 0.17088295 0.17432948 0.18093484
#>  [4,] 0.27708303 0.28208542 0.28782874
#>  [5,] 0.39017647 0.39509440 0.39993364
#>  [6,] 0.51210101 0.51603568 0.51990740
#>  [7,] 0.63169104 0.63466127 0.63758398
#>  [8,] 0.75213305 0.75413197 0.75609892
#>  [9,] 0.84366347 0.84492425 0.84616486
#> [10,] 0.91202853 0.91273798 0.91343607
#> [11,] 1.00813008 1.00806452 1.00800000
#>