Random Normal Probability Integral Transform

A function to create the normal conditional (randomized) quantile residuals. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

compnormRandPIT(object)

Arguments

object
an object class "cmp", obtained from a call to glm.cmp.

Value

A list consisting of two elements:

rt

the normal conditional randomized quantile residuals

rdMid

the midpoints of the predictive probability intervals

Details

The function compPredProb produces the non-randomized probability integral transform(PIT). It returns estimates of the cumulative predictive probabilities as upper and lower bounds of a collection of intervals. If the model is correct, a histogram drawn using these estimated probabilities should resemble a histogram obtained from a sample from the uniform distribution.

This function aims to produce observations which instead resemble a sample from a normal distribution. Such a sample can then be examined by the usual tools for checking normality, such as histograms and normal Q-Q plots.

For each of the intervals produced by compPredProb, a random uniform observation is generated, which is then converted to a normal observation by applying the inverse standard normal distribution function (using qnorm). The vector of these values is returned by the function in the list element rt. In addition non-random observations which should appear similar to a sample from a normal distribution are obtained by applying qnorm to the mid-points of the predictive distribution intervals. The vector of these values is returned by the function in the list element rtMid.

References

Berkowitz, J. (2001). Testing density forecasts, with applications to risk management. Journal of Business \& Economic Statistics, 19, 465--474.

Dunn, P. K. and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236--244.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Series Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)
M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
compnormRandPIT(M.bids)

#> $rt
#>   [1] -0.20016923 -1.81257340 -1.51742706 -0.32208383 -0.24344862  0.55642268
#>   [7]  0.15154483 -0.49320802  0.17992035 -0.71812233 -0.15836976  0.29653863
#>  [13]  0.09700498 -0.69146066  0.77970427  0.46652851  0.29964815  0.20915356
#>  [19]  0.65536683  0.36871878  1.78871486 -0.72276588 -0.22117067  0.68785484
#>  [25] -0.73335437 -0.54085675 -0.19483791  0.61936118  1.40171748 -0.34555970
#>  [31]  0.57975560 -0.61930375  1.66764307 -0.32674332 -1.19632111  3.00210062
#>  [37]  0.33943394  0.29955958 -0.05433438 -0.43315686 -0.03963827  0.21412794
#>  [43] -1.42418061  0.07068576  0.19905202 -0.42327149 -1.36287723 -0.82081147
#>  [49] -0.35647896  0.58105334 -0.02207692  0.72513896  0.40065667  0.21949232
#>  [55]  1.24537256 -0.94716029  0.57007365  0.19820170  0.35897732  0.55465832
#>  [61] -1.12223229  0.95901413 -2.57594557 -0.97646737  2.97256943 -0.21428494
#>  [67]  0.08282303 -0.24199080  0.69065481  0.09428069 -0.90708713  1.26858256
#>  [73]  1.06581807  2.19881896 -1.32972210  0.58666363  0.01491610 -0.32753924
#>  [79]  1.17220392 -0.17959368  0.53861133 -1.42780945 -0.75597382  0.32313284
#>  [85]  0.31710731  0.73548516  1.87905177  1.09433940 -1.27563082  0.08810100
#>  [91] -0.11024057 -0.33491828 -0.33476450 -1.10526951 -1.97613715 -0.61195475
#>  [97] -1.26430801 -0.29926234 -0.85859335  0.26497119  0.16016535 -1.04134216
#> [103] -0.66031188 -0.12961446  0.51059534  0.93994542 -1.48782449 -0.02032077
#> [109]  0.02041376 -0.82445095 -2.02222951 -0.05412629 -0.28020510  0.37054248
#> [115]  0.14630826 -0.90068202  0.55046942  2.90301483  0.09395407  1.13004446
#> [121]  0.78140356  1.93282171  0.88283407  1.07307178 -0.89773602 -2.10184543
#> 
#> $rtMid
#>   [1] -0.4898611323 -1.2821919153 -0.9046770512 -0.1188617721 -0.1341909866
#>   [6]  0.7653197820  0.2351342834  0.0171311442  0.3047564122 -1.1693445756
#>  [11]  0.0792291624  0.1389788809 -0.1034633916 -0.3370808604  0.3111062972
#>  [16]  0.1809609156  0.6063836876  0.1827395454  0.4748640145  0.1493596221
#>  [21]  2.0323934828 -0.4065197894 -0.0485085333  0.5966853562 -0.7121903189
#>  [26] -0.4681921021 -0.4067818741  0.0763152207  1.5778128638 -0.0213149356
#>  [31]  0.4153293460 -0.6181238891  1.5710161924 -0.4889149027 -0.6975914163
#>  [36]  2.8736490827  0.0293237945 -0.2299208557 -0.5073994133 -0.3107357141
#>  [41]  0.0040097052  0.3722310596 -1.0535886944  0.0433560807  0.2062884858
#>  [46] -0.6344091014 -1.0333889115 -1.0037232276  0.1400271386  0.6852327750
#>  [51] -0.3569771913  0.9064716727  0.3550394301  0.4009376435  1.1722269817
#>  [56] -0.5882166481  0.0362186712  0.1670730737  0.3196204704  0.7366036463
#>  [61] -1.0515728547  1.0488734656 -1.3465552896 -0.9436750406  3.0214249153
#>  [66]  0.1992821900  0.1892704657 -0.3062286212  0.3521637730  0.2239210507
#>  [71] -0.9333306526  1.2075715758  1.2234081811  2.2949676276 -1.3122885422
#>  [76]  0.1964275392  0.1269510964  0.0006073091  1.0115904438  0.1407701460
#>  [81] -0.0423072981 -1.0175050879 -0.7704859707 -0.0680241458  0.0658510437
#>  [86]  0.9119789346  1.8459583763  0.7570065985 -0.6610429080  0.1146655654
#>  [91] -0.4192371640 -0.6305049231 -0.2280870584 -0.7868244581 -1.8992387903
#>  [96] -0.7696486855 -1.2730134252 -0.4626263963 -0.8699281288 -0.1090793286
#> [101] -0.0729563025 -1.2752248509 -0.0715036311 -0.5068547314  0.0602529066
#> [106]  0.9397798742 -1.3557135706 -0.3086746890 -0.0736638092 -0.9880948765
#> [111] -2.2394335401 -0.1111342437  0.0010516271 -0.1014646815 -0.2458552865
#> [116] -0.6169574062  0.6335429515  2.7480721563  0.4121711442  1.3053286269
#> [121]  1.0357671748  1.8208362363  0.5919727771  0.6312394116 -0.9654768788
#> [126] -2.1524761412
#>