CMP Regression Diagnostic

This suite of functions provides the basic quantities which are used in forming a wide variety of diagnostics for checking the quality of regression fits.

Usage

# S3 method for class 'cmp'
influence(model, ...)

# S3 method for class 'cmp'
hatvalues(model, ...)

# S3 method for class 'cmp'
rstandard(
  model,
  infl = influence.cmp(model),
  type = c("deviance", "pearson"),
  ...
)

# S3 method for class 'cmp'
cooks.distance(
  model,
  infl = influence(model),
  res = infl$pear_res,
  hat = infl$hat,
  ...
)

Arguments

model

an object class 'cmp', obtained from a call to glm.cmp.

...

other arguments passed to or from other methods (currently unused).

infl

influence structure as returned by influence, only for rstudent and cooks.distance.

type

type of residuals for rstandard. The alternatives are: 'deviance' (default), and 'pearson'.

res

residuals, with proper default.

hat

hat values \(H[i,i]\), see default.

Value

influence.cmp returns a list with components hat (the leverage values), dev_res (the deviance residuals) and pear_res (the Pearson residuals).

hatvalues.cmp returns a numeric vector of leverage values.

rstandard.cmp returns a numeric vector of standardized residuals.

cooks.distance.cmp returns a numeric vector of Cook's distances.

Examples

data(attendance)
M.attendance <- glm.cmp(daysabs ~ gender + math + prog, data = attendance)
influence(M.attendance)
#> $hat
#>           1           2           3           4           5           6 
#> 0.011522827 0.010632763 0.012781978 0.014026784 0.020060946 0.013743373 
#>           7           8           9          10          11          12 
#> 0.011673460 0.018385257 0.009677630 0.012421447 0.016174725 0.011662760 
#>          13          14          15          16          17          18 
#> 0.014363242 0.033768524 0.023438692 0.020310790 0.010023059 0.009918405 
#>          19          20          21          22          23          24 
#> 0.034193688 0.017916760 0.012503108 0.011113570 0.009296421 0.011522827 
#>          25          26          27          28          29          30 
#> 0.013073726 0.029299981 0.017018891 0.012503108 0.013265512 0.034898026 
#>          31          32          33          34          35          36 
#> 0.019361683 0.010728631 0.009318287 0.009351266 0.012718943 0.012781978 
#>          37          38          39          40          41          42 
#> 0.015378079 0.018141277 0.042465745 0.009918405 0.010268665 0.010023059 
#>          43          44          45          46          47          48 
#> 0.009574397 0.010303698 0.012158816 0.011096970 0.014473168 0.009395409 
#>          49          50          51          52          53          54 
#> 0.011904733 0.017240084 0.013297415 0.010744457 0.009450190 0.010278484 
#>          55          56          57          58          59          60 
#> 0.019035244 0.010415895 0.040116864 0.041261488 0.011983752 0.009918405 
#>          61          62          63          64          65          66 
#> 0.009785413 0.011096970 0.014281254 0.009351980 0.010563029 0.009897636 
#>          67          68          69          70          71          72 
#> 0.031516121 0.014727603 0.016174725 0.009632306 0.013850382 0.016193259 
#>          73          74          75          76          77          78 
#> 0.021160080 0.011214951 0.012184511 0.036508084 0.009684666 0.012460161 
#>          79          80          81          82          83          84 
#> 0.019361683 0.012503108 0.010139795 0.015830799 0.036497013 0.010744457 
#>          85          86          87          88          89          90 
#> 0.014798646 0.032715111 0.009691062 0.015097830 0.014151026 0.010632763 
#>          91          92          93          94          95          96 
#> 0.011096970 0.013448620 0.023707877 0.031582967 0.016288485 0.013940826 
#>          97          98          99         100         101         102 
#> 0.012158816 0.014999820 0.009285619 0.018866880 0.032109019 0.009450769 
#>         103         104         105         106         107         108 
#> 0.014281254 0.018385257 0.032060744 0.011096970 0.043390956 0.011260414 
#>         109         110         111         112         113         114 
#> 0.009296990 0.009595345 0.015830799 0.019541135 0.011904733 0.010814960 
#>         115         116         117         118         119         120 
#> 0.009296990 0.009584872 0.009918405 0.039102177 0.019803600 0.018385257 
#>         121         122         123         124         125         126 
#> 0.017018891 0.039032828 0.014281254 0.018141277 0.020770702 0.013743373 
#>         127         128         129         130         131         132 
#> 0.035577291 0.009541068 0.011673460 0.010023059 0.009524193 0.010728631 
#>         133         134         135         136         137         138 
#> 0.019541135 0.009595345 0.030218327 0.011662760 0.017240084 0.014363242 
#>         139         140         141         142         143         144 
#> 0.012987414 0.031233627 0.015097830 0.015368184 0.039667426 0.010744457 
#>         145         146         147         148         149         150 
#> 0.010728631 0.010562513 0.009511052 0.010038289 0.011983752 0.030710982 
#>         151         152         153         154         155         156 
#> 0.009318287 0.009515373 0.029043898 0.030408634 0.033143131 0.038510492 
#>         157         158         159         160         161         162 
#> 0.009450769 0.009882075 0.016174725 0.011939315 0.011939315 0.011904622 
#>         163         164         165         166         167         168 
#> 0.016769717 0.009541068 0.012383792 0.012055711 0.012383792 0.012434894 
#>         169         170         171         172         173         174 
#> 0.013045220 0.013448620 0.014999321 0.011904622 0.011999583 0.010906595 
#>         175         176         177         178         179         180 
#> 0.013045220 0.011999583 0.012718209 0.025288351 0.012101869 0.013211355 
#>         181         182         183         184         185         186 
#> 0.010562513 0.028461203 0.018443158 0.014151026 0.012718209 0.013634188 
#>         187         188         189         190         191         192 
#> 0.028588754 0.012948487 0.011971824 0.028537886 0.013010583 0.013506390 
#>         193         194         195         196         197         198 
#> 0.019976532 0.014026784 0.025926910 0.031861247 0.012747764 0.010562513 
#>         199         200         201         202         203         204 
#> 0.012625687 0.014999321 0.012434894 0.019576775 0.016174725 0.011935193 
#>         205         206         207         208         209         210 
#> 0.009511052 0.014883818 0.033768524 0.016699591 0.028731543 0.012227377 
#>         211         212         213         214         215         216 
#> 0.012045613 0.013799395 0.029959091 0.011935193 0.013211355 0.012747764 
#>         217         218         219         220         221         222 
#> 0.024426372 0.011260414 0.013743373 0.012131868 0.009541068 0.013297415 
#>         223         224         225         226         227         228 
#> 0.019187983 0.010415895 0.009842343 0.014473168 0.013473769 0.014730566 
#>         229         230         231         232         233         234 
#> 0.012096107 0.025926910 0.011673460 0.012131868 0.012520978 0.013045220 
#>         235         236         237         238         239         240 
#> 0.030817136 0.016462081 0.019976532 0.036839194 0.010462361 0.019976532 
#>         241         242         243         244         245         246 
#> 0.017078742 0.015023254 0.012071143 0.013829159 0.009897636 0.013612900 
#>         247         248         249         250         251         252 
#> 0.012184511 0.010268665 0.009684666 0.012227377 0.013819928 0.013473769 
#>         253         254         255         256         257         258 
#> 0.010728631 0.013155728 0.015722998 0.012304053 0.013310028 0.012248540 
#>         259         260         261         262         263         264 
#> 0.012096107 0.012421447 0.009745111 0.011999583 0.028967602 0.009632306 
#>         265         266         267         268         269         270 
#> 0.044087556 0.015360645 0.014730566 0.012794687 0.010562513 0.011903398 
#>         271         272         273         274         275         276 
#> 0.032910957 0.012434894 0.018127941 0.012574067 0.012043855 0.010415895 
#>         277         278         279         280         281         282 
#> 0.013045220 0.012110839 0.012574067 0.014151026 0.017733749 0.013211355 
#>         283         284         285         286         287         288 
#> 0.013010583 0.013506390 0.029125642 0.013470124 0.012747764 0.009918405 
#>         289         290         291         292         293         294 
#> 0.012071143 0.012948487 0.014381798 0.009296421 0.015892955 0.014634234 
#>         295         296         297         298         299         300 
#> 0.011903398 0.025926910 0.031307980 0.015110001 0.012434894 0.014999321 
#>         301         302         303         304         305         306 
#> 0.012300177 0.010415895 0.013010583 0.017741648 0.011935193 0.017149520 
#>         307         308         309         310         311         312 
#> 0.011935193 0.012055711 0.013019187 0.010814960 0.013506390 0.010889963 
#>         313         314 
#> 0.015684207 0.013211355 
#> 
#> $dev_res
#>            1            2            3            4            5            6 
#> -0.264431012 -0.464137177 -1.137050858 -0.907077078 -0.975541893  0.836872639 
#>            7            8            9           10           11           12 
#>  0.561358082 -0.108788372  0.605119059  1.468062912  0.114354834 -0.304941513 
#>           13           14           15           16           17           18 
#> -0.205905445 -0.413390019 -0.834216769 -1.884087266 -1.420603965 -2.150003744 
#>           19           20           21           22           23           24 
#> -0.375219202  1.468969284 -1.132530084 -1.968586213 -1.254485960 -1.962546694 
#>           25           26           27           28           29           30 
#> -0.090579641 -0.343694861 -0.090124123 -2.192511846 -0.747483552 -0.759671179 
#>           31           32           33           34           35           36 
#> -2.150129851 -1.440686754 -0.616246146 -2.035063837 -1.947450419  1.559609126 
#>           37           38           39           40           41           42 
#>  0.069638238  0.063669866 -0.794306116  1.640264602 -0.441466783  0.990429449 
#>           43           44           45           46           47           48 
#>  1.120302460  0.324432972  0.262711559 -0.646728391 -0.728359958 -0.626501161 
#>           49           50           51           52           53           54 
#>  0.269537221  1.565492698 -0.175745577 -1.974626408  2.108171868  0.610269640 
#>           55           56           57           58           59           60 
#>  1.745950017 -1.188726166  1.116959999 -0.395382557 -0.872605764 -1.068970109 
#>           61           62           63           64           65           66 
#> -1.287173952 -0.170327623 -0.540058225  0.308774414 -0.294815927  1.086435428 
#>           67           68           69           70           71           72 
#> -0.702750468 -0.715281202 -0.146944863 -0.350633458 -1.935376059  0.679889712 
#>           73           74           75           76           77           78 
#> -1.712453459  0.161463314 -1.694337746  0.007218208 -0.425820818 -0.761790008 
#>           79           80           81           82           83           84 
#>  2.117266900  1.246871512  0.546128509  0.210352798 -0.903468584  0.089284059 
#>           85           86           87           88           89           90 
#> -1.926321678  0.039332889  0.459385767  0.205096654  0.657176570 -0.109085485 
#>           91           92           93           94           95           96 
#>  1.300390076 -0.223460455  0.137758085  1.456204375 -0.273958794 -1.364325963 
#>           97           98           99          100          101          102 
#>  0.262711559  0.430279366  1.350221783  0.469534272 -0.522836660 -2.041111396 
#>          103          104          105          106          107          108 
#> -0.540058225  0.263758194  1.092100817 -1.448709214 -0.234861993 -0.682397705 
#>          109          110          111          112          113          114 
#> -2.019948078  0.365951223  1.277017721 -0.316132979 -0.713006349  0.424810443 
#>          115          116          117          118          119          120 
#>  0.784627804 -0.193545952  2.117705421  0.121611969 -0.548959360  0.263758194 
#>          121          122          123          124          125          126 
#> -0.235700445 -0.709293864 -1.014259547  0.249111858  0.100598789  0.836872639 
#>          127          128          129          130          131          132 
#> -0.640162681 -0.390539047 -0.672232741  1.817248363  0.913757050  0.699075217 
#>          133          134          135          136          137          138 
#>  0.233655031 -1.279013570 -0.795133058  0.010430499 -1.229716231  0.518167288 
#>          139          140          141          142          143          144 
#> -0.465617735  1.118708377 -1.168627424  0.876233367 -0.942730520 -0.087759432 
#>          145          146          147          148          149          150 
#> -0.458350244 -0.485698591 -0.768160860  1.311929408  0.375666941  1.339704326 
#>          151          152          153          154          155          156 
#> -0.029313270  1.035007596  0.339418726 -0.109086199  0.140367332  0.290978301 
#>          157          158          159          160          161          162 
#> -2.041111396 -1.205215135  0.114354834 -1.586940927 -1.586940927 -0.148942645 
#>          163          164          165          166          167          168 
#> -0.898463787 -1.041569772 -1.545383985 -1.682353341 -1.545383985  1.009923018 
#>          169          170          171          172          173          174 
#> -0.054955887  0.373189124  0.934064138  0.496977926 -0.593956430 -0.164176365 
#>          175          176          177          178          179          180 
#> -1.527629573 -1.563172868  0.371421104 -0.607567428 -1.688344571  0.396413390 
#>          181          182          183          184          185          186 
#> -0.968001963  0.762113430 -0.919828531 -1.506963298 -0.708108700  2.799196939 
#>          187          188          189          190          191          192 
#> -0.191591771  2.382483209 -1.956507776 -0.040627461 -0.686403737 -0.221118065 
#>          193          194          195          196          197          198 
#> -0.207509039 -0.704779897 -0.771112884 -0.311721806  1.084637000 -2.083341944 
#>          199          200          201          202          203          204 
#> -1.287004495 -0.646701815 -1.649439662  1.509090051 -0.889902263 -1.569109946 
#>          205          206          207          208          209          210 
#> -1.380318704 -1.655543238 -0.737624877 -1.577897935 -0.923515010 -1.658409430 
#>          211          212          213          214          215          216 
#> -0.313230638 -1.512862681 -0.810463804 -1.569109946  0.396413390  0.129435743 
#>          217          218          219          220          221          222 
#> -0.842835445 -2.071235745 -1.270675656  0.321699117 -1.392421504  0.235490336 
#>          223          224          225          226          227          228 
#> -1.793459853 -0.298181310  0.115641994  1.134231842 -1.634626403 -1.498122674 
#>          229          230          231          232          233          234 
#> -1.557239513 -1.956382179 -0.940218373  0.562471880 -0.716849462 -1.527629573 
#>          235          236          237          238          239          240 
#> -0.986564713 -1.237924035 -0.518811736 -0.519898225 -0.270261300 -0.207509039 
#>          241          242          243          244          245          246 
#> -0.764958581 -0.610898704  0.763201394  1.215124984  1.507273921 -1.360281476 
#>          247          248          249          250          251          252 
#>  0.898860177 -0.441466783 -1.283094715 -0.734306600 -1.736340347  0.383646874 
#>          253          254          255          256          257          258 
#> -2.168211855  2.150848368 -0.874955325 -0.353811834 -0.399196432 -1.604801038 
#>          259          260          261          262          263          264 
#> -0.585109733  2.815366934 -1.059847693  1.930267052 -0.937998598 -0.350633458 
#>          265          266          267          268          269          270 
#>  0.807026150 -1.757370852  0.943779474  0.298195226 -0.968001963 -0.138549489 
#>          271          272          273          274          275          276 
#>  0.041945468 -1.649439662  0.239841692 -1.539461815 -0.589533989  0.520796394 
#>          277          278          279          280          281          282 
#> -0.540773571 -1.598844538 -1.539461815 -0.509640734  1.093265264  0.112424752 
#>          283          284          285          286          287          288 
#> -1.625671576 -0.681838341 -0.086190230 -0.838347762 -0.677635518 -0.120940762 
#>          289          290          291          292          293          294 
#>  0.763201394 -1.634502273  1.302245797 -2.029017066 -0.885618409 -2.116780814 
#>          295          296          297          298          299          300 
#>  0.509676348  1.464884344  0.126093021 -1.754365348 -1.649439662 -0.646701815 
#>          301          302          303          304          305          306 
#>  0.268845155 -1.188726166 -0.226477946 -0.489101521 -0.128143892 -0.575296118 
#>          307          308          309          310          311          312 
#>  0.227929365  0.928302045 -0.825410236 -1.315675470  0.123920620 -2.077288335 
#>          313          314 
#> -1.589792055 -0.231407020 
#> 
#> $pear_res
#>            1            2            3            4            5            6 
#> -0.243678918 -0.402285522 -0.791080887 -0.683891320 -0.721028357  1.051034713 
#>            7            8            9           10           11           12 
#>  0.656271807 -0.105384509  0.717390282  2.203676900  0.118457256 -0.278091560 
#>           13           14           15           16           17           18 
#> -0.193462029 -0.365945477 -0.628111607 -0.947929237 -0.888703362 -1.015626727 
#>           19           20           21           22           23           24 
#> -0.335054654  2.119655162 -0.788975833 -0.969648871 -0.824624314 -0.968111759 
#>           25           26           27           28           29           30 
#> -0.088251576 -0.310479536 -0.087780858 -1.026576817 -0.586218819 -0.598532062 
#>           31           32           33           34           35           36 
#> -1.015659051 -0.896099069 -0.507388453 -0.986482018 -0.964261437  2.280621532 
#>           37           38           39           40           41           42 
#>  0.071051280  0.064909307 -0.628345916  2.451497595 -0.386516983  1.282200480 
#>           43           44           45           46           47           48 
#>  1.496837262  0.355862700  0.283132445 -0.530552994 -0.574722140 -0.514194817 
#>           49           50           51           52           53           54 
#>  0.291054837  2.335522665 -0.166805584 -0.971184292  3.520651751  0.721916936 
#>           55           56           57           58           59           60 
#>  2.635343579 -0.797522458  1.464433651 -0.353448072 -0.664618465 -0.758802703 
#>           61           62           63           64           65           66 
#> -0.837696798 -0.162070225 -0.457433449  0.337572694 -0.270163558  1.448461443 
#>           67           68           69           70           71           72 
#> -0.573326465 -0.575038047 -0.140272077 -0.315410035 -0.961172209  0.824721355 
#>           73           74           75           76           77           78 
#> -0.901354359  0.169168704 -0.896163917  0.007231926 -0.373415591 -0.594719742 
#>           79           80           81           82           83           84 
#>  3.477722983  1.705834503  0.633720120  0.223933385 -0.681892418  0.091688601 
#>           85           86           87           88           89           90 
#> -0.958849674  0.039762170  0.522030473  0.217101701  0.802217669 -0.105603484 
#>           91           92           93           94           95           96 
#>  1.803171212 -0.208846385  0.143621221  2.071880392 -0.252983301 -0.867613110 
#>           97           98           99          100          101          102 
#>  0.283132445  0.485002824  1.917809215  0.534292408 -0.448511739 -0.988009276 
#>          103          104          105          106          107          108 
#> -0.457433449  0.284068046  1.448580031 -0.899035764 -0.220546730 -0.550491925 
#>          109          110          111          112          113          114 
#> -0.982663570  0.406180910  1.797872744 -0.288416230 -0.569810225  0.478778086 
#>          115          116          117          118          119          120 
#>  0.973583652 -0.182735164  3.478793691  0.125514493 -0.461557308  0.284068046 
#>          121          122          123          124          125          126 
#> -0.219781383 -0.567487497 -0.731909395  0.268260353  0.103709237  1.051034713 
#>          127          128          129          130          131          132 
#> -0.533617400 -0.347144380 -0.543990325  2.814603191  1.164140476  0.842705973 
#>          133          134          135          136          137          138 
#>  0.249106771 -0.834456851 -0.624576436  0.010462393 -0.814546406  0.599526181 
#>          139          140          141          142          143          144 
#> -0.401850805  1.461940859 -0.805641283  1.112213153 -0.716537232 -0.085456881 
#>          145          146          147          148          149          150 
#> -0.399297629 -0.418221101 -0.603665107  1.836745334  0.416531833  1.834142394 
#>          151          152          153          154          155          156 
#> -0.029058569  1.367593811  0.371049235 -0.105773678  0.145649569  0.314299367 
#>          157          158          159          160          161          162 
#> -0.988009276 -0.804424109  0.118457256 -0.863999218 -0.863999218 -0.141894068 
#>          163          164          165          166          167          168 
#> -0.662175038 -0.745445389 -0.850842044 -0.892695194 -0.850842044  1.350020959 
#>          169          170          171          172          173          174 
#> -0.053983064  0.415069379  1.226499675  0.578353972 -0.484913877 -0.156496608 
#>          175          176          177          178          179          180 
#> -0.845088652 -0.856526344  0.416175020 -0.498417434 -0.894432772  0.447552077 
#>          181          182          183          184          185          186 
#> -0.708449857  0.924596464 -0.673141703 -0.838287753 -0.556214944  5.619984170 
#>          187          188          189          190          191          192 
#> -0.181525849  4.362837919 -0.966572948 -0.040176059 -0.543129326 -0.205749700 
#>          193          194          195          196          197          198 
#> -0.194320475 -0.568344535 -0.600213806 -0.284751366  1.480014518 -0.998678614 
#>          199          200          201          202          203          204 
#> -0.837629509 -0.518627358 -0.883019742  2.261028390 -0.657730742 -0.858405575 
#>          205          206          207          208          209          210 
#> -0.873664312 -0.884830928 -0.589139139 -0.861172343 -0.699577809 -0.885678713 
#>          211          212          213          214          215          216 
#> -0.282819533 -0.840240446 -0.633822530 -0.858405575  0.447552077  0.134817414 
#>          217          218          219          220          221          222 
#> -0.632777790 -0.995617753 -0.831130338  0.355062571 -0.878212392  0.251831033 
#>          223          224          225          226          227          228 
#> -0.923862560 -0.271935497  0.119566707  1.542240640 -0.878590715 -0.835343935 
#>          229          230          231          232          233          234 
#> -0.854639280 -0.966541053 -0.694023091  0.665688900 -0.561424028 -0.845088652 
#>          235          236          237          238          239          240 
#> -0.726872538 -0.817902909 -0.438044824 -0.443049891 -0.249048408 -0.194320475 
#>          241          242          243          244          245          246 
#> -0.589481943 -0.503820989  0.954781812  1.712086193  2.211051886 -0.866075161 
#>          247          248          249          250          251          252 
#>  1.165049137 -0.386516983 -0.836079130 -0.571724559 -0.908106607  0.431468357 
#>          253          254          255          256          257          258 
#> -1.020299429  3.757260943 -0.659082753 -0.315251968 -0.350456157 -0.869525657 
#>          259          260          261          262          263          264 
#> -0.479126687  5.620376789 -0.754379906  3.231210793 -0.700846443 -0.315410035 
#>          265          266          267          268          269          270 
#>  0.999468717 -0.913968457  1.247218569  0.327401937 -0.708449857 -0.132440919 
#>          271          272          273          274          275          276 
#>  0.042455860 -0.883019742  0.258592361 -0.848932078 -0.482025665  0.604042771 
#>          277          278          279          280          281          282 
#> -0.449535858 -0.867690957 -0.848932078 -0.428161397  1.474194506  0.116475456 
#>          283          284          285          286          287          288 
#> -0.875890013 -0.540349087 -0.084080704 -0.630352302 -0.537781138 -0.116742643 
#>          289          290          291          292          293          294 
#>  0.954781812 -0.878553371  1.878016898 -0.984954654 -0.655495885 -1.007155089 
#>          295          296          297          298          299          300 
#>  0.595395869  2.147240547  0.130698558 -0.913135446 -0.883019742 -0.518627358 
#>          301          302          303          304          305          306 
#>  0.292494840 -0.797522458 -0.210368639 -0.416973598 -0.122910160 -0.484277417 
#>          307          308          309          310          311          312 
#>  0.244837547  1.213132532 -0.623311692 -0.848895241  0.128849686 -0.997147767 
#>          313          314 
#> -0.864886323 -0.214601165 
#> 
hatvalues(M.attendance)
#>           1           2           3           4           5           6 
#> 0.011522827 0.010632763 0.012781978 0.014026784 0.020060946 0.013743373 
#>           7           8           9          10          11          12 
#> 0.011673460 0.018385257 0.009677630 0.012421447 0.016174725 0.011662760 
#>          13          14          15          16          17          18 
#> 0.014363242 0.033768524 0.023438692 0.020310790 0.010023059 0.009918405 
#>          19          20          21          22          23          24 
#> 0.034193688 0.017916760 0.012503108 0.011113570 0.009296421 0.011522827 
#>          25          26          27          28          29          30 
#> 0.013073726 0.029299981 0.017018891 0.012503108 0.013265512 0.034898026 
#>          31          32          33          34          35          36 
#> 0.019361683 0.010728631 0.009318287 0.009351266 0.012718943 0.012781978 
#>          37          38          39          40          41          42 
#> 0.015378079 0.018141277 0.042465745 0.009918405 0.010268665 0.010023059 
#>          43          44          45          46          47          48 
#> 0.009574397 0.010303698 0.012158816 0.011096970 0.014473168 0.009395409 
#>          49          50          51          52          53          54 
#> 0.011904733 0.017240084 0.013297415 0.010744457 0.009450190 0.010278484 
#>          55          56          57          58          59          60 
#> 0.019035244 0.010415895 0.040116864 0.041261488 0.011983752 0.009918405 
#>          61          62          63          64          65          66 
#> 0.009785413 0.011096970 0.014281254 0.009351980 0.010563029 0.009897636 
#>          67          68          69          70          71          72 
#> 0.031516121 0.014727603 0.016174725 0.009632306 0.013850382 0.016193259 
#>          73          74          75          76          77          78 
#> 0.021160080 0.011214951 0.012184511 0.036508084 0.009684666 0.012460161 
#>          79          80          81          82          83          84 
#> 0.019361683 0.012503108 0.010139795 0.015830799 0.036497013 0.010744457 
#>          85          86          87          88          89          90 
#> 0.014798646 0.032715111 0.009691062 0.015097830 0.014151026 0.010632763 
#>          91          92          93          94          95          96 
#> 0.011096970 0.013448620 0.023707877 0.031582967 0.016288485 0.013940826 
#>          97          98          99         100         101         102 
#> 0.012158816 0.014999820 0.009285619 0.018866880 0.032109019 0.009450769 
#>         103         104         105         106         107         108 
#> 0.014281254 0.018385257 0.032060744 0.011096970 0.043390956 0.011260414 
#>         109         110         111         112         113         114 
#> 0.009296990 0.009595345 0.015830799 0.019541135 0.011904733 0.010814960 
#>         115         116         117         118         119         120 
#> 0.009296990 0.009584872 0.009918405 0.039102177 0.019803600 0.018385257 
#>         121         122         123         124         125         126 
#> 0.017018891 0.039032828 0.014281254 0.018141277 0.020770702 0.013743373 
#>         127         128         129         130         131         132 
#> 0.035577291 0.009541068 0.011673460 0.010023059 0.009524193 0.010728631 
#>         133         134         135         136         137         138 
#> 0.019541135 0.009595345 0.030218327 0.011662760 0.017240084 0.014363242 
#>         139         140         141         142         143         144 
#> 0.012987414 0.031233627 0.015097830 0.015368184 0.039667426 0.010744457 
#>         145         146         147         148         149         150 
#> 0.010728631 0.010562513 0.009511052 0.010038289 0.011983752 0.030710982 
#>         151         152         153         154         155         156 
#> 0.009318287 0.009515373 0.029043898 0.030408634 0.033143131 0.038510492 
#>         157         158         159         160         161         162 
#> 0.009450769 0.009882075 0.016174725 0.011939315 0.011939315 0.011904622 
#>         163         164         165         166         167         168 
#> 0.016769717 0.009541068 0.012383792 0.012055711 0.012383792 0.012434894 
#>         169         170         171         172         173         174 
#> 0.013045220 0.013448620 0.014999321 0.011904622 0.011999583 0.010906595 
#>         175         176         177         178         179         180 
#> 0.013045220 0.011999583 0.012718209 0.025288351 0.012101869 0.013211355 
#>         181         182         183         184         185         186 
#> 0.010562513 0.028461203 0.018443158 0.014151026 0.012718209 0.013634188 
#>         187         188         189         190         191         192 
#> 0.028588754 0.012948487 0.011971824 0.028537886 0.013010583 0.013506390 
#>         193         194         195         196         197         198 
#> 0.019976532 0.014026784 0.025926910 0.031861247 0.012747764 0.010562513 
#>         199         200         201         202         203         204 
#> 0.012625687 0.014999321 0.012434894 0.019576775 0.016174725 0.011935193 
#>         205         206         207         208         209         210 
#> 0.009511052 0.014883818 0.033768524 0.016699591 0.028731543 0.012227377 
#>         211         212         213         214         215         216 
#> 0.012045613 0.013799395 0.029959091 0.011935193 0.013211355 0.012747764 
#>         217         218         219         220         221         222 
#> 0.024426372 0.011260414 0.013743373 0.012131868 0.009541068 0.013297415 
#>         223         224         225         226         227         228 
#> 0.019187983 0.010415895 0.009842343 0.014473168 0.013473769 0.014730566 
#>         229         230         231         232         233         234 
#> 0.012096107 0.025926910 0.011673460 0.012131868 0.012520978 0.013045220 
#>         235         236         237         238         239         240 
#> 0.030817136 0.016462081 0.019976532 0.036839194 0.010462361 0.019976532 
#>         241         242         243         244         245         246 
#> 0.017078742 0.015023254 0.012071143 0.013829159 0.009897636 0.013612900 
#>         247         248         249         250         251         252 
#> 0.012184511 0.010268665 0.009684666 0.012227377 0.013819928 0.013473769 
#>         253         254         255         256         257         258 
#> 0.010728631 0.013155728 0.015722998 0.012304053 0.013310028 0.012248540 
#>         259         260         261         262         263         264 
#> 0.012096107 0.012421447 0.009745111 0.011999583 0.028967602 0.009632306 
#>         265         266         267         268         269         270 
#> 0.044087556 0.015360645 0.014730566 0.012794687 0.010562513 0.011903398 
#>         271         272         273         274         275         276 
#> 0.032910957 0.012434894 0.018127941 0.012574067 0.012043855 0.010415895 
#>         277         278         279         280         281         282 
#> 0.013045220 0.012110839 0.012574067 0.014151026 0.017733749 0.013211355 
#>         283         284         285         286         287         288 
#> 0.013010583 0.013506390 0.029125642 0.013470124 0.012747764 0.009918405 
#>         289         290         291         292         293         294 
#> 0.012071143 0.012948487 0.014381798 0.009296421 0.015892955 0.014634234 
#>         295         296         297         298         299         300 
#> 0.011903398 0.025926910 0.031307980 0.015110001 0.012434894 0.014999321 
#>         301         302         303         304         305         306 
#> 0.012300177 0.010415895 0.013010583 0.017741648 0.011935193 0.017149520 
#>         307         308         309         310         311         312 
#> 0.011935193 0.012055711 0.013019187 0.010814960 0.013506390 0.010889963 
#>         313         314 
#> 0.015684207 0.013211355 
rstandard(M.attendance, type = "pearson")
#>            1            2            3            4            5            6 
#> -0.245095104 -0.404441433 -0.796185666 -0.688738774 -0.728371278  1.058332401 
#>            7            8            9           10           11           12 
#>  0.660136155 -0.106366836  0.720887001  2.217492166  0.119427043 -0.279727542 
#>           13           14           15           16           17           18 
#> -0.194866549 -0.372285219 -0.635604644 -0.957705003 -0.893190887 -1.020701205 
#>           19           20           21           22           23           24 
#> -0.340934252  2.138902871 -0.793954897 -0.975082333 -0.828484275 -0.973738122 
#>           25           26           27           28           29           30 
#> -0.088834184 -0.315130517 -0.088537496 -1.033055333 -0.590146182 -0.609257407 
#>           31           32           33           34           35           36 
#> -1.025636608 -0.900945056 -0.509769100 -0.991127049 -0.970452753  2.295338191 
#>           37           38           39           40           41           42 
#>  0.071603979  0.065506209 -0.642128043  2.463746259 -0.388516906  1.288674977 
#>           43           44           45           46           47           48 
#>  1.504054788  0.357710341  0.284869580 -0.533521489 -0.578926862 -0.516627508 
#>           49           50           51           52           53           54 
#>  0.292802926  2.355919078 -0.167925810 -0.976444140  3.537406008  0.725655890 
#>           55           56           57           58           59           60 
#>  2.660789644 -0.801708644  1.494722318 -0.360973675 -0.668636930 -0.762593985 
#>           61           62           63           64           65           66 
#> -0.841825730 -0.162977023 -0.460735218  0.339162340 -0.271601835  1.455683269 
#>           67           68           69           70           71           72 
#> -0.582580296 -0.579319867 -0.141420458 -0.316940161 -0.967898462  0.831481025 
#>           73           74           75           76           77           78 
#> -0.911044786  0.170125367 -0.901673981  0.007367666 -0.375237034 -0.598459883 
#>           79           80           81           82           83           84 
#>  3.511887183  1.716599674  0.636957658  0.225727234 -0.694687251  0.092185178 
#>           85           86           87           88           89           90 
#> -0.966024242  0.040428989  0.524578523  0.218759378  0.807954732 -0.106169430 
#>           91           92           93           94           95           96 
#>  1.813260126 -0.210265058  0.145354580  2.105394434 -0.255069175 -0.873724706 
#>           97           98           99          100          101          102 
#>  0.284869580  0.488681741  1.926775731  0.539405083 -0.455890553 -0.992711355 
#>          103          104          105          106          107          108 
#> -0.460735218  0.286715946  1.472375026 -0.904065954 -0.225493165 -0.553617732 
#>          109          110          111          112          113          114 
#> -0.987263576  0.408143770  1.812274850 -0.291276204 -0.573232532  0.481388260 
#>          115          116          117          118          119          120 
#>  0.978141154 -0.183617256  3.496175138  0.128042832 -0.466196577  0.286715946 
#>          121          122          123          124          125          126 
#> -0.221675816 -0.578897964 -0.737192340  0.270727262  0.104803368  1.058332401 
#>          127          128          129          130          131          132 
#> -0.543370765 -0.348812390 -0.547193521  2.828815587  1.169724142  0.847263216 
#>          133          134          135          136          137          138 
#>  0.251576947 -0.838489344 -0.634232670  0.010523942 -0.821659942  0.603878695 
#>          139          140          141          142          143          144 
#> -0.404486003  1.485320847 -0.811792745  1.120859287 -0.731186109 -0.085919708 
#>          145          146          147          148          149          150 
#> -0.401456978 -0.420447487 -0.606556493  1.846034216  0.419050298  1.862972324 
#>          151          152          153          154          155          156 
#> -0.029194911  1.374147200  0.376557883 -0.107419527  0.148124914  0.320531881 
#>          157          158          159          160          161          162 
#> -0.992711355 -0.808428502  0.119427043 -0.869203647 -0.869203647 -0.142746282 
#>          163          164          165          166          167          168 
#> -0.667798105 -0.749027213 -0.856159811 -0.898125380 -0.856159811  1.358493744 
#>          169          170          171          172          173          174 
#> -0.054338657  0.417888905  1.235802793  0.581827559 -0.487849707 -0.157357075 
#>          175          176          177          178          179          180 
#> -0.850655359 -0.861712040  0.418847035 -0.504841614 -0.899894550  0.450538081 
#>          181          182          183          184          185          186 
#> -0.712221266  0.938041719 -0.679436337 -0.844282771 -0.559786074  5.658692400 
#>          187          188          189          190          191          192 
#> -0.184177644  4.391361295 -0.972411243 -0.040761899 -0.546697396 -0.207153404 
#>          193          194          195          196          197          198 
#> -0.196290972 -0.572372987 -0.608149296 -0.289398992  1.489539117 -1.003995045 
#>          199          200          201          202          203          204 
#> -0.842967938 -0.522561197 -0.888561608  2.283490557 -0.663115460 -0.863574508 
#>          205          206          207          208          209          210 
#> -0.877848920 -0.891490189 -0.599345550 -0.868454288 -0.709849666 -0.891143645 
#>          211          212          213          214          215          216 
#> -0.284538445 -0.846098550 -0.643535705 -0.863574508  0.450538081  0.135685028 
#>          217          218          219          220          221          222 
#> -0.640650547 -1.001271077 -0.836901156  0.357236154 -0.882432154  0.253522269 
#>          223          224          225          226          227          228 
#> -0.932855719 -0.273362883  0.120159494  1.553523820 -0.884570171 -0.841565298 
#>          229          230          231          232          233          234 
#> -0.859855555 -0.979319794 -0.698109731  0.669764042 -0.564972171 -0.850655359 
#>          235          236          237          238          239          240 
#> -0.738338301 -0.824719378 -0.442486795 -0.451443322 -0.250361538 -0.196290972 
#>          241          242          243          244          245          246 
#> -0.594581158 -0.507648687  0.960597168  1.724048768  2.222075881 -0.872030934 
#>          247          248          249          250          251          252 
#>  1.172212442 -0.388516906 -0.840157350 -0.575252290 -0.914447389  0.434404817 
#>          253          254          255          256          257          258 
#> -1.025817075  3.782222254 -0.664326043 -0.317209489 -0.352811991 -0.874900291 
#>          259          260          261          262          263          264 
#> -0.482051028  5.655611991 -0.758082750  3.250773621 -0.711223361 -0.316940161 
#>          265          266          267          268          269          270 
#>  1.022257128 -0.921069948  1.256507436  0.329516756 -0.712221266 -0.133236275 
#>          271          272          273          274          275          276 
#>  0.043172223 -0.888561608  0.260968591 -0.854320209 -0.484954874  0.607213384 
#>          277          278          279          280          281          282 
#> -0.452497008 -0.872993401 -0.854320209 -0.431223395  1.487442469  0.117252564 
#>          283          284          285          286          287          288 
#> -0.881644142 -0.544035555 -0.085332570 -0.634641141 -0.541242016 -0.117325936 
#>          289          290          291          292          293          294 
#>  0.960597168 -0.884297180  1.891668962 -0.989565102 -0.660767691 -1.014606445 
#>          295          296          297          298          299          300 
#>  0.598971439  2.175629439  0.132793842 -0.920113362 -0.888561608 -0.522561197 
#>          301          302          303          304          305          306 
#>  0.294310476 -0.801708644 -0.211750649 -0.420722455 -0.123650270 -0.488484165 
#>          307          308          309          310          311          312 
#>  0.246311849  1.220511909 -0.627409251 -0.853523198  0.129728748 -1.002621969 
#>          313          314 
#> -0.871749693 -0.216032953 
cooks.distance(M.attendance)
#>            1            2            3            4            5            6 
#> 1.400528e-04 3.515846e-04 1.641511e-03 1.349684e-03 2.172141e-03 3.121602e-03 
#>            7            8            9           10           11           12 
#> 1.029428e-03 4.238099e-05 1.015680e-03 1.236957e-02 4.689800e-05 1.846701e-04 
#>           13           14           15           16           17           18 
#> 1.106726e-04 9.687517e-04 1.939269e-03 3.803050e-03 1.615451e-03 2.087364e-03 
#>           19           20           21           22           23           24 
#> 8.230518e-04 1.669257e-02 1.596261e-03 2.137075e-03 1.288162e-03 2.210583e-03 
#>           25           26           27           28           29           30 
#> 2.090763e-05 5.995056e-04 2.714380e-05 2.702461e-03 9.364248e-04 2.684475e-03 
#>           31           32           33           34           35           36 
#> 4.153854e-03 1.760579e-03 4.888538e-04 1.854554e-03 2.426549e-03 1.364295e-02 
#>           37           38           39           40           41           42 
#> 1.601537e-05 1.585674e-05 3.657276e-03 1.216166e-02 3.132178e-04 3.362730e-03 
#>           43           44           45           46           47           48 
#> 4.373679e-03 2.664306e-04 1.997682e-04 6.388289e-04 9.844021e-04 5.062912e-04 
#>           49           50           51           52           53           54 
#> 2.065864e-04 1.947344e-02 7.600565e-05 2.071099e-03 2.387614e-02 1.093723e-03 
#>           55           56           57           58           59           60 
#> 2.747617e-02 1.353029e-03 1.867496e-02 1.121568e-03 1.084525e-03 1.165165e-03 
#>           61           62           63           64           65           66 
#> 1.400633e-03 5.961197e-05 6.151006e-04 2.171848e-04 1.575055e-04 4.236578e-03 
#>           67           68           69           70           71           72 
#> 2.208930e-03 1.003327e-03 6.576176e-05 1.953972e-04 2.631531e-03 2.275931e-03 
#>           73           74           75           76           77           78 
#> 3.588518e-03 6.565437e-05 2.005679e-03 4.113683e-07 2.753928e-04 9.037933e-04 
#>           79           80           81           82           83           84 
#> 4.870184e-02 7.461915e-03 8.312017e-04 1.639196e-04 3.656057e-03 1.845985e-05 
#>           85           86           87           88           89           90 
#> 2.803516e-03 1.105630e-05 5.385818e-04 1.467185e-04 1.874052e-03 2.422800e-05 
#>           91           92           93           94           95           96 
#> 7.379058e-03 1.205375e-04 1.026125e-04 2.891262e-02 2.154561e-04 2.158563e-03 
#>           97           98           99          100          101          102 
#> 1.997682e-04 7.273308e-04 6.959127e-03 1.119005e-03 1.378960e-03 1.880473e-03 
#>          103          104          105          106          107          108 
#> 6.151006e-04 3.079373e-04 1.436125e-02 1.834345e-03 4.612767e-04 6.981077e-04 
#>          109          110          111          112          113          114 
#> 1.829343e-03 3.227782e-04 1.056601e-02 3.381897e-04 7.917946e-04 5.067204e-04 
#>          115          116          117          118          119          120 
#> 1.795692e-03 6.525684e-05 2.448991e-02 1.334333e-04 8.782117e-04 3.079373e-04 
#>          121          122          123          124          125          126 
#> 1.701581e-04 2.722422e-03 1.574726e-03 2.708400e-04 4.659585e-05 3.121602e-03 
#>          127          128          129          130          131          132 
#> 2.178352e-03 2.344090e-04 7.073120e-04 1.620371e-02 2.631366e-03 1.557025e-03 
#>          133          134          135          136          137          138 
#> 2.522854e-04 1.362301e-03 2.506823e-03 2.613864e-07 2.368679e-03 1.062833e-03 
#>          139          140          141          142          143          144 
#> 4.305633e-04 1.422571e-02 2.020421e-03 3.921759e-03 4.416703e-03 1.603583e-05 
#>          145          146          147          148          149          150 
#> 3.495722e-04 3.774265e-04 7.065639e-04 6.911158e-03 4.259818e-04 2.199294e-02 
#>          151          152          153          154          155          156 
#> 1.603416e-06 3.628061e-03 8.482987e-04 7.237768e-05 1.504242e-04 8.230135e-04 
#>          157          158          159          160          161          162 
#> 1.880473e-03 1.304591e-03 4.689800e-05 1.825866e-03 1.825866e-03 4.909942e-05 
#>          163          164          165          166          167          168 
#> 1.521216e-03 1.080900e-03 1.838252e-03 1.968631e-03 1.838252e-03 4.647524e-03 
#>          169          170          171          172          173          174 
#> 7.805522e-06 4.761126e-04 4.651183e-03 8.157091e-04 5.781109e-04 5.460777e-05 
#>          175          176          177          178          179          180 
#> 1.912896e-03 1.803696e-03 4.519868e-04 1.322466e-03 1.984054e-03 5.435209e-04 
#>          181          182          183          184          185          186 
#> 1.083026e-03 5.155460e-03 1.734792e-03 2.046366e-03 8.073447e-04 8.852245e-02 
#>          187          188          189          190          191          192 
#> 1.996622e-04 5.059499e-02 2.291506e-03 9.761909e-06 7.879674e-04 1.175058e-04 
#>          193          194          195          196          197          198 
#> 1.570776e-04 9.321402e-04 1.968836e-03 5.512507e-04 5.729803e-03 2.152147e-03 
#>          199          200          201          202          203          204 
#> 1.817294e-03 8.316477e-04 1.988298e-03 2.082361e-02 1.445863e-03 1.801663e-03 
#>          205          206          207          208          209          210 
#> 1.479955e-03 2.401541e-03 2.510819e-03 2.561790e-03 2.981140e-03 1.966083e-03 
#>          211          212          213          214          215          216 
#> 1.974258e-04 2.003395e-03 2.558079e-03 1.801663e-03 5.435209e-04 4.754444e-05 
#>          217          218          219          220          221          222 
#> 2.055281e-03 2.283525e-03 1.952009e-03 3.134509e-04 1.500214e-03 1.732380e-04 
#>          223          224          225          226          227          228 
#> 3.404885e-03 1.573088e-04 2.870386e-05 7.088608e-03 2.137347e-03 2.117727e-03 
#>          229          230          231          232          233          234 
#> 1.810556e-03 5.105500e-03 1.151268e-03 1.101799e-03 8.094583e-04 1.912896e-03 
#>          235          236          237          238          239          240 
#> 3.466788e-03 2.276858e-03 7.982046e-04 1.559006e-03 1.325448e-04 1.570776e-04 
#>          241          242          243          244          245          246 
#> 1.228540e-03 7.861304e-04 2.254942e-03 8.336288e-03 9.871864e-03 2.098926e-03 
#>          247          248          249          250          251          252 
#> 3.389807e-03 3.132178e-04 1.380583e-03 8.192624e-04 2.343673e-03 5.154656e-04 
#>          253          254          255          256          257          258 
#> 2.282437e-03 3.814085e-02 1.409972e-03 2.506959e-04 3.358265e-04 1.898383e-03 
#>          259          260          261          262          263          264 
#> 5.690455e-04 8.046180e-02 1.131105e-03 2.566921e-02 3.018011e-03 1.953972e-04 
#>          265          266          267          268          269          270 
#> 9.639360e-03 2.646961e-03 4.720897e-03 2.814538e-04 1.083026e-03 4.277072e-05 
#>          271          272          273          274          275          276 
#> 1.268565e-05 1.988298e-03 2.514780e-04 1.858843e-03 5.734037e-04 7.761694e-04 
#>          277          278          279          280          281          282 
#> 5.412720e-04 1.868607e-03 1.858843e-03 5.338413e-04 7.988803e-03 3.681272e-05 
#>          283          284          285          286          287          288 
#> 2.049278e-03 8.104562e-04 4.368900e-05 1.099886e-03 7.565173e-04 2.757966e-05 
#>          289          290          291          292          293          294 
#> 2.254942e-03 2.051661e-03 1.044299e-02 1.837769e-03 1.410230e-03 3.057720e-03 
#>          295          296          297          298          299          300 
#> 8.643981e-04 2.519759e-02 1.139870e-04 2.597703e-03 1.988298e-03 8.316477e-04 
#>          301          302          303          304          305          306 
#> 2.157386e-04 1.353029e-03 1.182126e-04 6.394250e-04 3.693721e-05 8.327133e-04 
#>          307          308          309          310          311          312 
#> 1.465698e-04 3.635586e-03 1.038501e-03 1.592972e-03 4.608371e-05 2.213534e-03 
#>          313          314 
#> 2.421819e-03 1.249664e-04