Takeover Bids data set

This data set gives the number of bids received by 126 US firms that were successful targets of tender offers during the period 1978--1985, along with some explanatory variables on the defensive actions taken by management of target firm, firm-specific characteristics and intervention taken by federal regulators. The takeoverbids data frame has 126 observations on 14 variables. The descriptions below are taken from Sáez-Castillo and Conde-Sánchez (2013).

data(takeoverbids)

Format

A data frame with 126 observations on 14 variables.

bidprem

bid price divided by price 14 working days before bid

docno

doc no

finrest

indicator variable for proposed change in ownership structure

insthold

percentage of stock held by institutions

leglrest

indicator variable for legal defence by lawsuit

numbids

number of bids received after the initial bid

obs

Identifier

rearest

indicator variable for proposed changes in asset structure

regulatn

indicator variable for Department of Justice intervention

size

total book value of assets in billions of dollars

takeover

Indicator. 1 if the company was being taken over

weeks

time in weeks between the initial and final offers

whtknght

indicator variable for management invitation for friendly third-party bid

sizesq

book value squared

Source

Journal of Applied Econometrics data archive: http://qed.econ.queensu.ca/jae/.

References

Cameron, A.C. and Johansson, P. (1997). Count Data Regression Models using Series Expansions: with Applications. Journal of Applied Econometrics 12 203--223.

Cameron, A.C. and Trivedi P.K. (1998). Regression analysis of count data, Cambridge University Press, http://cameron.econ.ucdavis.edu/racd/racddata.html chapter 5.

Croissant Y (2011) Ecdat: Datasets for econometrics, R Package, version 0.1-6.1.

Jaggia, S. and Thosar, S. (1993). Multiple Bids as a Consequence of Target Management Resistance Review of Quantitative Finance and Accounting 3, 447--457.

Examples

## For examples see example(glm.cmp)