Update and Re-fit a COM-Poisson Model

update (i.e., update.cmp) will upate and (by-default) re-fit a model. It is identical to update in the stats package.

# S3 method for cmp
update(object, formula., ..., evaluate = TRUE)

Arguments

object
an object class 'cmp', obtained from a call to glm.cmp.

formula.
changes to the existing formula in object -- see update.formula for details

...
other arguments passed to or from other methods (currently unused).

evaluate
logical; if TRUE evaluate the new call otherwise simply ruturn the call

See also

Examples

data(takeoverbids)

## Fit full model 
M.bids.full <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
M.bids.full

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + whtknght + 
#>     bidprem + insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Linear Model Coefficients:
#> (Intercept)     leglrest      rearest      finrest     whtknght      bidprem  
#>     0.99000      0.26800     -0.17300      0.06770      0.48100     -0.68500  
#>    insthold         size       sizesq     regulatn  
#>    -0.36800      0.17900     -0.00758     -0.03760  
#> 
#> Dispersion (nu): 1.75
#> Degrees of Freedom: 125 Total (i.e. Null);  116 Residual
#> Null Deviance: 182.3899 
#> Residual Deviance: 131.2013 
#> AIC: 382.1753 
#> 

## Dropping whtknght
M.bids.null <- update(M.bids.full, .~.-whtknght)
M.bids.null

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + bidprem + 
#>     insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Linear Model Coefficients:
#> (Intercept)     leglrest      rearest      finrest      bidprem     insthold  
#>     1.33000      0.39100     -0.14600      0.13400     -0.74800     -0.44100  
#>        size       sizesq     regulatn  
#>     0.18800     -0.00794     -0.04460  
#> 
#> Dispersion (nu): 1.54
#> Degrees of Freedom: 125 Total (i.e. Null);  117 Residual
#> Null Deviance: 164.72 
#> Residual Deviance: 131.837 
#> AIC: 393.8799 
#>