Update and Re-fit a COM-Poisson Model

update (i.e., update.cmp) will update and (by-default) re-fit a model. It is identical to update in the stats package.

# S3 method for cmp
update(object, formula., formula_nu., ..., evaluate = TRUE)

Arguments

object
an object class 'cmp', obtained from a call to glm.cmp.

formula.
changes to the existing formula in object -- see update.formula

formula_nu.
changes to the existing formula_nu in object -- see update.formula for details. It also accepts NULL to not regressing on the dispersion.

...
other arguments passed to or from other methods (currently unused).

evaluate
logical; if TRUE evaluate the new call otherwise simply return the call

See also

Examples

# To update the mean regression formula
data(takeoverbids)

## Fit full model 
M.bids.full <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)
M.bids.full

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + whtknght + 
#>     bidprem + insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Linear Model Coefficients:
#> (Intercept)     leglrest      rearest      finrest     whtknght      bidprem  
#>     0.99000      0.26800     -0.17300      0.06770      0.48100     -0.68500  
#>    insthold         size       sizesq     regulatn  
#>    -0.36800      0.17900     -0.00758     -0.03760  
#> 
#> Dispersion (nu): 1.75
#> Degrees of Freedom: 125 Total (i.e. Null);  116 Residual
#> Null Deviance: 182.3899 
#> Residual Deviance: 131.2013 
#> AIC: 382.1753 
#> 

## Dropping whtknght
M.bids.null <- update(M.bids.full, .~.-whtknght)
M.bids.null

#> 
#> Call: glm.cmp(formula = numbids ~ leglrest + rearest + finrest + bidprem + 
#>     insthold + size + sizesq + regulatn, data = takeoverbids)
#> 
#> Linear Model Coefficients:
#> (Intercept)     leglrest      rearest      finrest      bidprem     insthold  
#>     1.33000      0.39100     -0.14600      0.13400     -0.74800     -0.44100  
#>        size       sizesq     regulatn  
#>     0.18800     -0.00794     -0.04460  
#> 
#> Dispersion (nu): 1.54
#> Degrees of Freedom: 125 Total (i.e. Null);  117 Residual
#> Null Deviance: 164.72 
#> Residual Deviance: 131.837 
#> AIC: 393.8799 
#> 

## To update the dispersion regression formula
data(sitophilus)

## Fit full model 
M.sit.full <- glm.cmp(formula = ninsect ~ extract, formula_nu = ~extract, data = sitophilus)
M.sit.full

#> 
#> Call: glm.cmp(formula = ninsect ~ extract, formula_nu = ~extract, data = sitophilus)
#> 
#> Mean Model Coefficients:
#>   (Intercept)    extractLeaf  extractBranch    extractSeed  
#>       3.45000       -0.00637       -0.05210       -3.35000  
#> 
#> Dispersion Model Coefficients:
#>   (Intercept)    extractLeaf  extractBranch    extractSeed  
#>        -0.665         -0.383         -0.372         -0.118  
#> 
#> Degrees of Freedom: 39 Total (i.e. Null);  32 Residual
#> Null Deviance: 257.2108 
#> Residual Deviance: 41.67931 
#> AIC: 260.8279 
#> 

## Dropping extract from the dispersion regression
M.sit.null1 <- update(M.sit.full, formula_nu. =  ~.-extract)
M.sit.null1

#> 
#> Call: glm.cmp(formula = ninsect ~ extract, formula_nu = ~1, data = sitophilus)
#> 
#> Mean Model Coefficients:
#>   (Intercept)    extractLeaf  extractBranch    extractSeed  
#>       3.45000       -0.00637       -0.05210       -3.35000  
#> 
#> Dispersion Model Coefficients:
#> (Intercept)  
#>      -0.927  
#> 
#> Degrees of Freedom: 39 Total (i.e. Null);  35 Residual
#> Null Deviance: 234.4694 
#> Residual Deviance: 41.40781 
#> AIC: 255.2668 
#> 

## To not regress on the dispersion at all
M.sit.null2 <- update(M.sit.full, formula_nu. = NULL)
M.sit.null2

#> 
#> Call: glm.cmp(formula = ninsect ~ extract, data = sitophilus)
#> 
#> Linear Model Coefficients:
#>   (Intercept)    extractLeaf  extractBranch    extractSeed  
#>       3.45000       -0.00637       -0.05210       -3.35000  
#> 
#> Dispersion (nu): 0.396
#> Degrees of Freedom: 39 Total (i.e. Null);  36 Residual
#> Null Deviance: 234.4703 
#> Residual Deviance: 41.40796 
#> AIC: 253.2668 
#>