Extracting the Variance-Covariance Matrix from a COM-Poisson Model Fit

Extracting the Variance-Covariance Matrix from a COM-Poisson Model Fit

# S3 method for cmp
vcov(object, ...)

Arguments

object
an object class 'cmp' object, obtained from a call to glm.cmp

...
other arguments passed to or from other methods (currently unused).

Value

The method will return the estimated covariances between the parameter estimates of the fitted cmp model.

Examples

data(attendance)
M.attendance <- glm.cmp(daysabs~ gender+math+prog, data=attendance)
vcov(M.attendance)

#>                  (Intercept)    gendermale          math  progAcademic
#> (Intercept)     0.0362548681 -6.220521e-03 -2.500916e-04 -2.278626e-02
#> gendermale     -0.0062205215  1.372357e-02  5.319846e-06 -7.355541e-04
#> math           -0.0002500916  5.319846e-06  5.692437e-06  1.039728e-05
#> progAcademic   -0.0227862603 -7.355541e-04  1.039728e-05  2.873855e-02
#> progVocational -0.0189576420 -5.111469e-04 -7.981397e-05  2.253264e-02
#>                progVocational
#> (Intercept)     -1.895764e-02
#> gendermale      -5.111469e-04
#> math            -7.981397e-05
#> progAcademic     2.253264e-02
#> progVocational   3.590206e-02